Predict response from new data

Predict response from new data

Usage

# S3 method for class 'pibblefit'
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1         s2          s3          s4          s5
#> log(c1/c10) -0.9017692 -0.3490987  0.17954940  1.55370109  1.76771468
#> log(c2/c10) -1.7342519 -1.4318339 -1.14256076 -0.39063307 -0.27352609
#> log(c3/c10)  1.1576901  0.6033109  0.07302835 -1.30537185 -1.52004711
#> log(c4/c10) -0.1763679 -0.3545812 -0.52504828 -0.96815523 -1.03716573
#> log(c5/c10) -3.4074491 -3.1973684 -2.99641909 -2.47407762 -2.39272694
#> log(c6/c10) -6.3970030 -5.5771398 -4.79291282 -2.75441711 -2.43693703
#> log(c7/c10)  1.4416937  1.1552574  0.88127124  0.16908009  0.05816177
#> log(c8/c10)  2.6587476  2.4420953  2.23486010  1.69617917  1.61228374
#> log(c9/c10) -1.3488834 -1.0256635 -0.71649264  0.08715666  0.21231888
#>                     s6         s7         s8         s9        s10        s11
#> log(c1/c10)  0.7701907  0.9357578  5.0650861 -0.6225535  3.0380339  6.2920063
#> log(c2/c10) -0.8193653 -0.7287680  1.5307761 -1.5814667  0.4215851  2.2021396
#> log(c3/c10) -0.5194391 -0.6855180 -4.8276131  0.8776111 -2.7942938 -6.0583266
#> log(c4/c10) -0.7155056 -0.7688941 -2.1004313 -0.2664034 -1.4467910 -2.4960622
#> log(c5/c10) -2.7719050 -2.7089697 -1.1393327 -3.3013138 -1.9098542 -0.6729567
#> log(c6/c10) -3.9167214 -3.6711098  2.4545735 -5.9827984 -0.5524724  4.2746576
#> log(c7/c10)  0.5751554  0.4893459 -1.6507898  1.2969827 -0.6002154 -2.2866743
#> log(c8/c10)  2.0033229  1.9384190  0.3196819  2.5492923  1.1143062 -0.1612829
#> log(c9/c10) -0.3710661 -0.2742370  2.1407306 -1.1855889  0.9552435  2.8582741
#>                    s12        s13        s14        s15        s16        s17
#> log(c1/c10)  2.6256118  3.8026271  3.7506556 -0.9179555  5.5594079  2.8681060
#> log(c2/c10)  0.1959101  0.8399660  0.8115275 -1.7431089  1.8012660  0.3286016
#> log(c3/c10) -2.3805966 -3.5612509 -3.5091188  1.1739264 -5.3234632 -2.6238406
#> log(c4/c10) -1.3138020 -1.6933406 -1.6765820 -0.1711485 -2.2598296 -1.3919963
#> log(c5/c10) -2.0666238 -1.6192176 -1.6389730 -3.4136018 -0.9514315 -1.9744470
#> log(c6/c10) -1.1642831  0.5817692  0.5046717 -6.4210147  3.1878789 -0.8045532
#> log(c7/c10) -0.3864665 -0.9964864 -0.9695508  1.4500827 -1.9069854 -0.5121457
#> log(c8/c10)  1.2759797  0.8145781  0.8349515  2.6650928  0.1259029  1.1809196
#> log(c9/c10)  0.7140454  1.4024029  1.3720082 -1.3583497  2.4298263  0.8558641
#>                    s18        s19         s20        s21        s22        s23
#> log(c1/c10)  0.7677743  0.4275617  1.71708459 -0.9241399 -0.6462982  4.2568774
#> log(c2/c10) -0.8206876 -1.0068499 -0.30123058 -1.7464930 -1.5944597  1.0885290
#> log(c3/c10) -0.5170152 -0.1757507 -1.46926049  1.1801300  0.9014293 -4.0169056
#> log(c4/c10) -0.7147264 -0.6050220 -1.02083963 -0.1691542 -0.2587467 -1.8398175
#> log(c5/c10) -2.7728235 -2.9021449 -2.41197241 -3.4159526 -3.3103396 -1.4465484
#> log(c6/c10) -3.9203061 -4.4249971 -2.51204461 -6.4301891 -6.0180227  1.2556302
#> log(c7/c10)  0.5764078  0.7527322  0.08440218  1.4532880  1.3092890 -1.2319139
#> log(c8/c10)  2.0042702  2.1376369  1.63213123  2.6675171  2.5586005  0.6365076
#> log(c9/c10) -0.3724793 -0.5714469  0.18270873 -1.3619666 -1.1994756  1.6680634
#>                    s24        s25         s26         s27         s28
#> log(c1/c10) -2.5901654 -0.1400643  2.24899060  1.94986214  1.73070497
#> log(c2/c10) -2.6581324 -1.3174515 -0.01017475 -0.17385607 -0.29377759
#> log(c3/c10)  2.8513064  0.3936302 -2.00281100 -1.70275773 -1.48292298
#> log(c4/c10)  0.3680699 -0.4219861 -1.19235726 -1.09590074 -1.02523164
#> log(c5/c10) -4.0492409 -3.1179104 -2.20978471 -2.32348919 -2.40679504
#> log(c6/c10) -8.9016672 -5.2670462 -1.72298463 -2.16672900 -2.49183936
#> log(c7/c10)  2.3167506  1.0469197 -0.19127245 -0.03624106  0.07734305
#> log(c8/c10)  3.3206154  2.3601518  1.42361887  1.54088015  1.62679191
#> log(c9/c10) -2.3363134 -0.9034133  0.49378493  0.31884473  0.19067438
#>                     s29         s30
#> log(c1/c10)  1.99119813  1.30963730
#> log(c2/c10) -0.15123727 -0.52418333
#> log(c3/c10) -1.74422151 -1.06055348
#> log(c4/c10) -1.10922988 -0.88945477
#> log(c5/c10) -2.30777659 -2.56685096
#> log(c6/c10) -2.10540881 -3.11647538
#> log(c7/c10) -0.05766455  0.29557272
#> log(c8/c10)  1.52467604  1.79185457
#> log(c9/c10)  0.34301938 -0.05557991
#> 
#> , , 2
#> 
#>                      s1         s2         s3         s4           s5
#> log(c1/c10)  0.06171952  0.3669083  0.6588318  1.4176487  1.535828598
#> log(c2/c10) -1.86343579 -1.6322356 -1.4110848 -0.8362322 -0.746703296
#> log(c3/c10)  0.50864727  0.1696200 -0.1546710 -0.9976234 -1.128906721
#> log(c4/c10) -0.32619309 -0.5514506 -0.7669170 -1.3269940 -1.414221661
#> log(c5/c10) -3.26357969 -3.1047726 -2.9528683 -2.5580128 -2.496517036
#> log(c6/c10) -5.36484746 -4.7719782 -4.2048786 -2.7307772 -2.501197174
#> log(c7/c10)  1.95035487  1.5879500  1.2412975  0.3402195  0.199883486
#> log(c8/c10)  1.70548514  1.5607776  1.4223599  1.0625612  1.006525320
#> log(c9/c10) -1.99898629 -1.5831767 -1.1854406 -0.1515777  0.009438568
#>                     s6         s7         s8         s9        s10       s11
#> log(c1/c10)  0.9849884  1.0764157  3.3566619  0.2159045  2.2373084  4.034176
#> log(c2/c10) -1.1640002 -1.0947381  0.6326951 -1.7466307 -0.2152872  1.145956
#> log(c3/c10) -0.5169909 -0.6185555 -3.1516287  0.3373667 -1.9081646 -3.904264
#> log(c4/c10) -1.0076507 -1.0751326 -2.7581647 -0.4399959 -1.9319785 -3.258233
#> log(c5/c10) -2.7831505 -2.7355756 -1.5490341 -3.1833485 -2.1314973 -1.196485
#> log(c6/c10) -3.5712766 -3.3936669  1.0360104 -5.0653229 -1.1384806  2.352171
#> log(c7/c10)  0.8539939  0.7454259 -1.9623151  1.7672636 -0.6331080 -2.766848
#> log(c8/c10)  1.2677103  1.2243594  0.1431636  1.6323772  0.6739134 -0.178085
#> log(c9/c10) -0.7410630 -0.6164963  2.4902637 -1.7889143  0.9651816  3.413355
#>                    s12        s13         s14         s15        s16        s17
#> log(c1/c10)  2.0095658  2.6595226  2.63082351  0.05278133  3.6296301  2.1434730
#> log(c2/c10) -0.3878169  0.1045672  0.08282582 -1.87020705  0.8394861 -0.2863736
#> log(c3/c10) -1.6551706 -2.3771926 -2.34531148  0.51857651 -3.4548629 -1.8039249
#> log(c4/c10) -1.7638835 -2.2436115 -2.22242898 -0.31959588 -2.9596405 -1.8627193
#> log(c5/c10) -2.2500047 -1.9117953 -1.92672901 -3.26823074 -1.4069932 -2.1803252
#> log(c6/c10) -1.5809005 -0.3182742 -0.37402590 -5.38221107  1.5662870 -1.3207682
#> log(c7/c10) -0.3626688 -1.1344778 -1.10039835  1.96096877 -2.2864588 -0.5216806
#> log(c8/c10)  0.7818993  0.4737174  0.48732531  1.70972325  0.0137337  0.7184062
#> log(c9/c10)  0.6548899  1.5404344  1.50133291 -2.01116428  2.8621739  0.8373339
#>                    s18        s19         s20         s21        s22        s23
#> log(c1/c10)  0.9836540  0.7957860  1.50787028  0.04936623  0.2027925  2.9103630
#> log(c2/c10) -1.1650111 -1.3073332 -0.76788352 -1.87279420 -1.7565639  0.2945950
#> log(c3/c10) -0.5155086 -0.3068104 -1.09784847  0.52237026  0.3519325 -2.6558455
#> log(c4/c10) -1.0066658 -0.8680019 -1.39358584 -0.31707522 -0.4303180 -2.4287549
#> log(c5/c10) -2.7838448 -2.8816032 -2.51106534 -3.27000781 -3.1901714 -1.7812688
#> log(c6/c10) -3.5738688 -3.9388270 -2.55550987 -5.38884535 -5.0907947  0.1690162
#> log(c7/c10)  0.8555784  1.0786674  0.23308336  1.96502413  1.7828339 -1.4323452
#> log(c8/c10)  1.2683430  1.3574220  1.01978196  1.71134255  1.6385943  0.3547796
#> log(c9/c10) -0.7428811 -0.9988450 -0.02865371 -2.01581724 -1.8067790  1.8821962
#>                    s24         s25        s26        s27         s28
#> log(c1/c10) -0.8706256  0.48233866  1.8015928  1.6364118  1.51539156
#> log(c2/c10) -2.5697472 -1.54478972 -0.5453698 -0.6705050 -0.76218567
#> log(c3/c10)  1.5443681  0.04139111 -1.4241381 -1.2406423 -1.10620368
#> log(c4/c10)  0.3619637 -0.63664884 -1.6103802 -1.4884613 -1.39913724
#> log(c5/c10) -3.7487318 -3.04470767 -2.3582248 -2.4441779 -2.50715159
#> log(c6/c10) -7.1760502 -4.54773960 -1.9849154 -2.3058011 -2.54089881
#> log(c7/c10)  3.0574937  1.45087910 -0.1157055  0.0804432  0.22415201
#> log(c8/c10)  2.1475635  1.50604544  0.8805112  0.9588330  1.01621569
#> log(c9/c10) -3.2692757 -1.42590663  0.3715335  0.1464799 -0.01840622
#>                     s29        s30
#> log(c1/c10)  1.65923784  1.2828748
#> log(c2/c10) -0.65321285 -0.9383321
#> log(c3/c10) -1.26599922 -0.8479062
#> log(c4/c10) -1.50530908 -1.2275184
#> log(c5/c10) -2.43230020 -2.6281432
#> log(c6/c10) -2.26145853 -2.9925931
#> log(c7/c10)  0.05333778  0.5002604
#> log(c8/c10)  0.94800991  1.1264652
#> log(c9/c10)  0.17757958 -0.3352026
#>