Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>             [,1]        [,2]        [,3]        [,4]       [,5]       [,6]
#>  [1,] -0.6795870  1.14628429 -0.53450922  0.42843560  1.5734056 -1.8343733
#>  [2,]  0.3371451  0.64463147 -2.04213725  0.75319328  1.5558628 -0.8488946
#>  [3,]  0.4611989 -0.18530184 -0.04386436 -0.05125347 -1.0062559 -0.8927638
#>  [4,] -0.4948672  1.64542724 -0.80602524  1.02499633  1.6177482 -0.7052257
#>  [5,]  0.4309283  1.30608312 -1.74108423 -2.78175207  1.8906229 -0.7559560
#>  [6,]  0.2964854  1.49094371 -1.13434697  0.60420963  0.7717225  0.2207179
#>  [7,]  0.9086010  0.74565527 -1.29740976 -0.36003191  1.1974594 -2.0381804
#>  [8,]  0.6267557  0.86646414 -1.85226653  0.79952373  1.3058054  0.3457720
#>  [9,] -2.0943810 -0.08631764 -0.41613399  0.62228271  0.8204320 -0.4046804
#>              [,7]        [,8]        [,9]     [,10]
#>  [1,]  0.08143240 -1.07125494 -0.48236644 3.2112007
#>  [2,] -1.89340518 -0.32034278 -0.15682744 3.2400790
#>  [3,] -0.29375360 -0.74844922 -2.53484684 3.1418517
#>  [4,] -0.01404607 -1.51632233 -1.32661328 2.4568682
#>  [5,]  0.09967403 -0.08994020 -0.03481101 2.6062105
#>  [6,] -0.65816720 -0.50995550  0.13873394 3.0385747
#>  [7,] -0.04331388 -0.91312594 -0.33918154 3.0554794
#>  [8,] -3.82010298 -0.39019632 -1.95661401 2.4061139
#>  [9,] -0.28543109 -0.03420247 -0.95850310 0.6342452