Summarise orthusfit object and print posterior quantiles

Default calculates median, mean, 50% and 95% credible interval

# S3 method for orthusfit
summary(
  object,
  pars = NULL,
  use_names = TRUE,
  as_factor = FALSE,
  gather_prob = FALSE,
  ...
)

Arguments

object

an object of class orthusfit

pars

character vector (default: c("Eta", "Lambda", "Sigma"))

use_names

should summary replace dimension indices with orthusfit names if names Y and X were named in call to orthus

as_factor

if use_names and as_factor then returns names as factors (useful for maintaining orderings when plotting)

gather_prob

if TRUE then prints quantiles in long format rather than wide (useful for some plotting functions)

...

other expressions to pass to summarise (using name 'val' unquoted is probably what you want)

Examples

if (FALSE) {
fit <- orthus(Y, Z, X)
summary(fit, pars="Eta", median = median(val))

# Some later functions make use of precomputation
fit$summary <- summary(fit)
}