Transform Lambda into IQLR (Inter-Quantile Log-Ratio)

Takes idea from Wu et al. (citation below) and calculates IQLR for Lambda, potentially useful if you believe there is an invariant group of categories (e.g., taxa / genes) that are not changing (in absolute abundance) between samples. IQLR is defined as $$IQLR_x = log(x_i/g(IQVF))$$ for i in 1,...,D. IQVF are the CLR coordinates whose variance is within the inter-quantile range (defined by probs argument to this function). A different IQVF is fit for each posteior sample as the IQVFs are calculted based on posterior estimates for Lambda. The variance of a CLR coordinate is defined as the norm of each row of Lambda[,focus.cov] (i.e., the covariation in Eta, explained by those covariates). This definition of variance allows uses to exclude variation from technical / trivial sources in calculation of IQVF/IQLR.

Usage

lambda_to_iqlr(m, focus.cov = NULL, probs = c(0.25, 0.75))

Arguments

m

object of class pibblefit (e.g., output of pibble)

focus.cov

vector of integers or characters specifying columns (covariates) of Lambda to include in calculating IQLR (if NULL, default, then uses all covariates)

probs

bounds for categories (i.e., features / genes / taxa) to include in calculation of iqlr (smaller bounds means more stringent inclusion criteria)

Value

array of dimension (D, Q, iter) where D is number of taxa, Q is number of covariates, and iter is number of posterior samples.

Details

Primarily intended for doing differential expression analysis under assumption that only small group of categories (e.g., taxa / genes) are changing

References

Jia R. Wu, Jean M. Macklaim, Briana L. Genge, Gregory B. Gloor (2017) Finding the center: corrections for asymmetry in high-throughput sequencing datasets. arxiv:1704.01841v1

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
# Use first two covariates to define iqlr, just show first 5 samples
lambda_to_iqlr(fit, 1:2)[,,1:5] 
#> , , 1
#> 
#>              [,1]       [,2]
#>  [1,]  2.58249657  6.8209952
#>  [2,] -3.11052569 -2.7167130
#>  [3,]  6.49900999 -0.9183486
#>  [4,]  0.96090295 -1.3229594
#>  [5,]  0.14008132  1.1903585
#>  [6,]  0.49972573 -1.3370605
#>  [7,]  0.08957466  0.2489264
#>  [8,]  0.98513706  2.2548793
#>  [9,] -2.44576574  0.4051406
#> [10,]  1.32535003 -0.1768482
#> 
#> , , 2
#> 
#>             [,1]       [,2]
#>  [1,]  2.5587883  6.1883850
#>  [2,] -3.0405870 -3.0912526
#>  [3,]  5.8512540 -0.6003824
#>  [4,]  0.9609548 -1.9479886
#>  [5,]  0.1591712  0.6891760
#>  [6,]  1.1055174 -0.9062200
#>  [7,] -0.1774132 -0.1424990
#>  [8,]  0.2053178  3.0132894
#>  [9,] -2.2717900 -0.1590809
#> [10,]  0.8722740 -0.3306140
#> 
#> , , 3
#> 
#>              [,1]        [,2]
#>  [1,]  2.60530636  5.21789900
#>  [2,] -2.54326545 -2.38212939
#>  [3,]  7.08497756 -0.21955481
#>  [4,]  1.31994025 -0.76897135
#>  [5,]  0.31169726  1.00375772
#>  [6,]  1.08657901 -1.45658403
#>  [7,]  0.04131485 -0.06377461
#>  [8,]  0.32288955  2.57115035
#>  [9,] -2.72940881 -0.34559497
#> [10,]  1.47805346 -0.38137550
#> 
#> , , 4
#> 
#>             [,1]        [,2]
#>  [1,]  3.3439506  4.57514801
#>  [2,] -3.7370334 -3.32615210
#>  [3,]  5.8688759 -0.94841909
#>  [4,]  1.0464624 -1.07088958
#>  [5,]  0.5365942 -0.07507684
#>  [6,]  0.8422877 -1.53300228
#>  [7,] -0.1707247 -0.73045076
#>  [8,]  0.4358403  2.39571374
#>  [9,] -1.1074033 -0.13226070
#> [10,]  1.2253342 -0.65615197
#> 
#> , , 5
#> 
#>             [,1]        [,2]
#>  [1,]  2.7120487  6.38024654
#>  [2,] -2.4680663 -2.88067266
#>  [3,]  5.2612425 -0.03676643
#>  [4,]  1.1926378 -1.48758213
#>  [5,] -0.1243978  0.88179646
#>  [6,]  0.8403210 -1.32344167
#>  [7,] -0.4971559 -0.48544153
#>  [8,]  0.9497406  2.72186980
#>  [9,] -2.9826994  0.08915400
#> [10,]  1.1840482  0.09361379
#>