Calculations for the Collapsed Pibble Model

Functions providing access to the Log Likelihood, Gradient, and Hessian of the collapsed pibble model. Note: These are convenience functions but are not as optimized as direct coding of the PibbleCollapsed C++ class due to a lack of Memoization. By contrast function optimPibbleCollapsed is much more optimized and massively cuts down on repeated calculations. A more efficient Rcpp module based implementation of these functions may following if the future. For model details see optimPibbleCollapsed documentation

loglikPibbleCollapsed(Y, upsilon, ThetaX, KInv, AInv, eta, sylv = FALSE)

gradPibbleCollapsed(Y, upsilon, ThetaX, KInv, AInv, eta, sylv = FALSE)

hessPibbleCollapsed(Y, upsilon, ThetaX, KInv, AInv, eta, sylv = FALSE)

Arguments

Y

D x N matrix of counts

upsilon

(must be > D)

ThetaX

D-1 x N matrix formed by Theta*X (Theta is Prior mean for regression coefficients)

KInv

Inverse of K for LTP (for Pibble defined as KInv = solve(Xi))

AInv

Inverse of A for LTP (for Pibble defined as AInv = solve(diag(N)+ X'GammaX) )

eta

matrix (D-1)xN of parameter values at which to calculate quantities

sylv

(default:false) if true and if N < D-1 will use sylvester determinant identity to speed computation

Value

see below

loglikPibbleCollapsed - double

gradPibbleCollapsed - vector

hessPibbleCollapsed- matrix

Examples

D <- 10
Q <- 2
N <- 30

# Simulate Data
Sigma <- diag(sample(1:8, D-1, replace=TRUE))
Sigma[2, 3] <- Sigma[3,2] <- -1
Gamma <- diag(sqrt(rnorm(Q)^2))
Theta <- matrix(0, D-1, Q)
Phi <-  Theta + t(chol(Sigma))%*%matrix(rnorm(Q*(D-1)), nrow=D-1)%*%chol(Gamma)
X <- matrix(rnorm(N*(Q-1)), Q-1, N)
X <- rbind(1, X)
Eta <- Phi%*%X + t(chol(Sigma))%*%matrix(rnorm(N*(D-1)), nrow=D-1)
Pi <- t(alrInv(t(Eta)))
Y <- matrix(0, D, N)
for (i in 1:N) Y[,i] <- rmultinom(1, sample(5000:10000), prob = Pi[,i])

# Priors
upsilon <- D+10
Xi <- Sigma*(upsilon-D)

# Precompute
KInv <- solve(Xi)
AInv <- solve(diag(N)+ t(X)%*%Gamma%*%X)
ThetaX <- Theta%*%X

loglikPibbleCollapsed(Y, upsilon, ThetaX, KInv, AInv, Eta)
#> [1] -112182.2
gradPibbleCollapsed(Y, upsilon, ThetaX, KInv, AInv, Eta)[1:5]
#> [1]  -0.7534645  14.4448444   1.2878366   0.6680285 -14.7969431
hessPibbleCollapsed(Y, upsilon, ThetaX, KInv, AInv, Eta)[1:5,1:5]
#>             [,1]          [,2]         [,3]         [,4]          [,5]
#> [1,] -0.62975819   0.401997663 -0.018924609  0.015266380  -0.014236647
#> [2,]  0.40199766 -70.378216040  0.369539014  0.007643699  55.986319595
#> [3,] -0.01892461   0.369539014 -0.611643520 -0.086384937  -0.008490053
#> [4,]  0.01526638   0.007643699 -0.086384937 -0.297012890  -0.002763937
#> [5,] -0.01423665  55.986319595 -0.008490053 -0.002763937 -56.291092139