Predict response from new data

Predict response from new data

Usage

# S3 method for class 'pibblefit'
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1         s2         s3          s4         s5         s6
#> log(c1/c10)  1.0442223  0.7064624  0.6294163 -0.60022384 -0.3204057  0.1363944
#> log(c2/c10) -0.6793906  0.2370479  0.4460959  3.78245873  3.0232327  1.7838043
#> log(c3/c10) -0.7904461 -1.5895193 -1.7717953 -4.68088111 -4.0188867 -2.9381878
#> log(c4/c10) -0.5596453  0.3500375  0.5575445  3.86931283  3.1156836  1.8853918
#> log(c5/c10) -0.7475619 -0.6292424 -0.6022527 -0.17150166 -0.2695238 -0.4295439
#> log(c6/c10)  1.9593575  1.5091154  1.4064110 -0.23272927  0.1402751  0.7492008
#> log(c7/c10)  1.7906498  1.6139957  1.5736993  0.93057706  1.0769266  1.3158407
#> log(c8/c10) -0.3612792 -0.3068995 -0.2944950 -0.09652191 -0.1415729 -0.2151181
#> log(c9/c10)  3.1589885  1.1798240  0.7283584 -6.47693780 -4.8372935 -2.1605915
#>                     s7         s8          s9        s10        s11        s12
#> log(c1/c10)  0.2888802  0.3297035  0.04580418  0.5932776  0.7218891  0.2941371
#> log(c2/c10)  1.3700671  1.2593019  2.02960127  0.5441504  0.1951908  1.3558036
#> log(c3/c10) -2.5774365 -2.4808567 -3.15250633 -1.8572923 -1.5530227 -2.5649997
#> log(c4/c10)  1.4747045  1.3647558  2.12937682  0.6548762  0.3084890  1.4605461
#> log(c5/c10) -0.4829606 -0.4972613 -0.39780952 -0.5895931 -0.6346465 -0.4848022
#> log(c6/c10)  0.9524681  1.0068865  0.62844184  1.3582373  1.5296796  0.9594757
#> log(c7/c10)  1.3955933  1.4169445  1.26846048  1.5547982  1.6220642  1.3983427
#> log(c8/c10) -0.2396684 -0.2462410 -0.20053299 -0.2886767 -0.3093833 -0.2405148
#> log(c9/c10) -1.2670738 -1.0278625 -2.69142107  0.5165974  1.2702197 -1.2362700
#>                    s13        s14        s15        s16        s17         s18
#> log(c1/c10)  0.4393032  1.1839619  0.1704123  0.7108586  0.4475010  0.09250168
#> log(c2/c10)  0.9619268 -1.0585439  1.6915042  0.2251199  0.9396836  1.90289769
#> log(c3/c10) -2.2215654 -0.4598497 -2.8577082 -1.5791189 -2.2021709 -3.04202926
#> log(c4/c10)  1.0695729 -0.9360037  1.7937721  0.3381974  1.0474937  2.00360725
#> log(c5/c10) -0.5356548 -0.7965137 -0.4414605 -0.6307824 -0.5385266 -0.41416796
#> log(c6/c10)  1.1529856  2.1456339  0.7945474  1.5149756  1.1639135  0.69069075
#> log(c7/c10)  1.4742670  1.8637359  1.3336326  1.6162950  1.4785546  1.29288405
#> log(c8/c10) -0.2638867 -0.3837774 -0.2205950 -0.3076073 -0.2652065 -0.20805132
#> log(c9/c10) -0.3856435  3.9778179 -1.9612578  1.2055841 -0.3376065 -2.41778870
#>                     s19        s20         s21         s22        s23
#> log(c1/c10) -0.59617026 -0.1132431  0.01669166 -0.45266696  0.8995317
#> log(c2/c10)  3.77146022  2.4611417  2.10859181  3.38209502 -0.2868040
#> log(c3/c10) -4.67129115 -3.5287808 -3.22138081 -4.33179066 -1.1327555
#> log(c4/c10)  3.85839540  2.5577361  2.20778507  3.47190046 -0.1699527
#> log(c5/c10) -0.17292165 -0.3420942 -0.38761121 -0.22319180 -0.6968759
#> log(c6/c10) -0.22732575  0.4164279  0.58963414 -0.03603235  1.7664815
#> log(c7/c10)  0.93269714  1.1852761  1.25323414  1.00775176  1.7149742
#> log(c8/c10) -0.09717454 -0.1749262 -0.19584584 -0.12027869 -0.3379839
#> log(c9/c10) -6.45318514 -3.6233870 -2.86201110 -5.61230188  2.3111481
#>                    s24        s25        s26        s27         s28        s29
#> log(c1/c10)  0.2787048  0.5645654  0.3968768 -0.1392652 -0.49044499  0.9264255
#> log(c2/c10)  1.3976758  0.6220547  1.0770417  2.5317471  3.48459754 -0.3597744
#> log(c3/c10) -2.6015095 -1.9252196 -2.3219379 -3.5903440 -4.42116603 -1.0691301
#> log(c4/c10)  1.5021097  0.7322062  1.1838391  2.6278210  3.57364737 -0.2423852
#> log(c5/c10) -0.4793961 -0.5795350 -0.5207926 -0.3329785 -0.20995791 -0.7062970
#> log(c6/c10)  0.9389040  1.3199632  1.0964302  0.3817398 -0.08639138  1.8023315
#> log(c7/c10)  1.3902714  1.5397812  1.4520773  1.1716661  0.98799322  1.7290401
#> log(c8/c10) -0.2380302 -0.2840540 -0.2570560 -0.1707366 -0.11419640 -0.3423138
#> log(c9/c10) -1.3266983  0.3483532 -0.6342485 -3.7758684 -5.83366900  2.4687371
#>                    s30
#> log(c1/c10) -0.1792545
#> log(c2/c10)  2.6402493
#> log(c3/c10) -3.6849507
#> log(c4/c10)  2.7355234
#> log(c5/c10) -0.3189700
#> log(c6/c10)  0.3284331
#> log(c7/c10)  1.1507510
#> log(c8/c10) -0.1642983
#> log(c9/c10) -4.0101926
#> 
#> , , 2
#> 
#>                       s1           s2            s3          s4          s5
#> log(c1/c10)  1.206251504  0.790660342  0.6958601726 -0.81713055 -0.47283290
#> log(c2/c10) -0.338484770  0.389713432  0.5558221437  3.20688218  2.60360436
#> log(c3/c10) -1.001586971 -1.775384319 -1.9518946151 -4.76896175 -4.12790720
#> log(c4/c10) -1.561101768 -0.217141890  0.0894277271  4.98221421  3.86880693
#> log(c5/c10) -0.911958888 -0.767436287 -0.7344693513 -0.20832401 -0.32805416
#> log(c6/c10)  1.179048668  0.812163370  0.7284734568 -0.60719992 -0.30325278
#> log(c7/c10)  0.742508054  0.961983811  1.0120482514  1.81106617  1.62924087
#> log(c8/c10) -0.009502774 -0.002011935 -0.0003032054  0.02696775  0.02076195
#> log(c9/c10)  3.047699972  1.218786835  0.8015949166 -5.85670019 -4.34153204
#>                      s6           s7           s8          s9          s10
#> log(c1/c10)  0.08922940  0.276853062  0.327083402 -0.02223586  0.651393938
#> log(c2/c10)  1.61875966  1.290005763  1.201992239  1.81406891  0.633735813
#> log(c3/c10) -3.08139236 -2.732052182 -2.638527327 -3.28893172 -2.034687171
#> log(c4/c10)  2.05118123  1.444434270  1.281996839  2.41164326  0.233224903
#> log(c5/c10) -0.52351237 -0.588758848 -0.606226542 -0.48475012 -0.719006135
#> log(c6/c10)  0.19293774  0.358572549  0.402916064  0.09453583  0.689218513
#> log(c7/c10)  1.33241298  1.233327988  1.206801096  1.39127833  1.035531091
#> log(c8/c10)  0.01063103  0.007249204  0.006343826  0.01264014  0.000498278
#> log(c9/c10) -1.86803593 -1.042350954 -0.821299748 -2.35856676  0.605910099
#>                      s11          s12          s13         s14          s15
#> log(c1/c10)  0.809641911  0.283321327  0.461938562  1.37819188  0.131086119
#> log(c2/c10)  0.356453956  1.278672081  0.965699234 -0.63975843  1.545418381
#> log(c3/c10) -1.740042162 -2.720008794 -2.387437856 -0.68144781 -3.003458518
#> log(c4/c10) -0.278525455  1.423516867  0.845895351 -2.11713132  1.915822841
#> log(c5/c10) -0.774037163 -0.591008199 -0.653122675 -0.97175147 -0.538068122
#> log(c6/c10)  0.828920364  0.364282754  0.521966660  1.33083822  0.229888993
#> log(c7/c10)  0.951959550  1.229912065  1.135583416  0.65170549  1.310308235
#> log(c8/c10) -0.002354069  0.007132617  0.003913124 -0.01260192  0.009876586
#> log(c9/c10)  1.302319990 -1.013885735 -0.227835797  3.80436672 -1.683834947
#>                      s16          s17         s18         s19         s20
#> log(c1/c10)  0.796069524  0.472025516  0.03522232 -0.81214289 -0.21793300
#> log(c2/c10)  0.380235472  0.948024888  1.71339077  3.19814281  2.15696914
#> log(c3/c10) -1.765312857 -2.368656758 -3.18194920 -4.75967513 -3.65330406
#> log(c4/c10) -0.234634378  0.813275645  2.22583206  4.96608487  3.04449864
#> log(c5/c10) -0.769317340 -0.656630433 -0.50473131 -0.21005848 -0.41669607
#> log(c6/c10)  0.816938614  0.530871457  0.14526011 -0.60279680 -0.07822627
#> log(c7/c10)  0.959127195  1.130256445  1.36093438  1.80843216  1.49462696
#> log(c8/c10) -0.002109433  0.003731311  0.01160449  0.02687785  0.01616749
#> log(c9/c10)  1.242591297 -0.183445625 -2.10570761 -5.83475075 -3.21978109
#>                     s21         s22          s23          s24          s25
#> log(c1/c10) -0.05805689 -0.63557158  1.028219331  0.264332919  0.616065489
#> log(c2/c10)  1.87683447  2.88875484 -0.026537067  1.311943540  0.695638268
#> log(c3/c10) -3.35562760 -4.43091355 -1.333068571 -2.755363684 -2.100465904
#> log(c4/c10)  2.52748314  4.39507958 -0.985372254  1.484922549  0.347471842
#> log(c5/c10) -0.47229329 -0.27146148 -0.850047868 -0.584404945 -0.706720601
#> log(c6/c10)  0.06291290 -0.44691904  1.021881257  0.347519723  0.658030437
#> log(c7/c10)  1.41019560  1.71518398  0.836527729  1.239939938  1.054188220
#> log(c8/c10)  0.01328580  0.02369524 -0.006293826  0.007474874  0.001135057
#> log(c9/c10) -2.51620619 -5.05770443  2.264224745 -1.097448986  0.450438398
#>                      s26         s27         s28          s29        s30
#> log(c1/c10)  0.409735700 -0.24995151 -0.68205495  1.061310347 -0.2991556
#> log(c2/c10)  1.057169009  2.21307194  2.97020292 -0.084519095  2.2992873
#> log(c3/c10) -2.484635390 -3.71291997 -4.51746184 -1.271455758 -3.8045341
#> log(c4/c10)  1.014711626  3.14804175  4.54539985 -1.092383669  3.3071606
#> log(c5/c10) -0.634969033 -0.40556158 -0.25529680 -0.861555330 -0.3884507
#> log(c6/c10)  0.475881795 -0.10649233 -0.48795472  1.051094119 -0.1499299
#> log(c7/c10)  1.163152007  1.51153610  1.73973207  0.819052199  1.5375210
#> log(c8/c10)  0.004854056  0.01674461  0.02453308 -0.006890276  0.0176315
#> log(c9/c10) -0.457567582 -3.36068661 -5.26226613  2.409850061 -3.5772217
#>