Predict response from new data

Predict response from new data

Usage

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                      s1          s2          s3          s4          s5
#> log(c1/c10) -0.17796083  0.62834337  0.33007341  0.29972910  0.61584499
#> log(c2/c10) -0.66714047 -0.67442760 -0.67173193 -0.67145768 -0.67431464
#> log(c3/c10)  0.01194827 -0.16521315 -0.09967718 -0.09300992 -0.16246701
#> log(c4/c10) -1.92446595 -2.35179063 -2.19371368 -2.17763181 -2.34516675
#> log(c5/c10)  1.22454477  0.80361548  0.95932664  0.97516782  0.81014023
#> log(c6/c10) -0.44837044 -1.05918303 -0.83322979 -0.81024257 -1.04971493
#> log(c7/c10)  0.62806714 -0.33397594  0.02190531  0.05811067 -0.31906349
#> log(c8/c10) -2.59949584 -2.57616941 -2.58479838 -2.58567624 -2.57653099
#> log(c9/c10) -0.73677652  0.07230161 -0.22699448 -0.25744319  0.05976024
#>                      s6         s7           s8          s9        s10
#> log(c1/c10)  0.26946078  1.7804766  0.549209754 -0.53527234  1.5498704
#> log(c2/c10) -0.67118413 -0.6848402 -0.673712413 -0.66391120 -0.6827561
#> log(c3/c10) -0.08635935 -0.4183602 -0.147825889  0.09045687 -0.3676914
#> log(c4/c10) -2.16159022 -2.9623976 -2.309851437 -1.73509819 -2.8401810
#> log(c5/c10)  0.99096933  0.2021469  0.844927004  1.41107844  0.3225344
#> log(c6/c10) -0.78731292 -1.9319770 -0.999235666 -0.17769050 -1.7572822
#> log(c7/c10)  0.09422537 -1.7086455 -0.239557551  1.05439391 -1.4334973
#> log(c8/c10) -2.58655191 -2.5428381 -2.578458751 -2.60983288 -2.5495096
#> log(c9/c10) -0.28781564  1.2283985 -0.007104249 -1.09531728  0.9969989
#>                    s11        s12        s13        s14         s15
#> log(c1/c10)  0.7833506  1.4191762 -1.5632197  3.3429403  2.05780267
#> log(c2/c10) -0.6758285 -0.6815749 -0.6546209 -0.6989613 -0.68734661
#> log(c3/c10) -0.1992714 -0.3389752  0.3163178 -0.7616653 -0.47929441
#> log(c4/c10) -2.4339413 -2.7709158 -1.1903072 -3.7904713 -3.10937472
#> log(c5/c10)  0.7226943  0.3907630  1.9477161 -0.6135338  0.05736947
#> log(c6/c10) -1.1766082 -1.6582753  0.6010270 -3.1156153 -2.14206432
#> log(c7/c10) -0.5189231 -1.2775593  2.2808909 -3.5729015 -2.03953749
#> log(c8/c10) -2.5716850 -2.5532906 -2.6395715 -2.4976360 -2.53481507
#> log(c9/c10)  0.2278422  0.8658551 -2.1268011  2.7962376  1.50667868
#>                      s16        s17        s18         s19         s20
#> log(c1/c10) -0.091900659 -0.9305473 -1.1588406  0.18150558  0.60250972
#> log(c2/c10) -0.667918257 -0.6603388 -0.6582756 -0.67038922 -0.67419412
#> log(c3/c10) -0.006960901  0.1773068  0.2274675 -0.06703381 -0.15953697
#> log(c4/c10) -1.970076079 -1.5256106 -1.4046198 -2.11497577 -2.33809932
#> log(c5/c10)  1.179617252  1.6174308  1.7366108  1.03688614  0.81710188
#> log(c6/c10) -0.513564985  0.1217485  0.2946912 -0.72068281 -1.03961285
#> log(c7/c10)  0.525384314  1.5260168  1.7984053  0.19916925 -0.30315249
#> log(c8/c10) -2.597006114 -2.6212682 -2.6278728 -2.58909646 -2.57691678
#> log(c9/c10) -0.650420271 -1.4919521 -1.7210308 -0.37607344  0.04637908
#>                    s21        s22        s23        s24          s25        s26
#> log(c1/c10)  2.9894787  2.8030632  1.2790793  0.7309004 -0.158360099  1.4552383
#> log(c2/c10) -0.6957668 -0.6940820 -0.6803087 -0.6753545 -0.667317617 -0.6819008
#> log(c3/c10) -0.6840026 -0.6430433 -0.3081931 -0.1877470  0.007641588 -0.3468988
#> log(c4/c10) -3.6031439 -3.5043475 -2.6966673 -2.4061437 -1.934853941 -2.7900280
#> log(c5/c10) -0.4290099 -0.3316921  0.4639002  0.7500758  1.214312252  0.3719368
#> log(c6/c10) -2.8478518 -2.7066335 -1.5521455 -1.1368747 -0.463218897 -1.6855941
#> log(c7/c10) -3.1511682 -2.9287462 -1.1104026 -0.4563420  0.604680496 -1.3205869
#> log(c8/c10) -2.5078616 -2.5132546 -2.5573436 -2.5732024 -2.598928790 -2.5522473
#> log(c9/c10)  2.4415599  2.2545031  0.7252763  0.1752114 -0.717108351  0.9020413
#>                    s27        s28         s29         s30
#> log(c1/c10)  0.6060679  3.1697522  0.22771251  0.21891577
#> log(c2/c10) -0.6742263 -0.6973961 -0.67080682 -0.67072732
#> log(c3/c10) -0.1603188 -0.7236123 -0.07718641 -0.07525359
#> log(c4/c10) -2.3399851 -3.6986852 -2.13946450 -2.13480241
#> log(c5/c10)  0.8152443 -0.5231213  1.01276392  1.01735623
#> log(c6/c10) -1.0423083 -2.9844174 -0.75568669 -0.74902276
#> log(c7/c10) -0.3073979 -3.3662618  0.14403737  0.15453321
#> log(c8/c10) -2.5768138 -2.5026463 -2.58775969 -2.58801418
#> log(c9/c10)  0.0499495  2.6224537 -0.32970754 -0.33853454
#> 
#> , , 2
#> 
#>                      s1          s2          s3         s4         s5
#> log(c1/c10) -0.70403507  0.31245879 -0.06356502 -0.1018196  0.2967023
#> log(c2/c10) -1.67884994 -1.46652633 -1.54506958 -1.5530601 -1.4698175
#> log(c3/c10)  0.65089646  0.17187600  0.34907638  0.3671038  0.1793012
#> log(c4/c10) -1.22650042 -1.11029913 -1.15328459 -1.1576577 -1.1121003
#> log(c5/c10)  1.72355960  1.21870285  1.40546065  1.4244603  1.2265285
#> log(c6/c10) -0.15669507 -0.72579934 -0.51527494 -0.4938574 -0.7169778
#> log(c7/c10)  0.25533429  0.05046872  0.12625308  0.1339630  0.0536443
#> log(c8/c10) -2.01623723 -1.52654592 -1.70769368 -1.7261226 -1.5341365
#> log(c9/c10) -0.03051471  0.32673385  0.19457962  0.1811350  0.3211962
#>                     s6         s7          s8          s9        s10
#> log(c1/c10) -0.1399783  1.7649334  0.21269640 -1.15449155  1.4742121
#> log(c2/c10) -1.5610307 -1.1631358 -1.48736453 -1.77294056 -1.2238612
#> log(c3/c10)  0.3850860 -0.5125994  0.21888881  0.86317307 -0.3755976
#> log(c4/c10) -1.1620198 -0.9442584 -1.12170355 -1.27799470 -0.9774924
#> log(c5/c10)  1.4434125  0.4973098  1.26825132  1.94728550  0.6417008
#> log(c6/c10) -0.4724934 -1.5389961 -0.66994539  0.09550194 -1.3762300
#> log(c7/c10)  0.1416535 -0.2422650  0.07057497  0.34611991 -0.1836726
#> log(c8/c10) -1.7445054 -0.8268229 -1.57460600 -2.23324259 -0.9668765
#> log(c9/c10)  0.1677240  0.8372086  0.29167218 -0.18882844  0.7350341
#>                     s11        s12        s13        s14        s15        s16
#> log(c1/c10)  0.50787381  1.3094482 -2.4504072  3.7347046  2.1145536 -0.5955405
#> log(c2/c10) -1.42570835 -1.2582768 -2.0436293 -0.7516931 -1.0901077 -1.6561878
#> log(c3/c10)  0.07978711 -0.2979530  1.4738704 -1.4408497 -0.6773571  0.5997686
#> log(c4/c10) -1.08796011 -0.9963275 -1.4261383 -0.7190825 -0.9042913 -1.2140978
#> log(c5/c10)  1.12164708  0.7235333  2.5909212 -0.4810063  0.3236657  1.6696742
#> log(c6/c10) -0.83520632 -1.2839836  0.8210460 -2.6418116 -1.7347379 -0.2174379
#> log(c7/c10)  0.01108451 -0.1504659  0.6073005 -0.6392554 -0.3127280  0.2334681
#> log(c8/c10) -1.43240562 -1.0462508 -2.8575441  0.1221055 -0.6583949 -1.9639705
#> log(c9/c10)  0.39541280  0.6771275 -0.6442803  1.5294882  0.9600833  0.0076159
#>                    s17        s18        s19         s20         s21        s22
#> log(c1/c10) -1.6528079 -1.9406133 -0.2508619  0.27989076  3.28910156  3.0540907
#> log(c2/c10) -1.8770281 -1.9371444 -1.5841919 -1.47332909 -0.84476995 -0.8938586
#> log(c3/c10)  1.0980035  1.2336312  0.4373396  0.18722362 -1.23086025 -1.1201119
#> log(c4/c10) -1.3349601 -1.3678608 -1.1746956 -1.11402217 -0.77002191 -0.7968874
#> log(c5/c10)  2.1947817  2.3377245  1.4984844  1.23487824 -0.25969094 -0.1429693
#> log(c6/c10)  0.3744942  0.5356278 -0.4104130 -0.70756548 -2.39233190 -2.2607564
#> log(c7/c10)  0.4465513  0.5045560  0.1640012  0.05703253 -0.54944797 -0.5020836
#> log(c8/c10) -2.4733042 -2.6119532 -1.7979231 -1.54223542 -0.09256177 -0.2057772
#> log(c9/c10) -0.3639626 -0.4651123  0.1287538  0.31528776  1.37288021  1.2902852
#>                    s23         s24         s25        s26         s27
#> log(c1/c10)  1.1328305  0.44175064 -0.67932476  1.3549111  0.28437648
#> log(c2/c10) -1.2951684 -1.43952005 -1.67368849 -1.2487806 -1.47239211
#> log(c3/c10) -0.2147223  0.11094751  0.63925178 -0.3193773  0.18510973
#> log(c4/c10) -1.0165177 -1.09551904 -1.22367564 -0.9911304 -1.11350938
#> log(c5/c10)  0.8112531  1.15448813  1.71128686  0.7009535  1.23265034
#> log(c6/c10) -1.1851007 -0.79818595 -0.17052963 -1.3094369 -0.71007690
#> log(c7/c10) -0.1148701  0.02441106  0.25035414 -0.1596285  0.05612847
#> log(c8/c10) -1.1313356 -1.46426016 -2.00433315 -1.0243493 -1.54007444
#> log(c9/c10)  0.6150549  0.37217370 -0.02183023  0.6931055  0.31686427
#>                     s28        s29        s30
#> log(c1/c10)  3.51636931 -0.1926096 -0.2036995
#> log(c2/c10) -0.79729863 -1.5720242 -1.5743407
#> log(c3/c10) -1.33795966  0.4098884  0.4151145
#> log(c4/c10) -0.74404162 -1.1680364 -1.1693042
#> log(c5/c10) -0.37256684  1.4695526  1.4750605
#> log(c6/c10) -2.51957226 -0.4430267 -0.4368178
#> log(c7/c10) -0.59525182  0.1522609  0.1544960
#> log(c8/c10)  0.01692344 -1.7698603 -1.7752029
#> log(c9/c10)  1.45275386  0.1492267  0.1453291
#>