Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1          s2          s3          s4           s5
#> log(c1/c10) -0.1050909 -3.21410875  0.26670481  -4.3985857 -0.940607044
#> log(c2/c10)  0.7489510 -0.43114794  0.89007460  -0.8807433  0.431811718
#> log(c3/c10) -0.8019256  0.03607504 -0.90213890   0.3553374 -0.576721636
#> log(c4/c10) -0.1203352 -8.18771343  0.84441214 -11.2612319 -2.288359006
#> log(c5/c10) -1.1818159 -3.00932035 -0.96327152  -3.7055650 -1.672938659
#> log(c6/c10)  2.8334149  1.42784457  3.00150176   0.8923489  2.455682533
#> log(c7/c10)  0.1406436 -0.38876113  0.20395309  -0.5904543 -0.001628421
#> log(c8/c10) -0.9910635 -9.45245137  0.02080204 -12.6760801 -3.264973239
#> log(c9/c10)  0.6481212 -2.27234957  0.99736933  -3.3849937 -0.136724849
#>                     s6         s7          s8          s9         s10
#> log(c1/c10)  3.1895286  2.3585525 -0.37987364 -0.53080771 -0.54146415
#> log(c2/c10)  1.9994993  1.6840833  0.64465093  0.58736044  0.58331555
#> log(c3/c10) -1.6899530 -1.4659728 -0.72786099 -0.68717841 -0.68430609
#> log(c4/c10)  8.4286482  6.2724049 -0.83335016 -1.22499869 -1.25265035
#> log(c5/c10)  0.7547867  0.2663325 -1.34333528 -1.43205548 -1.43831941
#> log(c6/c10)  4.3228948  3.9472149  2.70918711  2.64095062  2.63613290
#> log(c7/c10)  0.7016526  0.5601537  0.09385347  0.06815236  0.06633778
#> log(c8/c10)  7.9754512  5.7138973 -1.73890194 -2.14967854 -2.17868070
#> log(c9/c10)  3.7429379  2.9623565  0.39000276  0.24822212  0.23821194
#>                    s11         s12         s13         s14        s15
#> log(c1/c10)  1.0854360  1.94759330 -1.34941469 -1.85234532  1.6162255
#> log(c2/c10)  1.2008428  1.52809433  0.27663941  0.08574056  1.4023161
#> log(c3/c10) -1.1228186 -1.35520338 -0.46653216 -0.33097289 -1.2658869
#> log(c4/c10)  2.9688818  5.20603496 -3.34914606 -4.65416645  4.3461913
#> log(c5/c10) -0.4820151  0.02476759 -1.91323892 -2.20886539 -0.1700129
#> log(c6/c10)  3.3716458  3.76142246  2.27086278  2.04349053  3.6116129
#> log(c7/c10)  0.3433669  0.49017540 -0.07124034 -0.15687955  0.4337500
#> log(c8/c10)  2.2490305  4.59544596 -4.37756899 -5.74632641  3.6936076
#> log(c9/c10)  1.7664484  2.57631999 -0.52074027 -0.99317057  2.2650481
#>                    s16         s17        s18         s19         s20
#> log(c1/c10) -0.1908560 -0.99697212 -2.8589752 -0.62604058 -0.86086143
#> log(c2/c10)  0.7163969  0.41041706 -0.2963489  0.55121263  0.46208099
#> log(c3/c10) -0.7788086 -0.56152907 -0.0596472 -0.66150947 -0.59821617
#> log(c4/c10) -0.3428811 -2.43461690 -7.2662015 -1.47211197 -2.08143254
#> log(c5/c10) -1.2322292 -1.70607048 -2.8005701 -1.48803409 -1.62606357
#> log(c6/c10)  2.7946410  2.43020018  1.5883986  2.59789635  2.49173510
#> log(c7/c10)  0.1260395 -0.01122629 -0.3282889  0.05193607  0.01195069
#> log(c8/c10) -1.2244785 -3.41837435 -8.4859329 -2.40886080 -3.04794051
#> log(c9/c10)  0.5675574 -0.18967166 -1.9387531  0.15876466 -0.06181542
#>                     s21         s22        s23        s24        s25        s26
#> log(c1/c10) -2.00004498 -2.13735680  0.3588467  0.9694064  2.5374325  0.3748555
#> log(c2/c10)  0.02967778 -0.02244207  0.9250492  1.1568011  1.7519813  0.9311257
#> log(c3/c10) -0.29116212 -0.25415127 -0.9269747 -1.0915442 -1.5141879 -0.9312897
#> log(c4/c10) -5.03742221 -5.39372327  1.0835049  2.6678045  6.7365684  1.1250450
#> log(c5/c10) -2.29568438 -2.37639732 -0.9091098 -0.5502182  0.3714796 -0.8996997
#> log(c6/c10)  1.97671631  1.91463837  3.0431586  3.3191894  4.0280856  3.0503961
#> log(c7/c10) -0.18202990 -0.20541141  0.2196430  0.3236094  0.5906134  0.2223690
#> log(c8/c10) -6.14830033 -6.52200310  0.2715720  1.9332486  6.2007304  0.3151410
#> log(c9/c10) -1.13191295 -1.26089747  1.0839233  1.6574554  3.1303882  1.0989612
#>                    s27         s28        s29        s30
#> log(c1/c10) -0.1559894 -1.05538125  0.3071767 -3.3638481
#> log(c2/c10)  0.7296313  0.38824654  0.9054366 -0.4879850
#> log(c3/c10) -0.7882065 -0.54578555 -0.9130476  0.0764356
#> log(c4/c10) -0.2524081 -2.58617875  0.9494300 -8.5762619
#> log(c5/c10) -1.2117344 -1.74040381 -0.9394818 -3.0973383
#> log(c6/c10)  2.8104040  2.40379373  3.0197989  1.3601482
#> log(c7/c10)  0.1319766 -0.02117221  0.2108447 -0.4142588
#> log(c8/c10) -1.1295868 -3.57733846  0.1309489 -9.8599765
#> log(c9/c10)  0.6003095 -0.24453854  1.0353868 -2.4130080
#> 
#> , , 2
#> 
#>                      s1          s2          s3          s4          s5
#> log(c1/c10) -0.52310718 -3.17751488 -0.20567657  -4.1887940 -1.23645158
#> log(c2/c10) -0.04747959 -1.47118251  0.12277567  -2.0135864 -0.43008491
#> log(c3/c10) -0.61838230  0.14481079 -0.70964969   0.4355729 -0.41328212
#> log(c4/c10) -0.86607098 -9.03078474  0.11031637 -12.1413863 -3.06025268
#> log(c5/c10) -0.67965687 -1.89269893 -0.53459374  -2.3548450 -1.00564927
#> log(c6/c10)  2.77724952  0.85530051  3.00708816   0.1230743  2.26074576
#> log(c7/c10)  0.07691142 -0.09287705  0.09721579  -0.1575633  0.03128254
#> log(c8/c10) -1.31112777 -7.28962063 -0.59618234  -9.5673133 -2.91778540
#> log(c9/c10) -0.60610468 -3.99455496 -0.20089268  -5.2854904 -1.51671537
#>                     s6         s7          s8          s9         s10
#> log(c1/c10)  2.2897630  1.5802947 -0.75771033 -0.88657436 -0.89567258
#> log(c2/c10)  1.4612153  1.0806890 -0.17330999 -0.24242677 -0.24730664
#> log(c3/c10) -1.4271364 -1.2231507 -0.55092940 -0.51387852 -0.51126261
#> log(c4/c10)  7.7860587  5.6037997 -1.58768867 -1.98406257 -2.01204786
#> log(c5/c10)  0.6058012  0.2815802 -0.78686854 -0.84575832 -0.84991613
#> log(c6/c10)  4.8139348  4.3002376  2.60738289  2.51407767  2.50749002
#> log(c7/c10)  0.2568359  0.2114550  0.06190509  0.05366234  0.05308038
#> log(c8/c10)  5.0242683  3.4263409 -1.83952183 -2.12976085 -2.15025267
#> log(c9/c10)  2.9846288  2.0789662 -0.90558434 -1.07008407 -1.08169829
#>                    s11        s12          s13         s14          s15
#> log(c1/c10)  0.4933372  1.2294272 -1.585482109 -2.01487268  0.946513063
#> log(c2/c10)  0.4976947  0.8924997 -0.617288928 -0.84759440  0.740757512
#> log(c3/c10) -0.9106296 -1.1222695 -0.312929148 -0.18947113 -1.040926295
#> log(c4/c10)  2.2604185  4.5245636 -4.133838474 -5.45460436  3.654345634
#> log(c5/c10) -0.2151503  0.1212367 -1.165153302 -1.36138119 -0.008052737
#> log(c6/c10)  3.5132157  4.0461886  2.008026901  1.69712259  3.841341960
#> log(c7/c10)  0.1419280  0.1890118  0.008956895 -0.01850896  0.170915287
#> log(c8/c10)  0.9781986  2.6360860 -3.703902976 -4.67101458  1.998881522
#> log(c9/c10)  0.6914245  1.6310708 -1.962265842 -2.51039890  1.269920295
#>                     s16         s17         s18         s19         s20
#> log(c1/c10) -0.59633141 -1.28457478 -2.87431002 -0.96788199 -1.16836661
#> log(c2/c10) -0.08675372 -0.45589600 -1.30855729 -0.28603646 -0.39356724
#> log(c3/c10) -0.59732894 -0.39944577  0.05763358 -0.49050102 -0.43285785
#> log(c4/c10) -1.09130196 -3.20827523 -8.09815449 -2.23415735 -2.85082958
#> log(c5/c10) -0.71311972 -1.02764118 -1.75413685 -0.88291523 -0.97453501
#> log(c6/c10)  2.72423082  2.22590170  1.07483886  2.45520611  2.31004333
#> log(c7/c10)  0.07222765  0.02820435 -0.07348264  0.04846152  0.03563758
#> log(c8/c10) -1.47604988 -3.02617275 -6.60671570 -2.31288912 -2.76443838
#> log(c9/c10) -0.69957813 -1.57814643 -3.60750267 -1.17387629 -1.42980237
#>                     s21         s22        s23        s24        s25        s26
#> log(c1/c10) -2.14097524 -2.25820890 -0.1270079  0.3942738  1.7330183 -0.1133400
#> log(c2/c10) -0.91523005 -0.97810882  0.1649699  0.4445616  1.1626030  0.1723008
#> log(c3/c10) -0.15321422 -0.11950730 -0.7322684 -0.8821469 -1.2670616 -0.7361982
#> log(c4/c10) -5.84248423 -6.20308418  0.3522938  1.9557085  6.0735634  0.3943352
#> log(c5/c10) -1.41900901 -1.47258381 -0.4986428 -0.2604214  0.3513736 -0.4923967
#> log(c6/c10)  1.60581684  1.52093270  3.0640489  3.4414879  4.4108185  3.0739453
#> log(c7/c10) -0.02657507 -0.03407389  0.1022478  0.1355915  0.2212239  0.1031221
#> log(c8/c10) -4.95503397 -5.21907804 -0.4189979  0.7550792  3.7703185 -0.3882137
#> log(c9/c10) -2.67137352 -2.82102665 -0.1004692  0.5649663  2.2739235 -0.0830216
#>                     s27         s28         s29        s30
#> log(c1/c10) -0.56656310 -1.33444314 -0.17112258 -3.3053589
#> log(c2/c10) -0.07078736 -0.48264310  0.14130885 -1.5397522
#> log(c3/c10) -0.60588789 -0.38510766 -0.71958462  0.1815684
#> log(c4/c10) -0.99973738 -3.36166571  0.21660126 -9.4240212
#> log(c5/c10) -0.69951586 -1.05043060 -0.51880286 -1.9511226
#> log(c6/c10)  2.74578485  2.18979404  3.03210730  0.7627338
#> log(c7/c10)  0.07413178  0.02501453  0.09942602 -0.1010546
#> log(c8/c10) -1.40900305 -3.13849071 -0.51835678 -7.5775623
#> log(c9/c10) -0.66157778 -1.64180525 -0.15678323 -4.1577526
#>