Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                       s1          s2          s3          s4          s5
#> log(c1/c10) -2.025936025 -0.95515443 -1.11664768 -0.67705541 -0.92900560
#> log(c2/c10) -1.228628913 -0.77702296 -0.84513332 -0.65973371 -0.76599460
#> log(c3/c10)  6.249796254  1.89911279  2.55527463  0.76917106  1.79286771
#> log(c4/c10) -0.840398060 -0.65800158 -0.68551027 -0.61063031 -0.65354740
#> log(c5/c10)  5.549833135  1.44585651  2.06481049  0.37998848  1.34563609
#> log(c6/c10)  1.390988074 -0.17735901  0.05917613 -0.58468370 -0.21565854
#> log(c7/c10) -3.897668545 -0.47206002 -0.98870381  0.41762502 -0.38840557
#> log(c8/c10)  4.052169368  1.25200755  1.67432265  0.52476097  1.18362670
#> log(c9/c10) -0.002063687  0.06411894  0.05413740  0.08130763  0.06573514
#>                      s6         s7          s8          s9         s10
#> log(c1/c10) -1.23262777 -0.2092084 -0.48055746 -0.64444314 -0.99496227
#> log(c2/c10) -0.89404833 -0.4624175 -0.57686000 -0.64597937 -0.79381206
#> log(c3/c10)  3.02651233 -1.1317343 -0.02921807  0.63666444  2.06085569
#> log(c4/c10) -0.70526627 -0.5309374 -0.57715893 -0.60507515 -0.66478243
#> log(c5/c10)  2.50932652 -1.4131254 -0.37312775  0.25499569  1.59842773
#> log(c6/c10)  0.22904927 -1.2699276 -0.87248934 -0.63245007 -0.11905345
#> log(c7/c10) -1.35974342  1.9143456  1.04625467  0.52195709 -0.59941193
#> log(c8/c10)  1.97761795 -0.6986888  0.01090636  0.43947782  1.35610758
#> log(c9/c10)  0.04696893  0.1102242  0.09345273  0.08332332  0.06165851
#>                     s11        s12         s13         s14         s15
#> log(c1/c10) -0.86082024 -0.5495735 -0.62128026 -0.66240698 -0.50066110
#> log(c2/c10) -0.73723718 -0.6059678 -0.63621034 -0.65355568 -0.58533878
#> log(c3/c10)  1.51582437  0.2512003  0.54255153  0.70965315  0.05246487
#> log(c4/c10) -0.64193273 -0.5889151 -0.60112959 -0.60813510 -0.58058337
#> log(c5/c10)  1.08430258 -0.1086105  0.16621948  0.32384555 -0.29607667
#> log(c6/c10) -0.31552794 -0.7714033 -0.66637617 -0.60613889 -0.84304404
#> log(c7/c10) -0.17026926  0.8254609  0.59605901  0.46448778  0.98193977
#> log(c8/c10)  1.00531766  0.1913876  0.37890541  0.48645439  0.06347866
#> log(c9/c10)  0.06994953  0.0891870  0.08475496  0.08221301  0.09221017
#>                     s16         s17         s18         s19        s20
#> log(c1/c10) -0.81560745 -1.21345043 -1.04484882 -0.51116243 -0.3541907
#> log(c2/c10) -0.71816853 -0.88596022 -0.81485189 -0.58976775 -0.5235644
#> log(c3/c10)  1.33212067  2.94859302  2.26354930  0.09513271 -0.5426578
#> log(c4/c10) -0.63423120 -0.70199961 -0.67328008 -0.58237217 -0.5556337
#> log(c5/c10)  0.91101586  2.43582565  1.78962754 -0.25582833 -0.8574528
#> log(c6/c10) -0.38174998  0.20096069 -0.04598587 -0.82766301 -1.0575756
#> log(c7/c10) -0.02562602 -1.29839190 -0.75900731  0.94834429  1.4505230
#> log(c8/c10)  0.88708335  1.92746797  1.48656408  0.09094028 -0.3195508
#> log(c9/c10)  0.07274403  0.04815424  0.05857513  0.09156110  0.1012632
#>                     s21         s22         s23         s24         s25
#> log(c1/c10) -0.99573964 -0.73617203 -0.80672376 -0.76588375 -0.77135242
#> log(c2/c10) -0.79413992 -0.68466635 -0.71442180 -0.69719738 -0.69950381
#> log(c3/c10)  2.06401422  1.00936727  1.29602542  1.13008873  1.15230842
#> log(c4/c10) -0.66491484 -0.62070021 -0.63271796 -0.62576129 -0.62669282
#> log(c5/c10)  1.60140716  0.60656429  0.87696739  0.72044020  0.74139992
#> log(c6/c10) -0.11791485 -0.49809707 -0.39476170 -0.45457904 -0.44656922
#> log(c7/c10) -0.60189888  0.22850113  0.00279439  0.13344837  0.11595320
#> log(c8/c10)  1.35814046  0.67935467  0.86385193  0.75705272  0.77135363
#> log(c9/c10)  0.06161046  0.07765376  0.07329311  0.07581734  0.07547934
#>                     s26        s27         s28         s29         s30
#> log(c1/c10) -0.87344574 -0.0990504 -0.41546453 -1.17721054 -1.33819384
#> log(c2/c10) -0.74256203 -0.4159580 -0.54940682 -0.87067592 -0.93857120
#> log(c3/c10)  1.56712291 -1.5793161 -0.29369661  2.80134704  3.45543690
#> log(c4/c10) -0.64408335 -0.5121732 -0.56607103 -0.69582652 -0.72324835
#> log(c5/c10)  1.13269222 -1.8353270 -0.62260895  2.29692929  2.91392878
#> log(c6/c10) -0.29703569 -1.4312732 -0.96782934  0.14788103  0.38366925
#> log(c7/c10) -0.21066032  2.2667593  1.25449780 -1.18245446 -1.69746683
#> log(c8/c10)  1.03833414 -0.9867588 -0.15931580  1.83269836  2.25367990
#> log(c9/c10)  0.06916918  0.1170328  0.09747598  0.05039415  0.04044413
#> 
#> , , 2
#> 
#>                     s1           s2         s3          s4          s5
#> log(c1/c10) -2.1976903 -1.080782411 -1.2492323 -0.79070367 -1.05350717
#> log(c2/c10)  0.3597683 -0.378016810 -0.2667455 -0.56963135 -0.39603376
#> log(c3/c10)  4.8897594  0.739037967  1.3650419 -0.33897044  0.63767603
#> log(c4/c10) -4.9087607 -1.165910176 -1.7303998 -0.19383234 -1.07450857
#> log(c5/c10)  2.4796995  1.660528934  1.7840747  1.44777732  1.64052453
#> log(c6/c10)  1.6111655  0.021848090  0.2615459 -0.39092293 -0.01696355
#> log(c7/c10) -5.2899562 -1.231765699 -1.8438143 -0.17778905 -1.13266339
#> log(c8/c10)  3.8436685  1.252713738  1.6434766  0.57980157  1.18944180
#> log(c9/c10) -0.3792606  0.001007987 -0.0563434  0.09976979  0.01029427
#>                      s6         s7         s8          s9         s10
#> log(c1/c10) -1.37020852 -0.3027031 -0.5857411 -0.75668656 -1.12230506
#> log(c2/c10) -0.18683337 -0.8919852 -0.7050215 -0.59210171 -0.35058859
#> log(c3/c10)  1.81462104 -2.1525081 -1.1006648 -0.46538692  0.89334699
#> log(c4/c10) -2.13580100  1.4414980  0.4930140 -0.07983819 -1.30505601
#> log(c5/c10)  1.87280190  1.0898644  1.2974522  1.42282824  1.69098277
#> log(c6/c10)  0.43369045 -1.0853292 -0.6825768 -0.43932799  0.08093325
#> log(c7/c10) -2.28337108  1.5953199  0.5669250 -0.05419074 -1.38263475
#> log(c8/c10)  1.92411196 -0.5522412  0.1043383  0.50089014  1.34903620
#> log(c9/c10) -0.09753162  0.2659172  0.1695525  0.11135145 -0.01312904
#>                     s11        s12          s13        s14        s15
#> log(c1/c10) -0.98238457 -0.6577302 -0.732525885 -0.7754242 -0.6067108
#> log(c2/c10) -0.44301453 -0.6574684 -0.608061291 -0.5797243 -0.6911698
#> log(c3/c10)  0.37336589 -0.8331347 -0.555174302 -0.3957528 -1.0227361
#> log(c4/c10) -0.83617089  0.2517728  0.001126234 -0.1426297  0.4227430
#> log(c5/c10)  1.58836129  1.3502510  1.405108148  1.4365710  1.3128319
#> log(c6/c10) -0.11816829 -0.5801392 -0.473707712 -0.4126650 -0.6527378
#> log(c7/c10) -0.87424545  0.3053589  0.033595033 -0.1222725  0.4907336
#> log(c8/c10)  1.02445467  0.2713354  0.444843247  0.5443570  0.1529828
#> log(c9/c10)  0.03450906  0.1450427  0.119577325  0.1049719  0.1624130
#>                     s16         s17         s18        s19        s20
#> log(c1/c10) -0.93522414 -1.35020507 -1.17434059 -0.6176645 -0.4539308
#> log(c2/c10) -0.47416684 -0.20004686 -0.31621598 -0.6839342 -0.7920902
#> log(c3/c10)  0.19810541  1.74028298  1.08672460 -0.9820293 -1.5905063
#> log(c4/c10) -0.67813239 -2.06876778 -1.47943135  0.3860363  0.9347213
#> log(c5/c10)  1.55377255  1.85813082  1.72914704  1.3208657  1.2007790
#> log(c6/c10) -0.18527579  0.40522629  0.15497782 -0.6371511 -0.8701379
#> log(c7/c10) -0.70289201 -2.21069022 -1.57170144  0.4509342  1.0458466
#> log(c8/c10)  0.91505394  1.87770882  1.46974596  0.1783927 -0.2014295
#> log(c9/c10)  0.05056556 -0.09072114 -0.03084535  0.1586837  0.2144293
#>                     s21        s22         s23          s24         s25
#> log(c1/c10) -1.12311592 -0.8523669 -0.92595777 -0.883358487 -0.88906273
#> log(c2/c10) -0.35005297 -0.5288991 -0.48028784 -0.508427242 -0.50465925
#> log(c3/c10)  0.89636036 -0.1098139  0.16366913  0.005359068  0.02655752
#> log(c4/c10) -1.30777327 -0.4004708 -0.64708000 -0.504326286 -0.52344167
#> log(c5/c10)  1.69157748  1.4930028  1.54697634  1.515732871  1.51991652
#> log(c6/c10)  0.08208707 -0.3031786 -0.19846149 -0.259078651 -0.25096173
#> log(c7/c10) -1.38558094 -0.4018371 -0.66922341 -0.514442513 -0.53516839
#> log(c8/c10)  1.35091719  0.7228452  0.89355820  0.794738222  0.80797068
#> log(c9/c10) -0.01340511  0.0787756  0.05372044  0.068224025  0.06628193
#>                     s26        s27        s28        s29        s30
#> log(c1/c10) -0.99555394 -0.1877998 -0.5178442 -1.3124041 -1.4803221
#> log(c2/c10) -0.43431536 -0.9678858 -0.7498715 -0.2250167 -0.1140967
#> log(c3/c10)  0.42230670 -2.5795187 -1.3529876  1.5998046  2.2238318
#> log(c4/c10) -0.88030253  1.8265483  0.7205423 -1.9420935 -2.5048006
#> log(c5/c10)  1.59802006  1.0055913  1.2476548  1.8304065  1.9535622
#> log(c6/c10) -0.09942878 -1.2488322 -0.7791916  0.3514369  0.5903778
#> log(c7/c10) -0.92209524  2.0128112  0.8136229 -2.0733435 -2.6834594
#> log(c8/c10)  1.05500440 -0.8187888 -0.0531661  1.7900197  2.1795486
#> log(c9/c10)  0.03002534  0.3050378  0.1926690 -0.0778512 -0.1350215
#>