Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1           s2          s3          s4          s5
#> log(c1/c10) -1.3630309  0.154192855  0.10946807 -0.18424000  0.70550565
#> log(c2/c10)  7.6469392  1.382226459  1.56689792  2.77963705 -0.89417886
#> log(c3/c10) -3.0784740 -1.314624244 -1.36661908 -1.70806962 -0.67369511
#> log(c4/c10) -0.2066078 -0.007157816 -0.01303721 -0.05164724  0.06531621
#> log(c5/c10)  5.7068866 -0.366392555 -0.18736418  0.98831674 -2.57323675
#> log(c6/c10) -0.3455983  1.081062032  1.03900688  0.76283032  1.59946684
#> log(c7/c10)  0.3080240  0.401969924  0.39920058  0.38101429  0.43610699
#> log(c8/c10)  6.8113522  4.862780667  4.92022074  5.29743021  4.15472926
#> log(c9/c10) -2.4277094 -0.010729089 -0.08197693 -0.54986216  0.86752774
#>                      s6         s7         s8          s9         s10
#> log(c1/c10)   2.6037577  1.4349880  2.1997516 -1.04382458 -0.52495908
#> log(c2/c10)  -8.7321813 -3.9062575 -7.0640145  6.32891603  4.18648763
#> log(c3/c10)   1.5331193  0.1743652  1.0634417 -2.70738037 -2.10417287
#> log(c4/c10)   0.3148551  0.1612119  0.2617456 -0.16464581 -0.09643721
#> log(c5/c10) -10.1717299 -5.4932740 -8.5545380  4.42913884  2.35217757
#> log(c6/c10)   3.3844118  2.2854062  3.0045208 -0.04544546  0.44244884
#> log(c7/c10)   0.5536460  0.4812763  0.5286301  0.32778916  0.35991710
#> log(c8/c10)   1.7168028  3.2178546  2.2356682  6.40139529  5.73501596
#> log(c9/c10)   3.8914968  2.0296137  3.2479036 -1.91920464 -1.09263726
#>                     s11         s12        s13        s14         s15
#> log(c1/c10)   3.3629623  0.44431394 -3.3125228 -3.3096552   3.4077328
#> log(c2/c10) -11.8669848  0.18429819 15.6965140 15.6846735 -12.0518452
#> log(c3/c10)   2.4157332 -0.97734374 -5.3448573 -5.3415235   2.4677812
#> log(c4/c10)   0.4146580  0.03098068 -0.4628825 -0.4625055   0.4205434
#> log(c5/c10) -13.2107418 -1.52771520 13.5104870 13.4990083 -13.3899534
#> log(c6/c10)   4.0982992  1.35386572 -2.1787245 -2.1760281   4.1403974
#> log(c7/c10)   0.6006557  0.41993411  0.1873123  0.1874898   0.6034278
#> log(c8/c10)   0.7417559  4.49017796  9.3150859  9.3114031   0.6842571
#> log(c9/c10)   5.1009310  0.45144198 -5.5333048 -5.5287366   5.1722518
#>                     s16         s17        s18         s19         s20
#> log(c1/c10) -0.36963119   2.5618745  0.5683271   3.1282748  0.63008082
#> log(c2/c10)  3.54512900  -8.5592431 -0.3277599 -10.8979456 -0.58274498
#> log(c3/c10) -1.92359631   1.4844280 -0.8331722   2.1428971 -0.76138031
#> log(c4/c10) -0.07601825   0.3093492  0.0472831   0.3838066  0.05540107
#> log(c5/c10)  1.73041722 -10.0040762 -2.0241261 -12.2713140 -2.27131949
#> log(c6/c10)  0.58850517   3.3450286  1.4704765   3.8776203  1.52854413
#> log(c7/c10)  0.36953494   0.5510526  0.4276130   0.5861239  0.43143672
#> log(c8/c10)  5.53552825   1.7705934  4.3309078   1.0431651  4.25159742
#> log(c9/c10) -0.84519557   3.8247757  0.6489984   4.7270673  0.74737386
#>                    s21        s22        s23         s24        s25        s26
#> log(c1/c10) -1.8850591  1.5748274  1.8918669 -0.17757677  2.4641567 -2.2337387
#> log(c2/c10)  9.8024265 -4.4836634 -5.7927395  2.75212415 -8.1557600 11.2421465
#> log(c3/c10) -3.6853583  0.3369356  0.7055102 -1.70032328  1.3708261 -4.0907160
#> log(c4/c10) -0.2752321  0.1795948  0.2212719 -0.05077131  0.2965035 -0.3210686
#> log(c5/c10)  7.7965077 -6.0530358 -7.3221099  0.96164457 -9.6129225  9.1922336
#> log(c6/c10) -0.8364665  2.4168986  2.7150139  0.76909582  3.2531436 -1.1643333
#> log(c7/c10)  0.2757002  0.4899351  0.5095660  0.38142688  0.5450020  0.2541101
#> log(c8/c10)  7.4817934  3.0382587  2.6310846  5.28887262  1.8960925  7.9296028
#> log(c9/c10) -3.2593150  2.2523819  2.7574347 -0.53924745  3.6691084 -3.8147714
#>                    s27         s28        s29        s30
#> log(c1/c10) -1.6585300 -0.94184197 -1.2556909 -2.3437768
#> log(c2/c10)  8.8670736  5.90782338  7.2037256 11.6965006
#> log(c3/c10) -3.4220067 -2.58882040 -2.9536858 -4.2186408
#> log(c4/c10) -0.2454533 -0.15123946 -0.1924972 -0.3355339
#> log(c5/c10)  6.8897368  4.02091370  5.2772165  9.6327038
#> log(c6/c10) -0.6234590  0.05044979 -0.2446655 -1.2678032
#> log(c7/c10)  0.2897268  0.33410388  0.3146705  0.2472966
#> log(c8/c10)  7.1908619  6.27041895  6.6734954  8.0709248
#> log(c9/c10) -2.8984478 -1.75674346 -2.2567137 -3.9900651
#> 
#> , , 2
#> 
#>                    s1         s2         s3         s4         s5         s6
#> log(c1/c10) -1.451583  0.1847362  0.1365008 -0.1802620  0.7793245  2.8265805
#> log(c2/c10)  5.439022  1.3151491  1.4367128  2.2350226 -0.1833403 -5.3428622
#> log(c3/c10) -3.024624 -1.1937046 -1.2476766 -1.6021107 -0.5284043  1.7623239
#> log(c4/c10) -0.899448 -0.1503725 -0.1724537 -0.3174617  0.1218187  1.0590134
#> log(c5/c10)  5.022570 -0.4927180 -0.3301381  0.7375254 -2.4968052 -9.3971758
#> log(c6/c10) -1.726274  0.8433587  0.7676109  0.2701749  1.7770846  4.9920421
#> log(c7/c10)  0.177847  0.4733981  0.4646859  0.4074723  0.5807924  0.9505669
#> log(c8/c10)  7.348901  5.7638897  5.8106127  6.1174433  5.1879449  3.2048811
#> log(c9/c10) -1.919185  0.7293081  0.6512357  0.1385337  1.6916894  5.0053118
#>                     s7         s8         s9        s10        s11          s12
#> log(c1/c10)  1.5660678  2.3908617 -1.1073200 -0.5477260   3.645379  0.497630460
#> log(c2/c10) -2.1661013 -4.2447580  4.5714065  3.1611101  -7.406409  0.526588822
#> log(c3/c10)  0.3519034  1.2747868 -2.6394201 -2.0132757   2.678499 -0.843599101
#> log(c4/c10)  0.4819748  0.8595497 -0.7418511 -0.4856797   1.433844 -0.007135453
#> log(c5/c10) -5.1485598 -7.9285657  3.8622170  1.9760796 -12.156974 -1.547342437
#> log(c6/c10)  3.0125658  4.3078007 -1.1856528 -0.3068809   6.277862  1.334719626
#> log(c7/c10)  0.7228936  0.8718676  0.2400275  0.3411011   1.098458  0.529912960
#> log(c8/c10)  4.4258701  3.6269377  7.0154332  6.4733854   2.411756  5.460806318
#> log(c9/c10)  2.9650868  4.3000715 -1.3619724 -0.4562316   6.330593  1.235748577
#>                    s13        s14        s15         s16        s17         s18
#> log(c1/c10) -3.5541004 -3.5510077   3.693664 -0.38020556  2.7814097  0.63137807
#> log(c2/c10) 10.7378153 10.7300210  -7.528097  2.73892296 -5.2290222  0.18951631
#> log(c3/c10) -5.3771866 -5.3737261   2.732526 -1.82583276  1.7117811 -0.69394541
#> log(c4/c10) -1.8619405 -1.8605247   1.455948 -0.40899202  1.0383351  0.05409165
#> log(c5/c10) 12.1092032 12.0987791 -12.319720  1.41144436 -9.2449255 -1.99814491
#> log(c6/c10) -5.0280130 -5.0231563   6.353687 -0.04381121  4.9211070  1.54475339
#> log(c7/c10) -0.2019089 -0.2013503   1.107179  0.37135861  0.9424081  0.55407039
#> log(c8/c10)  9.3854942  9.3824985   2.364985  6.31111759  3.2486355  5.33125240
#> log(c9/c10) -5.3222521 -5.3172463   6.408745 -0.18508846  4.9321998  1.45222813
#>                    s19         s20         s21        s22        s23        s24
#> log(c1/c10)   3.392270  0.69797917 -2.01458750  1.7168840  2.0588096 -0.1730758
#> log(c2/c10)  -6.768520  0.02166731  6.85791425 -2.5461903 -3.4079158  2.2169116
#> log(c3/c10)   2.395288 -0.61942369 -3.65458539  0.5206555  0.9032450 -1.5940698
#> log(c4/c10)   1.317975  0.08458036 -1.15718087  0.5510155  0.7075426 -0.3141720
#> log(c5/c10) -11.303858 -2.22262701  6.92020404 -5.6568927 -6.8093688  0.7133038
#> log(c6/c10)   5.880386  1.64934203 -2.61040205  3.2494036  3.7863547  0.2814600
#> log(c7/c10)   1.052741  0.56609986  0.07615726  0.7501340  0.8118924  0.4087703
#> log(c8/c10)   2.656929  5.26673958  7.89425311  4.2797828  3.9485783  6.1104824
#> log(c9/c10)   5.920918  1.56002653 -2.83044626  3.2091930  3.7626227  0.1501652
#>                    s25          s26        s27        s28        s29
#> log(c1/c10)  2.6760214 -2.390636803 -1.7702769 -0.9973322 -1.3358169
#> log(c2/c10) -4.9634214  7.805638756  6.2421995  4.2942138  5.1472674
#> log(c3/c10)  1.5938595 -4.075356781 -3.3812199 -2.5163519 -2.8950912
#> log(c4/c10)  0.9900903 -1.329329060 -1.0453402 -0.6915008 -0.8464526
#> log(c5/c10) -8.8897097  8.187695567  6.0967441  3.4914982  4.6323763
#> log(c6/c10)  4.7556081 -3.200940093 -2.2267431 -1.0129308 -1.5444783
#> log(c7/c10)  0.9233730  0.008235404  0.1202845  0.2598935  0.1987565
#> log(c8/c10)  3.3507193  8.258511307  7.6576030  6.9088941  7.2367655
#> log(c9/c10)  4.7616218 -3.439107518 -2.4350131 -1.1839497 -1.7318100
#>                     s30
#> log(c1/c10) -2.50931234
#> log(c2/c10)  8.10472643
#> log(c3/c10) -4.20814593
#> log(c4/c10) -1.38365645
#> log(c5/c10)  8.58769692
#> log(c6/c10) -3.38730506
#> log(c7/c10) -0.01319972
#> log(c8/c10)  8.37346574
#> log(c9/c10) -3.63119190
#>