Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1         s2          s3         s4          s5         s6
#> log(c1/c10) -1.2074849 -1.2866954 -0.87475325 -0.7452220 -0.87438257 -4.9101179
#> log(c2/c10)  2.2250304  2.2670983  2.04832010  1.9795274  2.04812324  4.1914599
#> log(c3/c10)  0.7296990  0.9000598  0.01408094 -0.2645066  0.01328371  8.6930849
#> log(c4/c10)  7.0300522  7.0935187  6.76345451  6.6596690  6.76315751  9.9967467
#> log(c5/c10)  0.9977310  0.7919978  1.86193410  2.1983653  1.86289685 -8.6191075
#> log(c6/c10)  1.3966347  1.4067124  1.35430212  1.3378222  1.35425496  1.8677105
#> log(c7/c10)  8.3542077  8.1502441  9.21097713  9.5445144  9.21193160 -1.1799093
#> log(c8/c10)  0.3715309  0.3657812  0.39568335  0.4050858  0.39571026  0.1027633
#> log(c9/c10)  1.5841794  1.4980376  1.94602646  2.0868922  1.94642957 -2.4424490
#>                     s7         s8         s9        s10        s11        s12
#> log(c1/c10) -1.9665677 -1.0823516 -2.5313527 -1.0897166 -1.8927197 -1.7097687
#> log(c2/c10)  2.6281713  2.1585734  2.9281227  2.1624849  2.5889514  2.4917880
#> log(c3/c10)  2.3622856  0.4605702  3.5769890  0.4764103  2.2034580  1.8099787
#> log(c4/c10)  7.6382591  6.9297904  8.0907869  6.9356915  7.5790891  7.4325016
#> log(c5/c10) -0.9738327  1.3227395 -2.4407471  1.3036104 -0.7820274 -0.3068492
#> log(c6/c10)  1.4932107  1.3807143  1.5650667  1.3816513  1.4838152  1.4605388
#> log(c7/c10)  6.3996028  8.6764205  4.9453064  8.6574560  6.5897583  7.0608491
#> log(c8/c10)  0.3164304  0.3806141  0.2754337  0.3800795  0.3217909  0.3350710
#> log(c9/c10)  0.7586738  1.7202624  0.1444678  1.7122530  0.8389838  1.0379438
#>                    s13        s14        s15        s16        s17         s18
#> log(c1/c10) -2.4479096 -0.1330144  0.5332048 -0.2732752 -0.5089015  0.05491419
#> log(c2/c10)  2.8838069  1.6543904  1.3005684  1.7288814  1.8540200  1.55458348
#> log(c3/c10)  3.3975250 -1.5812035 -3.0140650 -1.2795395 -0.7727697 -1.98538826
#> log(c4/c10)  8.0239290  6.1691443  5.6353434  6.2815268  6.4703197  6.01856855
#> log(c5/c10) -2.2240207  3.7884504  5.5188196  3.4241513  2.8121600  4.27655679
#> log(c6/c10)  1.5544505  1.2599327  1.1751715  1.2777777  1.3077558  1.23602309
#> log(c7/c10)  5.1601686 11.1209221 12.8364071 10.7597566 10.1530294 11.60482988
#> log(c8/c10)  0.2814907  0.4495249  0.4978846  0.4393436  0.4222399  0.46316629
#> log(c9/c10)  0.2352125  2.7526705  3.4771865  2.6001362  2.3438918  2.95704357
#>                    s19        s20        s21        s22        s23        s24
#> log(c1/c10) -3.0637324 -2.3490342 -1.9597503 -3.0257952 -0.2861852 -1.7197426
#> log(c2/c10)  3.2108639  2.8312953  2.6245506  3.1907159  1.7357378  2.4970851
#> log(c3/c10)  4.7219970  3.1848701  2.3476231  4.6404043 -1.2517735  1.8314298
#> log(c4/c10)  8.5173503  7.9447062  7.6327967  8.4869536  6.2918707  7.4404930
#> log(c5/c10) -3.8234953 -1.9672118 -0.9561258 -3.7249614  3.3906202 -0.3327543
#> log(c6/c10)  1.6327999  1.5418708  1.4923433  1.6279733  1.2794202  1.4618078
#> log(c7/c10)  3.5744522  5.4147685  6.4171574  3.6721386 10.7265139  7.0351668
#> log(c8/c10)  0.2367892  0.2886679  0.3169253  0.2395430  0.4384065  0.3343470
#> log(c9/c10) -0.4344972  0.3427399  0.7660877 -0.3932404  2.5860966  1.0270971
#>                    s25        s26        s27        s28        s29        s30
#> log(c1/c10) -3.3856606 -3.2441721  0.5603346 -1.9704275 -2.9260967 -2.4285687
#> log(c2/c10)  3.3818366  3.3066936  1.2861600  2.6302212  3.1377671  2.8735352
#> log(c3/c10)  5.4143797  5.1100753 -3.0724141  2.3705869  4.4259790  3.3559278
#> log(c4/c10)  8.7752918  8.6619257  5.6136059  7.6413516  8.4070712  8.0084323
#> log(c5/c10) -4.6596387 -4.2921510  5.5892839 -0.9838576 -3.4660145 -2.1737866
#> log(c6/c10)  1.6737580  1.6557568  1.1717198  1.4937017  1.6152889  1.5519898
#> log(c7/c10)  2.7455011  3.1098278 12.9062652  6.3896642  3.9288581  5.2099707
#> log(c8/c10)  0.2134210  0.2236914  0.4998539  0.3161503  0.2467799  0.2828946
#> log(c9/c10) -0.7845955 -0.6307262  3.5066903  0.7544763 -0.2848178  0.2562458
#> 
#> , , 2
#> 
#>                     s1         s2         s3         s4         s5          s6
#> log(c1/c10) -1.9129800 -1.9224031 -1.8733970 -1.8579874 -1.8733529 -2.35345909
#> log(c2/c10)  2.1380467  2.2359713  1.7267045  1.5665705  1.7262462  6.71545597
#> log(c3/c10) -1.4669885 -1.2322488 -2.4530362 -2.8369010 -2.4541347  9.50573414
#> log(c4/c10)  6.3929163  6.5102197  5.9001713  5.7083475  5.8996224 11.87617452
#> log(c5/c10)  0.4138725  0.2972690  0.9036774  1.0943566  0.9042230 -5.03666811
#> log(c6/c10)  1.1881454  1.2324908  1.0018682  0.9293511  1.0016607  3.26103575
#> log(c7/c10)  7.4233070  7.3654305  7.6664231  7.7610674  7.6666940  4.71791501
#> log(c8/c10)  0.1691234  0.1665583  0.1798983  0.1840930  0.1799103  0.04922046
#> log(c9/c10)  0.8824662  0.7787931  1.3179555  1.4874899  1.3184406 -3.96365186
#>                     s7         s8         s9        s10        s11         s12
#> log(c1/c10) -2.0032833 -1.8980936 -2.0704722 -1.8989698 -1.9944981 -1.97273352
#> log(c2/c10)  3.0764688  1.9833496  3.7746887  1.9924546  2.9851736  2.75899892
#> log(c3/c10)  0.7825472 -1.8378201  2.4562829 -1.8159941  0.5636991  0.02152527
#> log(c4/c10)  7.5170480  6.2076053  8.3534425  6.2185122  7.4076858  7.13675221
#> log(c5/c10) -0.7035517  0.5980777 -1.5349556  0.5872360 -0.5948421 -0.32552508
#> log(c6/c10)  1.6131120  1.1180905  1.9293025  1.1222137  1.5717688  1.46934509
#> log(c7/c10)  6.8686702  7.5147379  6.4560004  7.5093565  6.9226286  7.05630485
#> log(c8/c10)  0.1445419  0.1731756  0.1262524  0.1729371  0.1469334  0.15285790
#> log(c9/c10) -0.1110444  1.0462445 -0.8502520  1.0366050 -0.0143898  0.22506203
#>                    s13        s14         s15        s16        s17        s18
#> log(c1/c10) -2.0605455 -1.7851569 -1.70590098 -1.8018429 -1.8298738 -1.7628002
#> log(c2/c10)  3.6715315  0.8097239 -0.01389476  0.9831225  1.2744173  0.5773958
#> log(c3/c10)  2.2090000 -4.6511732 -6.62550830 -4.2355114 -3.5372350 -5.2080980
#> log(c4/c10)  8.2298710  4.8017244  3.81511521  5.0094378  5.3583785  4.5234195
#> log(c5/c10) -1.4121214  1.9955700  2.97629231  1.7890960  1.4422374  2.2722143
#> log(c6/c10)  1.8825876  0.5866114  0.21363382  0.6651354  0.7970489  0.4814011
#> log(c7/c10)  6.5169695  8.2083873  8.69517172  8.1059033  7.9337390  8.3457005
#> log(c8/c10)  0.1289546  0.2039182  0.22549248  0.1993761  0.1917458  0.2100039
#> log(c9/c10) -0.7410392  2.2887658  3.16073354  2.1051881  1.7967934  2.5347324
#>                    s19        s20        s21        s22        s23         s24
#> log(c1/c10) -2.1338061 -2.0487830 -2.0024723 -2.1292930 -1.8033787 -1.97392005
#> log(c2/c10)  4.4328472  3.5492960  3.0680407  4.3859472  0.9990826  2.77132921
#> log(c3/c10)  4.0339856  1.9159835  0.7623438  3.9215592 -4.1972527  0.05108278
#> log(c4/c10)  9.1418477  8.0834456  7.5069520  9.0856663  5.0285563  7.15152261
#> log(c5/c10) -2.3186565 -1.2665697 -0.6935159 -2.2628103  1.7700916 -0.34020735
#> log(c6/c10)  2.2273511  1.8272329  1.6092954  2.2061123  0.6723629  1.47492889
#> log(c7/c10)  6.0670082  6.5892146  6.8736515  6.0947276  8.0964704  7.04901726
#> log(c8/c10)  0.1090123  0.1321565  0.1447627  0.1102408  0.1989580  0.15253491
#> log(c9/c10) -1.5470466 -0.6116281 -0.1021215 -1.4973933  2.0882911  0.21200791
#>                     s25        s26         s27        s28        s29        s30
#> log(c1/c10) -2.17210392 -2.1552719 -1.70267351 -2.0037425 -2.1174325 -2.0582447
#> log(c2/c10)  4.83083358  4.6559173 -0.04743423  3.0812404  4.2626940  3.6476212
#> log(c3/c10)  4.98801736  4.5687173 -6.70590733  0.7939855  3.6261033  2.1516833
#> log(c4/c10)  9.61859374  9.4090623  3.77493842  7.5227639  8.9380219  8.2012288
#> log(c5/c10) -2.79255790 -2.5842767  3.01622937 -0.7092335 -2.1160469 -1.3836502
#> log(c6/c10)  2.40758017  2.3283689  0.19844539  1.6152729  2.1502969  1.8717597
#> log(c7/c10)  5.83178578  5.9351668  8.71499460  6.8658500  6.1675740  6.5311013
#> log(c8/c10)  0.09858721  0.1031691  0.22637103  0.1444169  0.1134693  0.1295809
#> log(c9/c10) -1.96839603 -1.7832116  3.19624189 -0.1160961 -1.3669048 -0.7157252
#>