Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1         s2          s3         s4         s5         s6
#> log(c1/c10) -1.0185092 -1.1409277  -1.7211867 -1.3119869 -1.3660959 -0.9457429
#> log(c2/c10) -8.7234905 -7.5488439  -1.9810649 -5.9074735 -5.3882794 -9.4217076
#> log(c3/c10)  1.0620515  0.9863461   0.6275053  0.8805606  0.8470987  1.1070513
#> log(c4/c10)  6.5430602  4.4584583  -5.4224729  1.5455791  0.6241845  7.7821602
#> log(c5/c10) -8.1057473 -5.4780325   6.9772321 -1.8062449 -0.6447943 -9.6676769
#> log(c6/c10)  0.3856248  0.1615418  -0.9006024 -0.1515762 -0.2506209  0.5188210
#> log(c7/c10) -1.6711803 -1.4672980  -0.5009034 -1.1824067 -1.0922907 -1.7923693
#> log(c8/c10)  1.2156953  1.0662282   0.3577600  0.8573731  0.7913086  1.3045394
#> log(c9/c10)  7.3768143  4.2200271 -10.7430201 -0.1910499 -1.5863507  9.2532278
#>                     s7         s8         s9        s10        s11        s12
#> log(c1/c10) -1.3711854 -1.4008658 -1.2922416 -1.2680666 -1.3905672 -1.1401535
#> log(c2/c10) -5.3394438 -5.0546510 -6.0969365 -6.3289039 -5.1534689 -7.5562726
#> log(c3/c10)  0.8439513  0.8255965  0.8927714  0.9077215  0.8319653  0.9868249
#> log(c4/c10)  0.5375177  0.0321065  1.8818119  2.2934759  0.2074750  4.4716418
#> log(c5/c10) -0.5355477  0.1015411 -2.2300783 -2.7489954 -0.1195172 -5.4946508
#> log(c6/c10) -0.2599371 -0.3142659 -0.1154331 -0.0711815 -0.2954148  0.1629590
#> log(c7/c10) -1.0838143 -1.0343831 -1.2152917 -1.2555540 -1.0515348 -1.4685874
#> log(c8/c10)  0.7850946  0.7488563  0.8814812  0.9109978  0.7614303  1.0671735
#> log(c9/c10) -1.7175933 -2.4829556  0.3181195  0.9415170 -2.2173887  4.2399914
#>                    s13         s14          s15         s16          s17
#> log(c1/c10) -1.5034205 -1.25141257 -1.402822333 -1.26613328 -1.396376591
#> log(c2/c10) -4.0706040 -6.48870456 -5.035877304 -6.34745446 -5.097726354
#> log(c3/c10)  0.7621752  0.91802063  0.824386583  0.90891712  0.828372725
#> log(c4/c10) -1.7142452  2.57706826 -0.001210491  2.32639692  0.108550727
#> log(c5/c10)  2.3028795 -3.10647363  0.143538311 -2.79049350  0.005180397
#> log(c6/c10) -0.5019889 -0.04069693 -0.317847313 -0.06764268 -0.306048603
#> log(c7/c10) -0.8635829 -1.28329050 -1.031124538 -1.25877384 -1.041859624
#> log(c8/c10)  0.6236419  0.93133148  0.746467474  0.91335821  0.754337414
#> log(c9/c10) -5.1275184  1.37097100 -2.533408674  0.99137043 -2.367193339
#>                    s18        s19        s20        s21        s22        s23
#> log(c1/c10) -1.3530181 -1.5270311 -1.2893225 -1.2511749 -1.5179502 -1.3553849
#> log(c2/c10) -5.5137652 -3.8440523 -6.1249457 -6.4909856 -3.9311872 -5.4910558
#> log(c3/c10)  0.8551862  0.7475740  0.8945765  0.9181676  0.7531898  0.8537226
#> log(c4/c10)  0.8468794 -2.1162982  1.9315189  2.5811163 -1.9616632  0.8065779
#> log(c5/c10) -0.9255091  2.8096815 -2.2927357 -3.1115763  2.6147585 -0.8747076
#> log(c6/c10) -0.2266824 -0.5452073 -0.1100898 -0.0402618 -0.5285849 -0.2310146
#> log(c7/c10) -1.1140711 -0.8242605 -1.2201532 -1.2836864 -0.8393844 -1.1101295
#> log(c8/c10)  0.8072760  0.5948145  0.8850452  0.9316217  0.6059019  0.8043863
#> log(c9/c10) -1.2491158 -5.7363616  0.3933925  1.3771010 -5.5021922 -1.3101459
#>                   s24        s25        s26        s27        s28        s29
#> log(c1/c10) -1.098445 -1.1727994 -1.4693769 -1.1493947 -1.3304781 -1.3245309
#> log(c2/c10) -7.956480 -7.2430241 -4.3972643 -7.4676004 -5.7300446 -5.7871096
#> log(c3/c10)  1.012618  0.9666362  0.7832283  0.9811100  0.8691253  0.8728032
#> log(c4/c10)  5.181874  3.9157313 -1.1345333  4.3142785  1.2307025  1.3319737
#> log(c5/c10) -6.389924 -4.7939059  1.5721320 -5.2962888 -1.4093317 -1.5369877
#> log(c6/c10)  0.239305  0.1032018 -0.4396731  0.1460433 -0.1854236 -0.1745376
#> log(c7/c10) -1.538051 -1.4142171 -0.9202810 -1.4531966 -1.1516105 -1.1615153
#> log(c8/c10)  1.118098  1.0273144  0.6652076  1.0558904  0.8347963  0.8420575
#> log(c9/c10)  5.315522  3.3981562 -4.2496400  4.0016905 -0.6678788 -0.5145202
#>                    s30
#> log(c1/c10) -1.3758164
#> log(c2/c10) -5.2950083
#> log(c3/c10)  0.8410875
#> log(c4/c10)  0.4586596
#> log(c5/c10) -0.4361443
#> log(c6/c10) -0.2684139
#> log(c7/c10) -1.0761017
#> log(c8/c10)  0.7794404
#> log(c9/c10) -1.8370109
#> 
#> , , 2
#> 
#>                     s1         s2          s3         s4          s5
#> log(c1/c10) -1.3305329 -1.3583424  -1.4901585 -1.3972015 -1.40949332
#> log(c2/c10) -9.7842288 -8.3381320  -1.4836900 -6.3174559 -5.67828074
#> log(c3/c10)  1.2250923  0.9769941  -0.1989818  0.6303187  0.52065919
#> log(c4/c10)  5.2498682  3.6281113  -4.0589525  1.3619799  0.64516297
#> log(c5/c10) -8.6504070 -5.7213988   8.1619849 -1.6286045 -0.33398208
#> log(c6/c10)  0.5484224  0.3811594  -0.4116607  0.1474375  0.07350725
#> log(c7/c10) -1.8808174 -1.7819362  -1.3132434 -1.6437665 -1.60006100
#> log(c8/c10)  2.3078949  1.8319707  -0.4238913  1.1669470  0.95658836
#> log(c9/c10)  4.6916608  1.8930679 -11.3721522 -2.0174930 -3.25447181
#>                      s6          s7          s8         s9        s10
#> log(c1/c10)  -1.3140027 -1.41064950 -1.41739191 -1.3927160 -1.3872242
#> log(c2/c10) -10.6437975 -5.61815964 -5.26755384 -6.5507020 -6.8362750
#> log(c3/c10)   1.3725634  0.51034457  0.45019321  0.6703353  0.7193294
#> log(c4/c10)   6.2138503  0.57773884  0.18454431  1.6235589  1.9438209
#> log(c5/c10) -10.3914273 -0.21220931  0.49792797 -2.1010346 -2.6794507
#> log(c6/c10)   0.6478445  0.06655333  0.02600046  0.1744160  0.2074468
#> log(c7/c10)  -1.9395930 -1.59595004 -1.57197632 -1.6597154 -1.6792423
#> log(c8/c10)   2.5907871  0.93680193  0.82141423  1.2437105  1.3376953
#> log(c9/c10)   6.3551613 -3.37082260 -4.04934071 -1.5660980 -1.0134362
#>                     s11        s12        s13        s14         s15        s16
#> log(c1/c10) -1.41505242 -1.3581665 -1.4406890 -1.3834410 -1.41783637 -1.3867850
#> log(c2/c10) -5.38920769 -8.3472775 -4.0561026 -7.0330041 -5.24444171 -6.8591124
#> log(c3/c10)  0.47106464  0.9785631  0.2423517  0.7530810  0.44622800  0.7232475
#> log(c4/c10)  0.32097568  3.6383677 -1.1740642  2.1644471  0.15862470  1.9694325
#> log(c5/c10)  0.25152321 -5.7399226  2.9516715 -3.0779173  0.54474062 -2.7257069
#> log(c6/c10)  0.04007157  0.3822172 -0.1141222  0.2302015  0.02332719  0.2100883
#> log(c7/c10) -1.58029477 -1.7825616 -1.4891397 -1.6926942 -1.57039596 -1.6808039
#> log(c8/c10)  0.86145167  1.8349806  0.4227141  1.4024407  0.81380781  1.3452113
#> log(c9/c10) -3.81390723  1.9107669 -6.3938303 -0.6327115 -4.09406901 -0.9692396
#>                     s17         s18        s19        s20        s21        s22
#> log(c1/c10) -1.41637211 -1.40652247 -1.4460526 -1.3920529 -1.3833870 -1.4439897
#> log(c2/c10) -5.32058351 -5.83276511 -3.7771968 -6.5851839 -7.0358123 -3.8844677
#> log(c3/c10)  0.45929120  0.54716315  0.1945015  0.6762512  0.7535628  0.2129053
#> log(c4/c10)  0.24401559  0.81841284 -1.4868492  1.6622294  2.1675964 -1.3665478
#> log(c5/c10)  0.39051861 -0.64688366  3.5165835 -2.1708763 -3.0836050  3.2993106
#> log(c6/c10)  0.03213415  0.09137571 -0.1463819  0.1784043  0.2305263 -0.1339744
#> log(c7/c10) -1.57560238 -1.61062432 -1.4700687 -1.6620732 -1.6928863 -1.4774037
#> log(c8/c10)  0.83886680  1.00743063  0.3309235  1.2550588  1.4033649  0.3662274
#> log(c9/c10) -3.94671380 -2.95550231 -6.9335893 -1.4993661 -0.6272770 -6.7259906
#>                     s23        s24        s25         s26        s27        s28
#> log(c1/c10) -1.40706011 -1.3486917 -1.3655826 -1.43295541 -1.3602658 -1.4014021
#> log(c2/c10) -5.80480778 -8.8399686 -7.9616401 -4.45825105 -8.2381139 -6.0990247
#> log(c3/c10)  0.54236668  1.0630912  0.9124016  0.31134594  0.9598346  0.5928437
#> log(c4/c10)  0.78705948  4.1909069  3.2058862 -0.72306599  3.5159438  1.1170155
#> log(c5/c10) -0.59025726 -6.7378477 -4.9588303  2.13713670 -5.5188164 -1.1861813
#> log(c6/c10)  0.08814202  0.4392044  0.3376124 -0.06760767  0.3695907  0.1221727
#> log(c7/c10) -1.60871266 -1.8162508 -1.7561925 -1.51663783 -1.7750972 -1.6288306
#> log(c8/c10)  0.99822961  1.9971299  1.7080636  0.55506499  1.7990538  1.0950592
#> log(c9/c10) -3.00960739  2.8642589  1.1644530 -5.61556297  1.6995055 -2.4402171
#>                    s29         s30
#> log(c1/c10) -1.4000511 -1.41170150
#> log(c2/c10) -6.1692770 -5.56345548
#> log(c3/c10)  0.6048965  0.50095930
#> log(c4/c10)  1.1958014  0.51638968
#> log(c5/c10) -1.3284743 -0.10140835
#> log(c6/c10)  0.1302984  0.06022597
#> log(c7/c10) -1.6336343 -1.59220948
#> log(c8/c10)  1.1181799  0.91879827
#> log(c9/c10) -2.3042598 -3.47669009
#>