Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1        s2         s3         s4         s5         s6
#> log(c1/c10)  -2.685091 -1.709961 -2.4227792 -2.5497684 -2.5515253 -2.3932158
#> log(c2/c10)  -3.224075  7.474178 -0.3462298 -1.7394410 -1.7587168 -0.0218857
#> log(c3/c10)   5.011733 -3.120882  2.8240478  3.8831412  3.8977943  2.5774873
#> log(c4/c10)  -8.709447 11.796886 -3.1932145 -5.8637115 -5.9006591 -2.5715141
#> log(c5/c10)   1.382025 -1.492664  0.6087296  0.9830944  0.9882739  0.5215762
#> log(c6/c10)  -4.576717  6.897219 -1.4902119 -2.9844386 -3.0051119 -1.1423510
#> log(c7/c10)   3.273325 -6.737687  0.5803484  1.8840615  1.9020990  0.2768397
#> log(c8/c10) -10.460049 12.984981 -4.1533025 -7.2064998 -7.2487422 -3.4425082
#> log(c9/c10)  -1.490790  5.161728  0.2987468 -0.5675967 -0.5795830  0.5004344
#>                      s7         s8          s9         s10        s11
#> log(c1/c10) -2.47099600 -2.1595679 -2.27440702 -2.22894069 -1.9235752
#> log(c2/c10) -0.87522107  2.5414904  1.28157785  1.78039383  5.1305916
#> log(c3/c10)  3.22617715  0.6288561  1.58661836  1.20742765 -1.3393310
#> log(c4/c10) -4.20718107  2.3419462 -0.07304481  0.88308196  7.3047164
#> log(c5/c10)  0.75087291 -0.1672194  0.17132718  0.03729212 -0.8629276
#> log(c6/c10) -2.05755792  1.6068839  0.25562080  0.79060368  4.3837093
#> log(c7/c10)  1.07535795 -2.1218684 -0.94289101 -1.40966370 -4.5446491
#> log(c8/c10) -5.31257734  2.1750842 -0.58599134  0.50715488  7.8490530
#> log(c9/c10) -0.03019696  2.0944241  1.31096995  1.62114976  3.7044105
#>                    s12        s13        s14       s15        s16       s17
#> log(c1/c10) -2.0670565  -2.846089 -2.3768749 -1.763211 -1.9030633 -2.573098
#> log(c2/c10)  3.5564427  -4.990406  0.1573914  6.889964  5.3556298 -1.995392
#> log(c3/c10) -0.1426920   6.354465  2.4412042 -2.676773 -1.5104008  4.077710
#> log(c4/c10)  4.2873993 -12.095137 -2.2278770 10.677068  7.7360679 -6.354316
#> log(c5/c10) -0.4399436   1.856650  0.4734033 -1.335682 -0.9233968  1.051870
#> log(c6/c10)  2.6954258  -6.471117 -0.9500753  6.270646  4.6250639 -3.258947
#> log(c7/c10) -3.0716215   4.926189  0.1090791 -6.191002 -4.7552312  2.123570
#> log(c8/c10)  4.3993339 -14.330931 -3.0496256 11.704686  8.3422201 -7.767411
#> log(c9/c10)  2.7255540  -2.589151  0.6119147  4.798445  3.8443464 -0.726755
#>                    s18        s19        s20        s21         s22        s23
#> log(c1/c10) -2.4895635 -2.4831538 -1.8957012  -2.828101 -2.39294854 -2.4022910
#> log(c2/c10) -1.0789275 -1.0086053  5.4363996  -4.793063 -0.01895396 -0.1214513
#> log(c3/c10)  3.3810310  3.3275734 -1.5718005   6.204448  2.57525868  2.6531753
#> log(c4/c10) -4.5976440 -4.4628509  7.8908868 -11.716871 -2.56589459 -2.7623608
#> log(c5/c10)  0.8056101  0.7867141 -0.9451001   1.803622  0.52078843  0.5483301
#> log(c6/c10) -2.2760342 -2.2006132  4.7116900  -6.259465 -1.13920673 -1.2491357
#> log(c7/c10)  1.2659785  1.2001737 -4.8308124   4.741523  0.27409633  0.3700094
#> log(c8/c10) -5.7589962 -5.6048864  8.5192256 -13.898457 -3.43608339 -3.6607045
#> log(c9/c10) -0.1568681 -0.1131396  3.8945717  -2.466437  0.50225746  0.4385213
#>                    s24         s25        s26        s27        s28       s29
#> log(c1/c10) -2.5582257 -2.45583287  -2.746641 -2.1785159 -1.9560144 -2.637427
#> log(c2/c10) -1.8322271 -0.70886470  -3.899356  2.3336097  4.7746976 -2.701154
#> log(c3/c10)  3.9536755  3.09971610   5.525069  0.7868831 -1.0687869  4.614218
#> log(c4/c10) -6.0415631 -3.88831039 -10.003821  1.9434821  6.6225414 -7.707117
#> log(c5/c10)  1.0080266  0.70617188   1.563477 -0.1113605 -0.7672966  1.241513
#> log(c6/c10) -3.0839521 -1.87913979  -5.300959  1.3839308  4.0020110 -4.015881
#> log(c7/c10)  1.9708871  0.91968809   3.905227 -1.9273417 -4.2116173  2.783996
#> log(c8/c10) -7.4098387 -4.94801034 -11.939915  1.7195175  7.0691178 -9.314077
#> log(c9/c10) -0.6252941  0.07324878  -1.910701  1.9651572  3.4831042 -1.165621
#>                    s30
#> log(c1/c10) -2.5516758
#> log(c2/c10) -1.7603675
#> log(c3/c10)  3.8990492
#> log(c4/c10) -5.9038232
#> log(c5/c10)  0.9887174
#> log(c6/c10) -3.0068823
#> log(c7/c10)  1.9036437
#> log(c8/c10) -7.2523598
#> log(c9/c10) -0.5806095
#> 
#> , , 2
#> 
#>                     s1        s2         s3         s4         s5           s6
#> log(c1/c10)  -2.065776 -1.609722 -1.9430967 -2.0024876 -2.0033093 -1.929270293
#> log(c2/c10)  -2.580520  7.502641  0.1318644 -1.1812447 -1.1994122  0.437560488
#> log(c3/c10)   4.294144 -2.114844  2.5701171  3.4047462  3.4162937  2.375812685
#> log(c4/c10)  -8.922090 11.860276 -3.3316045 -6.0380487 -6.0754936 -2.701535466
#> log(c5/c10)   1.611765 -1.732806  0.7120710  1.1476276  1.1536537  0.610672055
#> log(c6/c10)  -2.242794  5.267673 -0.2224683 -1.2005407 -1.2140728  0.005230168
#> log(c7/c10)   4.362720 -8.178891  0.9890087  2.6222765  2.6448735  0.608778633
#> log(c8/c10) -10.359792 13.218569 -4.0171799 -7.0877407 -7.1302234 -3.302343313
#> log(c9/c10)  -1.699880  5.435318  0.2194983 -0.7097035 -0.7225595  0.435819580
#>                     s7         s8         s9         s10        s11        s12
#> log(c1/c10) -1.9656470 -1.8199966 -1.8737052 -1.85244126 -1.7096263 -1.7767304
#> log(c2/c10) -0.3667127  2.8535563  1.6660819  2.13621868  5.2937982  3.8101543
#> log(c3/c10)  2.8870191  0.8401745  1.5949486  1.29612357 -0.7108748  0.2321484
#> log(c4/c10) -4.3592199  2.2780643 -0.1694346  0.79956249  7.3076377  4.2497049
#> log(c5/c10)  0.8774484 -0.1907106  0.2031731  0.04722938 -1.0001356 -0.5080129
#> log(c6/c10) -0.5938346  1.8047904  0.9202973  1.27047975  3.6224104  2.5173147
#> log(c7/c10)  1.6091475 -2.3962787 -0.9192776 -1.50404176 -5.4314937 -3.5861118
#> log(c8/c10) -5.1830475  2.3471964 -0.4295814  0.66978155  8.0534331  4.5840955
#> log(c9/c10) -0.1333122  2.1454626  1.3051643  1.63784946  3.8722631  2.8223848
#>                    s13        s14       s15        s16        s17        s18
#> log(c1/c10)  -2.141072 -1.9216279 -1.634626 -1.7000332 -2.0133985 -1.9743307
#> log(c2/c10)  -4.245297  0.6065302  6.952016  5.5058979 -1.4224795 -0.5587071
#> log(c3/c10)   5.352297  2.2684134 -1.764860 -0.8456881  3.5580781  3.0090532
#> log(c4/c10) -12.353355 -2.3532727 10.725384  7.7447956 -6.5352568 -4.7549388
#> log(c5/c10)   2.163970  0.5546250 -1.550164 -1.0704888  1.2276449  0.9411327
#> log(c6/c10)  -3.482807  0.1310876  4.857538  3.7803933 -1.3802250 -0.7368421
#> log(c7/c10)   6.433397  0.3986112 -7.494014 -5.6953070  2.9223285  1.8479535
#> log(c8/c10) -14.252689 -2.9072264 11.930993  8.5494049 -7.6518417 -5.6320052
#> log(c9/c10)  -2.877934  0.5553884  5.045676  4.0223522 -0.8804097 -0.2691742
#>                    s19        s20        s21          s22         s23
#> log(c1/c10) -1.9713330 -1.6965901  -2.132660 -1.929145317 -1.93351466
#> log(c2/c10) -0.4924281  5.5820239  -4.059300  0.440323661  0.34371934
#> log(c3/c10)  2.9669254 -0.8940747   5.234075  2.374056377  2.43545933
#> log(c4/c10) -4.6183313  7.9016985 -11.969997 -2.695840300 -2.89495110
#> log(c5/c10)  0.9191480 -1.0957396   2.102275  0.609755514  0.64179904
#> log(c6/c10) -0.6874740  3.8370959  -3.344267  0.007288324 -0.06466763
#> log(c7/c10)  1.7655144 -5.7899937   6.202051  0.605341750  0.72549988
#> log(c8/c10) -5.4770190  8.7274170 -13.817755 -3.295881937 -3.52178049
#> log(c9/c10) -0.2222728  4.0762216  -2.746316  0.437774898  0.36941430
#>                    s24        s25        s26        s27        s28       s29
#> log(c1/c10) -2.0064430 -1.9585554  -2.094562 -1.8288583 -1.7247977 -2.043484
#> log(c2/c10) -1.2686960 -0.2099209  -3.216976  2.6576277  4.9583662 -2.087665
#> log(c3/c10)  3.4603314  2.7873602   4.698684  0.9647093 -0.4976699  3.980878
#> log(c4/c10) -6.2182943 -4.0360570 -10.233888  1.8742365  6.6162801 -7.906268
#> log(c5/c10)  1.1766351  0.8254408   1.822877 -0.1257213 -0.8888732  1.448286
#> log(c6/c10) -1.2656790 -0.4770479  -2.716860  1.6588525  3.3725630 -1.875690
#> log(c7/c10)  2.7310500  1.4141272   5.154355 -2.1525792 -5.0142776  3.749697
#> log(c8/c10) -7.2922360 -4.8164072 -11.848075  1.8890389  7.2690624 -9.207304
#> log(c9/c10) -0.7715871 -0.0223609  -2.150258  2.0068167  3.6348997 -1.351118
#>                    s30
#> log(c1/c10) -2.0033797
#> log(c2/c10) -1.2009680
#> log(c3/c10)  3.4172826
#> log(c4/c10) -6.0787003
#> log(c5/c10)  1.1541698
#> log(c6/c10) -1.2152317
#> log(c7/c10)  2.6468087
#> log(c8/c10) -7.1338615
#> log(c9/c10) -0.7236604
#>