#' Predict response from new data #' #' #' @param object An object of class pibblefit #' @param ... Other objects to be passed to the `predict` function #' #' @return (if summary==FALSE) array D x N x iter; (if summary==TRUE) #' tibble with calculated posterior summaries #' #' @export #' @importFrom stats median predict runif #' @examples #' sim <- pibble_sim() #' fit <- pibble(sim$Y, sim$X) #' predict(fit)[,,1:2] # just show 2 samples predict <- function(object, ...) UseMethod("predict") Predict response from new data

#' Predict response from new data #' #' #' @param object An object of class pibblefit #' @param ... Other objects to be passed to the `predict` function #' #' @return (if summary==FALSE) array D x N x iter; (if summary==TRUE) #' tibble with calculated posterior summaries #' #' @export #' @importFrom stats median predict runif #' @examples #' sim <- pibble_sim() #' fit <- pibble(sim$Y, sim$X) #' predict(fit)[,,1:2] # just show 2 samples predict <- function(object, ...) UseMethod("predict") Predict response from new data

Usage

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                      s1         s2          s3         s4          s5
#> log(c1/c10)  2.02040135  1.2297566  1.64184915  0.8092051  0.62735352
#> log(c2/c10) -0.07730111 -0.9626076 -0.50117634 -1.4335106 -1.63713478
#> log(c3/c10) -0.04150090  0.1663964  0.05803811  0.2769790  0.32479628
#> log(c4/c10) -0.15292404 -0.4485374 -0.29446056 -0.6057771 -0.67376936
#> log(c5/c10)  2.87525803  1.2495113  2.09686797  0.3847610  0.01083252
#> log(c6/c10)  1.36343884  0.4004278  0.90235952 -0.1118070 -0.33330350
#> log(c7/c10)  2.95379234  2.0261201  2.50963284  1.5326823  1.31931386
#> log(c8/c10)  0.34941901  0.2185128  0.28674253  0.1488825  0.11877358
#> log(c9/c10)  1.84714795  0.4345951  1.17083286 -0.3167552 -0.64164823
#>                     s6         s7          s8          s9         s10
#> log(c1/c10)  1.3942290  0.7694058  1.69271640  1.92946772  2.07146051
#> log(c2/c10) -0.7784434 -1.4780749 -0.44421890 -0.17912198 -0.02012878
#> log(c3/c10)  0.1231490  0.2874441  0.04466274 -0.01759021 -0.05492673
#> log(c4/c10) -0.3870430 -0.6206576 -0.27544185 -0.18692313 -0.13383358
#> log(c5/c10)  1.5877041  0.3029247  2.20146269  2.68827765  2.98024735
#> log(c6/c10)  0.6007563 -0.1602827  0.96431619  1.25268101  1.42562925
#> log(c7/c10)  2.2190973  1.4859854  2.56931595  2.84709890  3.01370061
#> log(c8/c10)  0.2457443  0.1422930  0.29516456  0.33436322  0.35787281
#> log(c9/c10)  0.7284387 -0.3878599  1.26171144  1.68468743  1.93836939
#>                    s11          s12        s13         s14         s15
#> log(c1/c10)  1.3263470  0.621568108  1.2043400  0.88001846 -0.16391754
#> log(c2/c10) -0.8544527 -1.643612863 -0.9910672 -1.35421895 -2.52314257
#> log(c3/c10)  0.1409983  0.326317533  0.1730796  0.25835889  0.53285831
#> log(c4/c10) -0.4124234 -0.675932464 -0.4580404 -0.57930071 -0.96961694
#> log(c5/c10)  1.4481233 -0.001063606  1.1972490  0.53036954 -1.61620206
#> log(c6/c10)  0.5180756 -0.340350172  0.3694702 -0.02555577 -1.29707737
#> log(c7/c10)  2.1394507  1.312525775  1.9962985  1.61576845  0.39090675
#> log(c8/c10)  0.2345052  0.117815695  0.2143046  0.16060704 -0.01223632
#> log(c9/c10)  0.6071620 -0.651984344  0.3891863 -0.19024122 -2.05531998
#>                    s16        s17         s18        s19         s20
#> log(c1/c10)  0.7689462  3.0388708  2.01317991  2.9428850  0.11772164
#> log(c2/c10) -1.4785896  1.0631069 -0.08538715  0.9556290 -2.20778350
#> log(c3/c10)  0.2875650 -0.3093040 -0.03960205 -0.2840648  0.45880225
#> log(c4/c10) -0.6208294  0.2278705 -0.15562406  0.1919825 -0.86431514
#> log(c5/c10)  0.3019796  4.9694645  2.86040910  4.7720958 -1.03708739
#> log(c6/c10) -0.1608426  2.6039419  1.35464307  2.4870306 -0.95403882
#> log(c7/c10)  1.4854461  4.1487738  2.94531934  4.0361527  0.72135712
#> log(c8/c10)  0.1422169  0.5180459  0.34822336  0.5021536  0.03439438
#> log(c9/c10) -0.3886811  3.6667285  1.83424625  3.4952420 -1.55214810
#>                     s21         s22        s23         s24         s25
#> log(c1/c10)  1.02923218  2.24470000  1.2278535  2.37357511  0.06135052
#> log(c2/c10) -1.18714028  0.17385220 -0.9647386  0.31815718 -2.27090378
#> log(c3/c10)  0.21912366 -0.10047947  0.1668969 -0.13436674  0.47362484
#> log(c4/c10) -0.52351134 -0.06906123 -0.4492490 -0.02087624 -0.88539168
#> log(c5/c10)  0.83718712  3.33646744  1.2455981  3.60146419 -1.15299932
#> log(c6/c10)  0.15618761  1.63663621  0.3981098  1.79360703 -1.02269924
#> log(c7/c10)  1.79084255  3.21696443  2.0238872  3.36817502  0.65521627
#> log(c8/c10)  0.18531219  0.38655589  0.2181977  0.40789360  0.02506107
#> log(c9/c10)  0.07634153  2.24787620  0.4311951  2.47812233 -1.65285980
#>                      s26        s27         s28         s29        s30
#> log(c1/c10) -0.058215188 -0.6823088  0.86245123  2.59408703  1.2527811
#> log(c2/c10) -2.404784767 -3.1035994 -1.37388946  0.56507039 -0.9368265
#> log(c3/c10)  0.505064233  0.6691675  0.26297814 -0.19234959  0.1603422
#> log(c4/c10) -0.930095991 -1.1634378 -0.58586890  0.06157075 -0.4399289
#> log(c5/c10) -1.398853806 -2.6821333  0.49424728  4.05488721  1.2968549
#> log(c6/c10) -1.168331141 -1.9284817 -0.04695279  2.06219213  0.4284718
#> log(c7/c10)  0.514928495 -0.2173275  1.59515662  3.62690410  2.0531350
#> log(c8/c10)  0.005264708 -0.0980658  0.15769845  0.44440352  0.2223249
#> log(c9/c10) -1.866473913 -2.9814692 -0.22162655  2.87208526  0.4757303
#> 
#> , , 2
#> 
#>                      s1          s2          s3         s4         s5
#> log(c1/c10)  1.11385559  0.85099551  0.98800101  0.7111777  0.6507188
#> log(c2/c10)  0.42309023 -0.43327891  0.01306990 -0.8887898 -1.0857583
#> log(c3/c10) -0.07765775  0.03969635 -0.02146987  0.1021181  0.1291100
#> log(c4/c10) -0.64782678 -0.72333467 -0.68397915 -0.7634980 -0.7808652
#> log(c5/c10)  2.55842555  0.36288013  1.50722202 -0.8049515 -1.3099360
#> log(c6/c10)  1.11388593  0.02152583  0.59087567 -0.5595110 -0.8107583
#> log(c7/c10)  2.05806581  0.18082275  1.15926203 -0.8177008 -1.2494745
#> log(c8/c10) -0.52066119  0.19060298 -0.18011556  0.5689312  0.7325249
#> log(c9/c10)  0.76040862 -0.42510644  0.19279676 -1.0556932 -1.3283666
#>                      s6         s7          s8          s9         s10
#> log(c1/c10)  0.90567647  0.6979459  1.00491248  1.08362352  1.13083086
#> log(c2/c10) -0.25513436 -0.9318975  0.06816562  0.32459751  0.47839381
#> log(c3/c10)  0.01528399  0.1080254 -0.02902001 -0.06416062 -0.08523637
#> log(c4/c10) -0.70762729 -0.7672989 -0.67912124 -0.65651109 -0.64295054
#> log(c5/c10)  0.81960426 -0.9154702  1.64847554  2.30591149  2.70021197
#> log(c6/c10)  0.24876195 -0.6144979  0.66115420  0.98825141  1.18442960
#> log(c7/c10)  0.57133265 -0.9121969  1.28003712  1.84216034  2.17929654
#> log(c8/c10)  0.04264361  0.6047346 -0.22587575 -0.43885728 -0.56659402
#> log(c9/c10) -0.17849194 -1.1153694  0.26906855  0.62406017  0.83696816
#>                     s11        s12         s13        s14        s15        s16
#> log(c1/c10)  0.88310824  0.6487954  0.84254544  0.7347205  0.3876505  0.6977931
#> log(c2/c10) -0.32865916 -1.0920246 -0.46080829 -0.8120899 -1.9428058 -0.9323953
#> log(c3/c10)  0.02535959  0.1299688  0.04346889  0.0916074  0.2465571  0.1080937
#> log(c4/c10) -0.71411013 -0.7814177 -0.72576200 -0.7567352 -0.8564329 -0.7673428
#> log(c5/c10)  0.63110255 -1.3260016  0.29230073 -0.6083095 -3.5072207 -0.9167466
#> log(c6/c10)  0.15497580 -0.8187515 -0.01358987 -0.4616747 -1.9039837 -0.6151329
#> log(c7/c10)  0.41015925 -1.2632109  0.12047571 -0.6495672 -3.1282046 -0.9132882
#> log(c8/c10)  0.10371022  0.7377295  0.21346773  0.5052275  1.4443525  0.6051481
#> log(c9/c10) -0.28027604 -1.3370414 -0.46321676 -0.9495137 -2.5148208 -1.1160585
#>                    s17         s18        s19        s20         s21        s22
#> log(c1/c10)  1.4524589  1.11145472  1.4205472  0.4812851  0.78432857  1.1884266
#> log(c2/c10)  1.5262225  0.41526849  1.4222577 -1.6377547 -0.65047244  0.6660343
#> log(c3/c10) -0.2288275 -0.07658588 -0.2145805  0.2047539  0.06945986 -0.1109500
#> log(c4/c10) -0.5505613 -0.64851644 -0.5597280 -0.8295359 -0.74248509 -0.6264059
#> log(c5/c10)  5.3866184  2.53837230  5.1200751 -2.7251354 -0.19395716  3.1812816
#> log(c6/c10)  2.5210101  1.10390874  2.3883956 -1.5148691 -0.25552006  1.4237785
#> log(c7/c10)  4.4762373  2.04091981  4.2483366 -2.4595033 -0.29528625  2.5906224
#> log(c8/c10) -1.4368765 -0.51416478 -1.3505277  1.1909898  0.37099479 -0.7224403
#> log(c9/c10)  2.2875304  0.74958059  2.1436067 -2.0925231 -0.72577843  1.0967283
#>                     s23        s24        s25        s26        s27         s28
#> log(c1/c10)  0.85036280  1.2312728  0.4625438  0.4227926  0.2153046  0.72888009
#> log(c2/c10) -0.43534019  0.8056225 -1.6988118 -1.8283167 -2.5042897 -0.83111749
#> log(c3/c10)  0.03997882 -0.1300787  0.2131209  0.2308679  0.3235011  0.09421488
#> log(c4/c10) -0.72351642 -0.6140981 -0.8349194 -0.8463381 -0.9059401 -0.75841295
#> log(c5/c10)  0.35759544  3.5391555 -2.8816726 -3.2136953 -4.9467440 -0.65709201
#> log(c6/c10)  0.01889651  1.6018332 -1.5927519 -1.7579446 -2.6201966 -0.48594574
#> log(c7/c10)  0.17630422  2.8966131 -2.5933463 -2.8772335 -4.3590310 -0.69127746
#> log(c8/c10)  0.19231500 -0.8383764  1.2417013  1.3492626  1.9106974  0.52103096
#> log(c9/c10) -0.42795999  1.2899673 -2.1770475 -2.3563277 -3.2921113 -0.97585449
#>                    s29         s30
#> log(c1/c10)  1.3045848  0.85865029
#> log(c2/c10)  1.0444650 -0.40834046
#> log(c3/c10) -0.1628090  0.03627886
#> log(c4/c10) -0.5930388 -0.72113580
#> log(c5/c10)  4.1514962  0.42681692
#> log(c6/c10)  1.9064939  0.05333660
#> log(c7/c10)  3.4201788  0.23549021
#> log(c8/c10) -1.0367489  0.16989016
#> log(c9/c10)  1.6206091 -0.39058289
#>