#' Predict response from new data #' #' #' @param object An object of class pibblefit #' @param ... Other objects to be passed to the `predict` function #' #' @return (if summary==FALSE) array D x N x iter; (if summary==TRUE) #' tibble with calculated posterior summaries #' #' @export #' @importFrom stats median predict runif #' @examples #' sim <- pibble_sim() #' fit <- pibble(sim$Y, sim$X) #' predict(fit)[,,1:2] # just show 2 samples predict <- function(object, ...) UseMethod("predict") Predict response from new data

#' Predict response from new data #' #' #' @param object An object of class pibblefit #' @param ... Other objects to be passed to the `predict` function #' #' @return (if summary==FALSE) array D x N x iter; (if summary==TRUE) #' tibble with calculated posterior summaries #' #' @export #' @importFrom stats median predict runif #' @examples #' sim <- pibble_sim() #' fit <- pibble(sim$Y, sim$X) #' predict(fit)[,,1:2] # just show 2 samples predict <- function(object, ...) UseMethod("predict") Predict response from new data

Usage

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                      s1          s2         s3          s4         s5
#> log(c1/c10) -2.99927176 -2.21003622 -3.4245500 -3.08104525 -3.6717625
#> log(c2/c10)  0.09985666  0.96615368 -0.3669461  0.01009875 -0.6382965
#> log(c3/c10)  3.63282642  2.79801292  4.0826643  3.71932229  4.3441532
#> log(c4/c10) -1.94485688 -0.76671198 -2.5796983 -2.06692566 -2.9487290
#> log(c5/c10)  1.12956387  0.92188763  1.2414699  1.15108142  1.3065204
#> log(c6/c10)  0.12557356 -0.09031211  0.2419032  0.14794169  0.3095251
#> log(c7/c10) -2.40445011 -1.15225100 -3.0791955 -2.53419172 -3.4714222
#> log(c8/c10) -0.28223895 -0.18028979 -0.3371741 -0.29280201 -0.3691077
#> log(c9/c10)  0.94773015 -0.66986498  1.8193686  1.11533082  2.3260484
#>                      s6         s7         s8         s9        s10         s11
#> log(c1/c10) -2.73569301 -3.3710531 -2.9895132 -2.9745660 -3.3291010 -3.04978506
#> log(c2/c10)  0.38917141 -0.3082257  0.1105680  0.1269747 -0.2621773  0.04441121
#> log(c3/c10)  3.35402613  4.0260780  3.6225043  3.6066939  3.9817031  3.68625684
#> log(c4/c10) -1.55139518 -2.4998399 -1.9302897 -1.9079769 -2.4372151 -2.02026147
#> log(c5/c10)  1.06020683  1.2273929  1.1269961  1.1230629  1.2163538  1.14285574
#> log(c6/c10)  0.05347484  0.2272698  0.1229042  0.1188156  0.2157943  0.13939085
#> log(c7/c10) -1.98625674 -2.9943175 -2.3889672 -2.3652520 -2.9277563 -2.48459438
#> log(c8/c10) -0.24819128 -0.3302636 -0.2809784 -0.2790476 -0.3248445 -0.28876399
#> log(c9/c10)  0.40750651  1.7097228  0.9277293  0.8970939  1.6237389  1.05126081
#>                    s12         s13         s14        s15         s16
#> log(c1/c10) -3.1843502 -2.77947001 -2.37085091 -3.2229193 -2.75426957
#> log(c2/c10) -0.1032930  0.34112000  0.78963694 -0.1456280  0.36878103
#> log(c3/c10)  3.8285931  3.40033122  2.96811457  3.8693895  3.37367547
#> log(c4/c10) -2.2211359 -1.61674404 -1.00677087 -2.2787106 -1.57912565
#> log(c5/c10)  1.1782647  1.07172613  0.96420375  1.1884136  1.06509499
#> log(c6/c10)  0.1761995  0.06544949 -0.04632323  0.1867496  0.05855622
#> log(c7/c10) -2.6980951 -2.05571321 -1.40739918 -2.7592888 -2.01573025
#> log(c8/c10) -0.3061464 -0.25384615 -0.20106296 -0.3111285 -0.25059090
#> log(c9/c10)  1.3270618  0.49723062 -0.34026368  1.4061120  0.44558049
#>                    s17        s18        s19         s20        s21         s22
#> log(c1/c10) -2.9755920 -3.2872614 -2.4327182 -2.84575788 -3.2568607 -2.76246386
#> log(c2/c10)  0.1258486 -0.2162524  0.7217289  0.26835974 -0.1828835  0.35978665
#> log(c3/c10)  3.6077791  3.9374473  3.0335546  3.47044719  3.9052910  3.38234297
#> log(c4/c10) -1.9095085 -2.3747583 -1.0991243 -1.71569640 -2.3293773 -1.59135781
#> log(c5/c10)  1.1233329  1.2053443  0.9804833  1.08916885  1.1973448  1.06725120
#> log(c6/c10)  0.1190962  0.2043496 -0.0294002  0.08358173  0.1960339  0.06079767
#> log(c7/c10) -2.3668798 -2.8613737 -1.5055576 -2.16088537 -2.8131401 -2.02873129
#> log(c8/c10) -0.2791801 -0.3194399 -0.2090546 -0.26240886 -0.3155129 -0.25164939
#> log(c9/c10)  0.8991967  1.5379856 -0.2134623  0.63309239  1.4756773  0.46237528
#>                     s23        s24         s25         s26        s27
#> log(c1/c10) -2.37092419 -3.2732683 -2.40383120 -2.75873371 -3.3603918
#> log(c2/c10)  0.78955651 -0.2008931  0.75343643  0.36388101 -0.2965234
#> log(c3/c10)  2.96819208  3.9226461  3.00299946  3.37839741  4.0148010
#> log(c4/c10) -1.00688025 -2.3538699 -1.05600276 -1.58578957 -2.4839250
#> log(c5/c10)  0.96422303  1.2016622  0.97288205  1.06626966  1.2245875
#> log(c6/c10) -0.04630319  0.2005219 -0.03730188  0.05977733  0.2243535
#> log(c7/c10) -1.40751544 -2.8391723 -1.45972562 -2.02281305 -2.9774023
#> log(c8/c10) -0.20107242 -0.3176323 -0.20532317 -0.25116755 -0.3288865
#> log(c9/c10) -0.34011350  1.5093058 -0.27266821  0.45473008  1.6878717
#>                    s28        s29        s30
#> log(c1/c10) -3.3685663 -3.2422700 -3.2055454
#> log(c2/c10) -0.3054961 -0.1668681 -0.1265576
#> log(c3/c10)  4.0234476  3.8898577  3.8510122
#> log(c4/c10) -2.4961277 -2.3075967 -2.2527753
#> log(c5/c10)  1.2267386  1.1935054  1.1838419
#> log(c6/c10)  0.2265896  0.1920427  0.1819972
#> log(c7/c10) -2.9903720 -2.7899905 -2.7317233
#> log(c8/c10) -0.3299424 -0.3136281 -0.3088843
#> log(c9/c10)  1.7046260  1.4457726  1.3705028
#> 
#> , , 2
#> 
#>                      s1         s2          s3           s4          s5
#> log(c1/c10) -3.42965130 -1.9040546 -4.25171657 -3.587719914 -4.72957962
#> log(c2/c10)  0.50615449 -0.6480486  1.12809495  0.625742639  1.48962630
#> log(c3/c10)  3.65979111  2.6984201  4.17782432  3.759399732  4.47895483
#> log(c4/c10) -2.05450564 -0.9225119 -2.66447861 -2.171792653 -3.01905332
#> log(c5/c10)  1.50869234  1.1399726  1.70737638  1.546895764  1.82287057
#> log(c6/c10) -0.03022021 -0.3858912  0.16143251  0.006631219  0.27283943
#> log(c7/c10) -3.06847203 -1.2165713 -4.06636571 -3.260349337 -4.64643709
#> log(c8/c10) -0.20653315 -0.4956560 -0.05073977 -0.176576838  0.03982227
#> log(c9/c10)  0.72760183 -0.6058981  1.44615609  0.865767106  1.86384864
#>                     s6          s7          s8          s9         s10
#> log(c1/c10) -2.9201521 -4.14830673 -3.41078799 -3.38189493 -4.06721301
#> log(c2/c10)  0.1206886  1.04985935  0.49188329  0.47002400  0.98850722
#> log(c3/c10)  3.3387248  4.11265950  3.64790419  3.62969692  4.06155743
#> log(c4/c10) -1.6764569 -2.58774844 -2.04050905 -2.01907038 -2.52757685
#> log(c5/c10)  1.3855520  1.68238337  1.50413328  1.49715015  1.66278392
#> log(c6/c10) -0.1490026  0.13732399 -0.03461792 -0.04135392  0.11841816
#> log(c7/c10) -2.4499980 -3.94083794 -3.04557412 -3.01050123 -3.84239940
#> log(c8/c10) -0.3030907 -0.07033744 -0.21010802 -0.21558368 -0.08570589
#> log(c9/c10)  0.2822566  1.35576718  0.71111371  0.68585874  1.28488445
#>                      s11         s12        s13        s14         s15
#> log(c1/c10) -3.527293791 -3.78740901 -3.0047733 -2.2149103 -3.86196332
#> log(c2/c10)  0.580026743  0.77681912  0.1847094 -0.4128681  0.83322382
#> log(c3/c10)  3.721321564  3.88523595  3.3920497  2.8943091  3.93221715
#> log(c4/c10) -2.126956429 -2.31996210 -1.7392459 -1.1531670 -2.37528145
#> log(c5/c10)  1.532291442  1.59515840  1.4060040  1.2151029  1.61317735
#> log(c6/c10) -0.007456265  0.05318587 -0.1292744 -0.3134196  0.07056714
#> log(c7/c10) -3.186998900 -3.50274918 -2.5527185 -1.5939147 -3.59324963
#> log(c8/c10) -0.188028472 -0.13873283 -0.2870537 -0.4367443 -0.12460370
#> log(c9/c10)  0.812949587  1.04031219  0.3562227 -0.3341841  1.10547894
#>                    s16         s17         s18        s19         s20
#> log(c1/c10) -2.9560607 -3.38387813 -3.98633680 -2.3345001 -3.13290812
#> log(c2/c10)  0.1478555  0.47152441  0.92731963 -0.3223915  0.28165091
#> log(c3/c10)  3.3613529  3.63094665  4.01059242  2.9696699  3.47279522
#> log(c4/c10) -1.7031011 -2.02054191 -2.46756665 -1.2419027 -1.83432199
#> log(c5/c10)  1.3942307  1.49762947  1.64323704  1.2440065  1.43697280
#> log(c6/c10) -0.1406311 -0.04089157  0.09956303 -0.2855390 -0.09940163
#> log(c7/c10) -2.4935869 -3.01290861 -3.74422487 -1.7390831 -2.70825957
#> log(c8/c10) -0.2962855 -0.21520784 -0.10103311 -0.4140803 -0.26277032
#> log(c9/c10)  0.3136438  0.68759223  1.21419183 -0.2296525  0.46822330
#>                     s21        s22        s23         s24        s25        s26
#> log(c1/c10) -3.92757224 -2.9719003 -2.2150520 -3.95928810 -2.2786614 -2.9646899
#> log(c2/c10)  0.88286080  0.1598391 -0.4127610  0.90685570 -0.3646367  0.1543840
#> log(c3/c10)  3.97356131  3.3713344  2.8943983  3.99354739  2.9344825  3.3667907
#> log(c4/c10) -2.42396331 -1.7148541 -1.1532721 -2.44749649 -1.2004703 -1.7095040
#> log(c5/c10)  1.62903430  1.3980590  1.2151372  1.63669967  1.2305109  1.3963163
#> log(c6/c10)  0.08586291 -0.1369383 -0.3133866  0.09325701 -0.2985570 -0.1386193
#> log(c7/c10) -3.67289139 -2.5128144 -1.5940866 -3.71139083 -1.6713012 -2.5040618
#> log(c8/c10) -0.11216985 -0.2932837 -0.4367175 -0.10615924 -0.4246626 -0.2946501
#> log(c9/c10)  1.16282665  0.3274889 -0.3340603  1.19054899 -0.2784603  0.3211864
#>                     s27         s28         s29         s30
#> log(c1/c10) -4.12769835 -4.14349978 -3.89936825 -3.82837932
#> log(c2/c10)  1.03426791  1.04622261  0.86152283  0.80781555
#> log(c3/c10)  4.09967291  4.10963035  3.95578826  3.91105383
#> log(c4/c10) -2.57245701 -2.58418168 -2.40303592 -2.35036209
#> log(c5/c10)  1.67740255  1.68122158  1.62221770  1.60506047
#> log(c6/c10)  0.13251945  0.13620332  0.07928756  0.06273751
#> log(c7/c10) -3.91582171 -3.93500284 -3.63865495 -3.55248246
#> log(c8/c10) -0.07424303 -0.07124843 -0.11751492 -0.13096836
#> log(c9/c10)  1.33775373  1.35156550  1.13817399  1.07612369
#>