#' Predict response from new data #' #' #' @param object An object of class pibblefit #' @param ... Other objects to be passed to the `predict` function #' #' @return (if summary==FALSE) array D x N x iter; (if summary==TRUE) #' tibble with calculated posterior summaries #' #' @export #' @importFrom stats median predict runif #' @examples #' sim <- pibble_sim() #' fit <- pibble(sim$Y, sim$X) #' predict(fit)[,,1:2] # just show 2 samples predict <- function(object, ...) UseMethod("predict") Predict response from new data

#' Predict response from new data #' #' #' @param object An object of class pibblefit #' @param ... Other objects to be passed to the `predict` function #' #' @return (if summary==FALSE) array D x N x iter; (if summary==TRUE) #' tibble with calculated posterior summaries #' #' @export #' @importFrom stats median predict runif #' @examples #' sim <- pibble_sim() #' fit <- pibble(sim$Y, sim$X) #' predict(fit)[,,1:2] # just show 2 samples predict <- function(object, ...) UseMethod("predict") Predict response from new data

Usage

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1         s2         s3         s4         s5         s6
#> log(c1/c10) -0.9613273 -1.1710972 -1.1446457 -0.6167494 -1.4763081  0.2047930
#> log(c2/c10) -0.3682075 -0.3650041 -0.3654081 -0.3734696 -0.3603432 -0.3860154
#> log(c3/c10) -0.3542165 -0.4370414 -0.4265974 -0.2181643 -0.5575501  0.1062113
#> log(c4/c10)  1.8197043  1.7888409  1.7927327  1.8704020  1.7439353  1.9912755
#> log(c5/c10)  1.9611314  1.9081950  1.9148702  2.0480872  1.8311737  2.2554070
#> log(c6/c10) -1.1551970 -1.2055283 -1.1991816 -1.0725204 -1.2787593 -0.8754030
#> log(c7/c10)  1.4056740  1.3846105  1.3872666  1.4402740  1.3539635  1.5227672
#> log(c8/c10) -0.2326739 -0.5457900 -0.5063068  0.2816655 -1.0013676  1.5079531
#> log(c9/c10)  0.5759203  0.2038511  0.2507682  1.1870987 -0.3375019  2.6442699
#>                      s7         s8         s9         s10        s11        s12
#> log(c1/c10) -1.30400274 -0.5673455 -1.4277063 -1.23714657 -0.9567673 -2.7562892
#> log(c2/c10) -0.36297451 -0.3742240 -0.3610854 -0.36399547 -0.3682772 -0.3407966
#> log(c3/c10) -0.48951750 -0.1986577 -0.5383603 -0.46312019 -0.3524160 -1.0629344
#> log(c4/c10)  1.76928654  1.8776708  1.7510860  1.77912308  1.8203753  1.5556118
#> log(c5/c10)  1.87465575  2.0605545  1.8434386  1.89152720  1.9622821  1.5081649
#> log(c6/c10) -1.23741706 -1.0606667 -1.2670980 -1.22137587 -1.1541029 -1.5858725
#> log(c7/c10)  1.37126514  1.4452348  1.3588437  1.37797833  1.4061319  1.2254374
#> log(c8/c10) -0.74417347  0.3554090 -0.9288215 -0.64437959 -0.2258673 -2.9119506
#> log(c9/c10) -0.03188364  1.2747265 -0.2512969  0.08669927  0.5840085 -2.6078065
#>                     s13          s14        s15        s16         s17
#> log(c1/c10) -1.34186909 -1.280905209 -1.6577733 -2.9711461  0.05662173
#> log(c2/c10) -0.36239625 -0.363327232 -0.3575721 -0.3375155 -0.38375264
#> log(c3/c10) -0.50446855 -0.480397730 -0.6291993 -1.1477679  0.04770781
#> log(c4/c10)  1.76371527  1.772684880  1.7172363  1.5240000  1.96947506
#> log(c5/c10)  1.86510001  1.880484516  1.7853802  1.4539448  2.21801534
#> log(c6/c10) -1.24650255 -1.231875132 -1.3222993 -1.6374243 -0.91095462
#> log(c7/c10)  1.36746288  1.373584418  1.3357421  1.2038630  1.50788890
#> log(c8/c10) -0.80069524 -0.709696583 -1.2722344 -3.2326599  1.28678303
#> log(c9/c10) -0.09904724  0.009084499 -0.6593671 -2.9888985  2.38145828
#>                    s18        s19         s20        s21        s22        s23
#> log(c1/c10) -2.4682788 -1.5364327 -0.86479906  0.4810048 -1.8025686 -1.8367546
#> log(c2/c10) -0.3451948 -0.3594251 -0.36968160 -0.3902334 -0.3553609 -0.3548388
#> log(c3/c10) -0.9492171 -0.5812895 -0.31610350  0.2152701 -0.6863699 -0.6998678
#> log(c4/c10)  1.5979868  1.7350892  1.83390653  2.0319145  1.6959326  1.6909028
#> log(c5/c10)  1.5808456  1.8160010  1.98549074  2.3251103  1.7488405  1.7402135
#> log(c6/c10) -1.5167685 -1.2931853 -1.13203640 -0.8091299 -1.3570408 -1.3652433
#> log(c7/c10)  1.2543572  1.3479263  1.41536670  1.5505023  1.3212029  1.3177702
#> log(c8/c10) -2.4820475 -1.0911135 -0.08858951  1.9202448 -1.4883653 -1.5393935
#> log(c9/c10) -2.0969620 -0.4441451  0.74713267  3.1341873 -0.9161908 -0.9768265
#>                      s24        s25        s26         s27        s28
#> log(c1/c10) -0.043988984  0.4073736  0.3179397 -0.03007566 -1.1791762
#> log(c2/c10) -0.382216215 -0.3891090 -0.3877432 -0.38242869 -0.3648807
#> log(c3/c10)  0.007982945  0.1861977  0.1508859  0.01347645 -0.4402313
#> log(c4/c10)  1.954672213  2.0210811  2.0079227  1.95671928  1.7876523
#> log(c5/c10)  2.192625783  2.3065291  2.2839601  2.19613687  1.9061563
#> log(c6/c10) -0.935094723 -0.8267967 -0.8482551 -0.93175642 -1.2074667
#> log(c7/c10)  1.497786325  1.5431088  1.5341285  1.49918340  1.3837993
#> log(c8/c10)  1.136604943  1.8103381  1.6768432  1.15737288 -0.5578493
#> log(c9/c10)  2.203004891  3.0035875  2.8449584  2.22768298  0.1895214
#>                     s29         s30
#> log(c1/c10)  0.16987630  0.08490533
#> log(c2/c10) -0.38548216 -0.38418456
#> log(c3/c10)  0.09242494  0.05887523
#> log(c4/c10)  1.98613819  1.97363642
#> log(c5/c10)  2.24659563  2.22515283
#> log(c6/c10) -0.88378080 -0.90416837
#> log(c7/c10)  1.51926108  1.51072893
#> log(c8/c10)  1.45583416  1.32900096
#> log(c9/c10)  2.58233810  2.43162494
#> 
#> , , 2
#> 
#>                     s1         s2         s3         s4          s5         s6
#> log(c1/c10) -0.7665346 -0.8676528 -0.8549020 -0.6004330 -1.01477774 -0.2044138
#> log(c2/c10) -0.1616803 -0.1009352 -0.1085951 -0.2614631 -0.01255237 -0.4993651
#> log(c3/c10) -1.6436508 -1.5473503 -1.5594936 -1.8018386 -1.40723510 -2.1789897
#> log(c4/c10)  2.2564963  2.1896874  2.1981119  2.3662396  2.09248197  2.6278896
#> log(c5/c10)  3.2945192  3.4281366  3.4112877  3.0750330  3.62254708  2.5517342
#> log(c6/c10) -0.8719222 -0.9655691 -0.9537604 -0.7180934 -1.10182336 -0.3513348
#> log(c7/c10)  1.1013325  1.1585925  1.1513721  1.0072745  1.24190463  0.7830215
#> log(c8/c10) -0.5804032 -0.5912342 -0.5898685 -0.5626116 -0.60699320 -0.5201928
#> log(c9/c10)  0.5787701  0.3933945  0.4167699  0.8832769  0.12367674  1.6092817
#>                     s7         s8          s9         s10        s11        s12
#> log(c1/c10) -0.9317191 -0.5766182 -0.99134959 -0.89949148 -0.7643365 -1.6317844
#> log(c2/c10) -0.0624485 -0.2757695 -0.02662645 -0.08180869 -0.1630008  0.3581042
#> log(c3/c10) -1.4863365 -1.8245188 -1.42954703 -1.51702861 -1.6457442 -0.8196254
#> log(c4/c10)  2.1473588  2.3819741  2.10796095  2.16865162  2.2579486  1.6848255
#> log(c5/c10)  3.5127936  3.0435641  3.59158917  3.47020810  3.2916146  4.4378582
#> log(c6/c10) -1.0249016 -0.6960381 -1.08012624 -0.99505526 -0.8698864 -1.6732412
#> log(c7/c10)  1.1948712  0.9937889  1.22863802  1.17662171  1.1000877  1.5912957
#> log(c8/c10) -0.5980966 -0.5600607 -0.60448375 -0.59464456 -0.5801677 -0.6730825
#> log(c9/c10)  0.2759446  0.9269356  0.16662656  0.33502603  0.5827998 -1.0074547
#>                     s13         s14         s15        s16        s17
#> log(c1/c10) -0.94997230 -0.92058505 -1.10225189 -1.7353548 -0.2758388
#> log(c2/c10) -0.05148317 -0.06913708  0.03999628  0.4203224 -0.4564577
#> log(c3/c10) -1.46895292 -1.49694003 -1.32392860 -0.7209896 -2.1109677
#> log(c4/c10)  2.13529894  2.15471510  2.03468777  1.6163966  2.5806991
#> log(c5/c10)  3.53691333  3.49808109  3.73813521  4.5747158  2.6461150
#> log(c6/c10) -1.04180620 -1.01459028 -1.18283431 -1.7691591 -0.4174824
#> log(c7/c10)  1.20520738  1.18856632  1.29143844  1.6499443  0.8234672
#> log(c8/c10) -0.60005171 -0.59690396 -0.61636281 -0.6841763 -0.5278434
#> log(c9/c10)  0.24248174  0.29635610 -0.03668582 -1.1973258  1.4783413
#>                    s18          s19        s20         s21         s22
#> log(c1/c10) -1.4929508 -1.043760436 -0.7200038 -0.07126768 -1.17204954
#> log(c2/c10)  0.2747022  0.004858501 -0.1896330 -0.57935046  0.08192606
#> log(c3/c10) -0.9518443 -1.379633265 -1.6879647 -2.30579215 -1.25745642
#> log(c4/c10)  1.7765529  2.073333093  2.2872392  2.71585927  1.98857245
#> log(c5/c10)  4.2544038  3.660844742  3.2330335  2.37579518  3.83036565
#> log(c6/c10) -1.5446657 -1.128664607 -0.8288294 -0.22802644 -1.24747481
#> log(c7/c10)  1.5126787  1.258316603  1.0749836  0.70762507  1.33096262
#> log(c8/c10) -0.6582117 -0.610097622 -0.5754191 -0.50593116 -0.62383903
#> log(c9/c10) -0.7529372  0.070544029  0.6640731  1.85337271 -0.16464281
#>                     s23        s24        s25        s26        s27         s28
#> log(c1/c10) -1.18852866 -0.3243376 -0.1067611 -0.1498722 -0.3176307 -0.87154726
#> log(c2/c10)  0.09182562 -0.4273229 -0.5580283 -0.5321301 -0.4313519 -0.09859572
#> log(c3/c10) -1.24176244 -2.0647796 -2.2719898 -2.2309327 -2.0711669 -1.54364144
#> log(c4/c10)  1.97768470  2.5486559  2.6924087  2.6639253  2.5530872  2.18711438
#> log(c5/c10)  3.85214112  2.7102011  2.4226961  2.4796630  2.7013387  3.43328267
#> log(c6/c10) -1.26273634 -0.4623977 -0.2608974 -0.3008231 -0.4561865 -0.96917575
#> log(c7/c10)  1.34029422  0.8509304  0.7277239  0.7521363  0.8471326  1.16079779
#> log(c8/c10) -0.62560416 -0.5330382 -0.5097330 -0.5143507 -0.5323198 -0.59165138
#> log(c9/c10) -0.19485326  1.3894307  1.7883041  1.7092705  1.4017260  0.38625496
#>                    s29        s30
#> log(c1/c10) -0.2212452 -0.2622049
#> log(c2/c10) -0.4892540 -0.4646481
#> log(c3/c10) -2.1629602 -2.1239520
#> log(c4/c10)  2.6167691  2.5897070
#> log(c5/c10)  2.5739751  2.6280991
#> log(c6/c10) -0.3669226 -0.4048559
#> log(c7/c10)  0.7925525  0.8157467
#> log(c8/c10) -0.5219957 -0.5263830
#> log(c9/c10)  1.5784255  1.5033358
#>