Predict response from new data

Predict response from new data

Usage

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                      s1          s2         s3          s4          s5
#> log(c1/c10) -0.28372866  0.01053469 -0.8154688 -1.83366374 -1.52918855
#> log(c2/c10) -0.98444563 -1.16086659 -0.6656492 -0.05520647 -0.23774978
#> log(c3/c10) -0.59202037 -0.71631876 -0.3674105  0.06268037 -0.06593155
#> log(c4/c10)  3.25612105  2.42449992  4.7588783  7.63641091  6.77592997
#> log(c5/c10)  1.88208790  2.42886118  0.8940569 -0.99785989 -0.43211190
#> log(c6/c10) -0.86459716 -0.88449734 -0.8286371 -0.75977955 -0.78037033
#> log(c7/c10)  1.33880514  1.19003681  1.6076327  2.12239322  1.96846217
#> log(c8/c10)  0.06893603 -0.25171031  0.6483508  1.75783481  1.42606105
#> log(c9/c10)  1.47587659  0.73844137  2.8084378  5.36007324  4.59704675
#>                     s6         s7          s8         s9        s10        s11
#> log(c1/c10) -0.1473011  0.1878314 -0.07331021  0.5873790 -0.8559509  0.1667615
#> log(c2/c10) -1.0662386 -1.2671621 -1.11059870 -1.5067045 -0.6413788 -1.2545299
#> log(c3/c10) -0.6496481 -0.7912098 -0.68090223 -0.9599808 -0.3503107 -0.7823098
#> log(c4/c10)  2.8705616  1.9234395  2.66145497  0.7942733  4.8732851  1.9829854
#> log(c5/c10)  2.1355851  2.7582978  2.27306825  3.5007007  0.8188368  2.7191476
#> log(c6/c10) -0.8738234 -0.8964874 -0.87882715 -0.9235076 -0.8258994 -0.8950625
#> log(c7/c10)  1.2698326  1.1004023  1.23242559  0.8984063  1.6280989  1.1110545
#> log(c8/c10) -0.0797233 -0.4449031 -0.16034807 -0.8802732  0.6924624 -0.4219441
#> log(c9/c10)  1.1339839  0.2941290  0.94855989 -0.7071527  2.9098874  0.3469310
#>                     s12        s13         s14        s15         s16
#> log(c1/c10) -0.21531589 -0.5728201  2.32717013 -0.9209134 -0.26688778
#> log(c2/c10) -1.02546144 -0.8111254 -2.54976887 -0.6024315 -0.99454232
#> log(c3/c10) -0.62091828 -0.4699066 -1.69487776 -0.3228702 -0.59913405
#> log(c4/c10)  3.06277889  4.0731257 -4.12257089  5.0568765  3.20852683
#> log(c5/c10)  2.00920627  1.3449246  6.73342154  0.6981294  1.91338010
#> log(c6/c10) -0.86922372 -0.8450467 -1.04116466 -0.8215062 -0.86573606
#> log(c7/c10)  1.30421825  1.4849587  0.01883427  1.6609415  1.33029103
#> log(c8/c10) -0.00561048  0.3839467 -2.77605007  0.7632494  0.05058523
#> log(c9/c10)  1.30443155  2.2003508 -5.06713573  3.0726860  1.43367269
#>                    s17        s18        s19        s20        s21         s22
#> log(c1/c10) -1.0746071  0.1169662 -2.7702039  0.4740726  1.6108286  0.83544771
#> log(c2/c10) -0.5102869 -1.2246759  0.5062814 -1.4387735 -2.1202976 -1.65543020
#> log(c3/c10) -0.2579491 -0.7612760  0.4582798 -0.9121196 -1.3922913 -1.06476636
#> log(c4/c10)  5.4912320  2.1237125 10.2831780  1.1144898 -2.0981095  0.09320347
#> log(c5/c10)  0.4125498  2.6266225 -2.7380532  3.2901650  5.4023811  3.96163921
#> log(c6/c10) -0.8111123 -0.8916950 -0.6964441 -0.9158451 -0.9927206 -0.94028381
#> log(c7/c10)  1.7386431  1.1362291  2.5958722  0.9556897  0.3809892  0.77299224
#> log(c8/c10)  0.9307229 -0.3676842  2.7783430 -0.7568080 -1.9954829 -1.15058318
#> log(c9/c10)  3.4578483  0.4717199  7.7070786 -0.4232025 -3.2719563 -1.32882225
#>                    s23         s24        s25        s26         s27        s28
#> log(c1/c10) -2.2023443  0.33640502 -2.1336159 -0.5854155 -2.01509655  0.4475976
#> log(c2/c10)  0.1658301 -1.35623701  0.1246251 -0.8035741  0.05356871 -1.4229007
#> log(c3/c10)  0.2184130 -0.85396810  0.1893817 -0.4645863  0.13931854 -0.9009364
#> log(c4/c10)  8.6783434  1.50355392  8.4841093  4.1087217  8.14916032  1.1893114
#> log(c5/c10) -1.6829084  3.03436366 -1.5552036  1.3215210 -1.33498176  3.2409715
#> log(c6/c10) -0.7348468 -0.90653500 -0.7394947 -0.8441950 -0.74750978 -0.9140546
#> log(c7/c10)  2.3087840  1.02528920  2.2740376  1.4913265  2.21411873  0.9690745
#> log(c8/c10)  2.1595704 -0.60679744  2.0846800  0.3976714  1.95553448 -0.7279592
#> log(c9/c10)  6.2840008 -0.07820215  6.1117649  2.2319153  5.81475081 -0.3568549
#>                     s29        s30
#> log(c1/c10) -1.34263220 -1.3082029
#> log(c2/c10) -0.34959670 -0.3702383
#> log(c3/c10) -0.14473394 -0.1592771
#> log(c4/c10)  6.24870087  6.1513997
#> log(c5/c10) -0.08546993 -0.0214965
#> log(c6/c10) -0.79298659 -0.7953149
#> log(c7/c10)  1.87414639  1.8567402
#> log(c8/c10)  1.22277848  1.1852623
#> log(c9/c10)  4.12952940  4.0432482
#> 
#> , , 2
#> 
#>                      s1         s2         s3         s4         s5         s6
#> log(c1/c10)  0.02481566  0.2520972 -0.3858871 -1.1723151 -0.9371462  0.1301889
#> log(c2/c10) -0.71875454 -1.1783942  0.1118242  1.7022458  1.2266552 -0.9318545
#> log(c3/c10) -1.68322185 -1.7877833 -1.4942771 -1.1324789 -1.2406690 -1.7316991
#> log(c4/c10)  3.35617498  2.4505957  4.9925761  8.1260148  7.1890092  2.9363267
#> log(c5/c10)  1.93002351  2.3382280  1.1923892 -0.2200590  0.2023115  2.1192769
#> log(c6/c10) -0.63426980 -0.8229939 -0.2932414  0.3597719  0.1644985 -0.7217668
#> log(c7/c10)  1.25687248  1.0607582  1.6112552  2.2898399  2.0869198  1.1659492
#> log(c8/c10)  0.26082937  0.0598926  0.6239264  1.3191975  1.1112876  0.1676703
#> log(c9/c10)  0.77777403 -0.2511829  2.6371213  6.1974651  5.1328002  0.3007249
#>                     s7          s8         s9        s10         s11
#> log(c1/c10)  0.3890368  0.18733756  0.6976373 -0.4171544  0.37276290
#> log(c2/c10) -1.4553319 -1.04742839 -2.0794257  0.1750572 -1.42242068
#> log(c3/c10) -1.8507827 -1.75799046 -1.9927551 -1.4798925 -1.84329584
#> log(c4/c10)  1.9049748  2.70862379  0.6753891  5.1171574  1.96981631
#> log(c5/c10)  2.5841755  2.22191770  3.1384312  1.1362322  2.55494713
#> log(c6/c10) -0.9367021 -0.76922043 -1.1929496 -0.2672784 -0.92318901
#> log(c7/c10)  0.9425973  1.11663732  0.6763154  1.6382348  0.95663948
#> log(c8/c10) -0.0611739  0.11714582 -0.3340037  0.6515695 -0.04678637
#> log(c9/c10) -0.8711402  0.04199933 -2.2682469  2.7786757 -0.79746457
#>                     s12        s13        s14        s15         s16        s17
#> log(c1/c10)  0.07765596 -0.1984713  2.0414080 -0.4673298  0.03782313 -0.5860389
#> log(c2/c10) -0.82561537 -0.2671934 -4.7969813  0.2765288 -0.74506001  0.5165985
#> log(c3/c10) -1.70753116 -1.5804982 -2.6109601 -1.4568091 -1.68920597 -1.4021967
#> log(c4/c10)  3.14563843  4.2458379 -4.6787221  5.3170760  3.30434809  5.7900599
#> log(c5/c10)  2.02492627  1.5289935  5.5518835  1.0461156  1.95338533  0.8329105
#> log(c6/c10) -0.67814593 -0.4488627 -2.3087543 -0.2256151 -0.64507060 -0.1270446
#> log(c7/c10)  1.21127821  1.4495399 -0.4831832  1.6815297  1.24564873  1.7839601
#> log(c8/c10)  0.21411393  0.4582345 -1.5220143  0.6959289  0.24932963  0.8008781
#> log(c9/c10)  0.53855361  1.7886462 -8.3518115  3.0058316  0.71888615  3.5432556
#>                     s18        s19        s20          s21         s22
#> log(c1/c10)  0.33430227 -1.8956751  0.6101223  1.488123878  0.88923926
#> log(c2/c10) -1.34464037  3.1651225 -1.9024410 -3.678054993 -2.46690925
#> log(c3/c10) -1.82560193 -0.7996954 -1.9524936 -2.356420407 -2.08090206
#> log(c4/c10)  2.12305858 11.0081654  1.0240833 -2.474220466 -0.08802862
#> log(c5/c10)  2.48587069 -1.5192350  2.9812516  4.558168645  3.48255404
#> log(c6/c10) -0.89125309  0.9604164 -1.1202812 -1.849332883 -1.35204695
#> log(c7/c10)  0.98982598  2.9140050  0.7518294 -0.005771262  0.51098789
#> log(c8/c10) -0.01278382  1.9587108 -0.2566328 -1.032862845 -0.50339656
#> log(c9/c10) -0.62334430  9.4722850 -1.8720459 -5.846964663 -3.13567388
#>                    s23        s24        s25        s26        s27        s28
#> log(c1/c10) -1.4570747  0.5037913 -1.4039906 -0.2081996 -1.3124492  0.5896736
#> log(c2/c10)  2.2781252 -1.6874040  2.1707714 -0.2475194  1.9856440 -1.8610869
#> log(c3/c10) -1.0014746 -1.9035758 -1.0258960 -1.5760226 -1.0680098 -1.9430861
#> log(c4/c10)  9.2606088  1.4477478  9.0491011  4.2845995  8.6843643  1.1055589
#> log(c5/c10) -0.7314956  2.7902780 -0.6361551  1.5115211 -0.4717441  2.9445252
#> log(c6/c10)  0.5962230 -1.0319889  0.5521445 -0.4407847  0.4761328 -1.1033015
#> log(c7/c10)  2.5355502  0.8435791  2.4897455  1.4579342  2.4107573  0.7694739
#> log(c8/c10)  1.5709498 -0.1626269  1.5240188  0.4668352  1.4430883 -0.2385543
#> log(c9/c10)  7.4866383 -1.3906606  7.2463143  1.8326887  6.8318851 -1.7794700
#>                     s29         s30
#> log(c1/c10) -0.79305473 -0.76646239
#> log(c2/c10)  0.93525395  0.88147530
#> log(c3/c10) -1.30695856 -1.31919244
#> log(c4/c10)  6.61489227  6.50893791
#> log(c5/c10)  0.46110394  0.50886459
#> log(c6/c10)  0.04485174  0.02277069
#> log(c7/c10)  1.96258773  1.93964201
#> log(c8/c10)  0.98389819  0.96038824
#> log(c9/c10)  4.48046464  4.36007487
#>