Predict response from new data

Predict response from new data

Usage

# S3 method for class 'pibblefit'
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1         s2         s3          s4          s5         s6
#> log(c1/c10)  0.6863508  0.9219946  1.1473958  1.73329703  1.82454664  1.3992294
#> log(c2/c10) -1.6295976 -1.4148640 -1.2094641 -0.67555370 -0.59240125 -0.9799774
#> log(c3/c10)  1.3595185  0.7966756  0.2582972 -1.14114704 -1.35909975 -0.3432154
#> log(c4/c10) -1.1939518 -1.3017092 -1.4047828 -1.67270915 -1.71443663 -1.5199436
#> log(c5/c10) -4.6737184 -4.3095702 -3.9612500 -3.05583713 -2.91482602 -3.5720831
#> log(c6/c10) -7.0836708 -6.2227212 -5.3991936 -3.25854127 -2.92515108 -4.4790930
#> log(c7/c10)  1.0213983  0.8170428  0.6215698  0.11346333  0.03432964  0.4031742
#> log(c8/c10)  2.6929886  2.3794906  2.0796191  1.30014233  1.17874479  1.7445824
#> log(c9/c10) -2.2387673 -1.7504600 -1.2833774 -0.06925722  0.11983269 -0.7615212
#>                     s7         s8         s9         s10         s11        s12
#> log(c1/c10)  1.4698227  3.2304568  0.8054008  2.36617646  3.75358246  2.1903308
#> log(c2/c10) -0.9156482  0.6887534 -1.5211116 -0.09883376  1.16545871 -0.2590755
#> log(c3/c10) -0.5118299 -4.7171624  1.0751636 -2.65280037 -5.96666524 -2.2327872
#> log(c4/c10) -1.5522251 -2.3573444 -1.2483921 -1.96211818 -2.59656419 -1.8817058
#> log(c5/c10) -3.4629929 -0.7422257 -4.4897464 -2.07782722  0.06617803 -2.3495674
#> log(c6/c10) -4.2211728  2.2114913 -6.6487088 -0.94624890  4.12278659 -1.5887197
#> log(c7/c10)  0.3419541 -1.1849068  0.9181555 -0.43538376 -1.63857303 -0.2828865
#> log(c8/c10)  1.6506658 -0.6916634  2.5346057  0.45816615 -1.38762428  0.6921094
#> log(c9/c10) -0.6152358  3.0331974 -1.9920686  1.24221254  4.11723251  0.8778204
#>                     s13         s14        s15        s16        s17        s18
#> log(c1/c10)  2.69217837  2.67001913  0.6794494  3.4412223  2.2937238  1.3981991
#> log(c2/c10)  0.19823989  0.17804698 -1.6358866  0.8808163 -0.1648572 -0.9809162
#> log(c3/c10) -3.43146666 -3.37853858  1.3760027 -5.2205827 -2.4797448 -0.3407545
#> log(c4/c10) -2.11119539 -2.10106220 -1.1907958 -2.4537252 -1.9289864 -1.5194724
#> log(c5/c10) -1.57404548 -1.60828890 -4.6843834 -0.4165227 -2.1897907 -3.5736753
#> log(c6/c10)  0.24483377  0.16387263 -7.1088858  2.9815454 -1.2109624 -4.4828573
#> log(c7/c10) -0.71809985 -0.69888286  1.0273834 -1.3676873 -0.3725512  0.4040677
#> log(c8/c10)  0.02445662  0.05393704  2.7021702 -0.9720636  0.5545564  1.7459532
#> log(c9/c10)  1.91776245  1.87184348 -2.2530686  3.4699513  1.0920742 -0.7636562
#>                      s19         s20        s21        s22        s23
#> log(c1/c10)  1.253141583  1.80295934  0.6768125  0.7952767  2.8858584
#> log(c2/c10) -1.113101833 -0.61207297 -1.6382895 -1.5303374  0.3747335
#> log(c3/c10)  0.005720035 -1.30753777  1.3823010  1.0993453 -3.8940778
#> log(c4/c10) -1.453139168 -1.70456498 -1.1895900 -1.2437624 -2.1997632
#> log(c5/c10) -3.797837524 -2.94818561 -4.6884583 -4.5053915 -1.2747452
#> log(c6/c10) -5.012840292 -3.00402259 -7.1185199 -6.6856983  0.9524645
#> log(c7/c10)  0.529864787  0.05305062  1.0296701  0.9269354 -0.8860635
#> log(c8/c10)  1.938936111  1.20746431  2.7056783  2.5480747 -0.2332133
#> log(c9/c10) -1.064248225  0.07509890 -2.2585328 -2.0130481  2.3191115
#>                     s24        s25        s26         s27         s28
#> log(c1/c10) -0.03353579  1.0111212  2.0297497  1.90220941  1.80876670
#> log(c2/c10) -2.28560398 -1.3336462 -0.4054072 -0.52163001 -0.60678093
#> log(c3/c10)  3.07899127  0.5837937 -1.8492340 -1.54459981 -1.32140884
#> log(c4/c10) -0.86475529 -1.3424658 -1.8082738 -1.74995098 -1.70722063
#> log(c5/c10) -5.78618344 -4.1718398 -2.5977188 -2.79481114 -2.93921130
#> log(c6/c10) -9.71385303 -5.8970876 -2.1754199 -2.64140191 -2.98280477
#> log(c7/c10)  1.64569996  0.7397503 -0.1436270 -0.03302123  0.04801435
#> log(c8/c10)  3.65071823  2.2609175  0.9057448  1.07542307  1.19973826
#> log(c9/c10) -3.73053564 -1.5657694  0.5450600  0.28076755  0.08713307
#>                     s29        s30
#> log(c1/c10)  1.91983395  1.6292348
#> log(c2/c10) -0.50556940 -0.7703818
#> log(c3/c10) -1.58669662 -0.8925910
#> log(c4/c10) -1.75801050 -1.6251226
#> log(c5/c10) -2.76757533 -3.2166480
#> log(c6/c10) -2.57700875 -3.6387436
#> log(c7/c10) -0.04830563  0.2037084
#> log(c8/c10)  1.05197555  1.4385856
#> log(c9/c10)  0.31728961 -0.2848977
#> 
#> , , 2
#> 
#>                     s1         s2          s3         s4         s5         s6
#> log(c1/c10)  0.2554892  0.6278036  0.98393505  1.9096522  2.0538256  1.3818291
#> log(c2/c10) -2.1370698 -1.8312776 -1.53877705 -0.7784599 -0.6600464 -1.2119756
#> log(c3/c10)  2.4393810  1.7523679  1.09521648 -0.6129628 -0.8789986  0.3610025
#> log(c4/c10)  0.8846957  0.4577851  0.04943046 -1.0120340 -1.1773489 -0.4068109
#> log(c5/c10) -3.5373904 -3.2557568 -2.98636478 -2.2861153 -2.1770568 -2.6853818
#> log(c6/c10) -5.9087214 -5.2678389 -4.65481308 -3.0613325 -2.8131601 -3.9698992
#> log(c7/c10)  2.1163553  1.7799539  1.45817460  0.6217512  0.4914847  1.0986610
#> log(c8/c10)  1.6198074  1.4238384  1.23638750  0.7491332  0.6732471  1.0269546
#> log(c9/c10) -0.5649927 -0.3421060 -0.12890733  0.4252748  0.5115845  0.1092927
#>                     s7         s8         s9         s10        s11         s12
#> log(c1/c10)  1.4933658  4.2751475  0.4435868  2.90959448  5.1016801  2.63176036
#> log(c2/c10) -1.1203675  1.1643868 -1.9825800  0.04282032  1.8432409 -0.18537254
#> log(c3/c10)  0.1551896 -4.9778922  2.0922938 -2.45810594 -6.5030512 -1.94543261
#> log(c4/c10) -0.5347032 -3.7244058  0.6690154 -2.15860788 -4.6721412 -1.84003213
#> log(c5/c10) -2.6010110 -0.4967600 -3.3951058 -1.52971895  0.1284623 -1.73988380
#> log(c6/c10) -3.7779059  1.0105062 -5.5849399 -1.34008468  2.4332557 -1.81833366
#> log(c7/c10)  0.9978831 -1.5155707  1.9464013 -0.28173759 -2.2623767 -0.03070308
#> log(c8/c10)  0.9682470 -0.4959529  1.5208015  0.22281038 -0.9310010  0.36904928
#> log(c9/c10)  0.1760642  1.8413827 -0.4523877  1.02389199  2.3361878  0.85756612
#>                     s13         s14        s15        s16         s17
#> log(c1/c10)  3.42467377  3.38966242  0.2445850  4.6081546  2.79512013
#> log(c2/c10)  0.46586918  0.43711340 -2.1460257  1.4378947 -0.05120064
#> log(c3/c10) -3.40855583 -3.34395115  2.4595019 -5.5923734 -2.24687218
#> log(c4/c10) -2.74921846 -2.70907304  0.8971989 -4.1062450 -2.02734700
#> log(c5/c10) -1.14009237 -1.16657636 -3.5456387 -0.2448602 -1.61631193
#> log(c6/c10) -0.45345441 -0.51372109 -5.9274911  1.5837269 -1.53713478
#> log(c7/c10) -0.74713283 -0.71549865  2.1262076 -1.8164562 -0.17830533
#> log(c8/c10) -0.04830328 -0.02987494  1.6255468 -0.6712322  0.28306433
#> log(c9/c10)  1.33224509  1.31128549 -0.5715205  2.0407379  0.95536172
#>                    s18         s19        s20        s21        s22         s23
#> log(c1/c10)  1.3802013  1.15101211  2.0197179  0.2404188  0.4275909  3.73068604
#> log(c2/c10) -1.2133126 -1.40155203 -0.6880600 -2.1494475 -1.9957179  0.71720553
#> log(c3/c10)  0.3640064  0.78691759 -0.8160614  2.4671896  2.1218104 -3.97322487
#> log(c4/c10) -0.4049443 -0.14214680 -1.1382397  0.9019760  0.6873570 -3.10010440
#> log(c5/c10) -2.6866132 -2.85998100 -2.2028572 -3.5487902 -3.4072058 -0.90861245
#> log(c6/c10) -3.9727013 -4.36721545 -2.8718712 -5.9346627 -5.6124746  0.07329895
#> log(c7/c10)  1.1001318  1.30721363  0.5223024  2.1299720  1.9608544 -1.02362746
#> log(c8/c10)  1.0278115  1.14844594  0.6911998  1.6277397  1.5292211 -0.20937383
#> log(c9/c10)  0.1083181 -0.02888636  0.4911658 -0.5740146 -0.4619637  1.51543986
#>                    s24        s25        s26        s27        s28        s29
#> log(c1/c10) -0.8819234  0.7686226  2.3780441  2.1765318  2.0288934  2.2043784
#> log(c2/c10) -3.0712582 -1.7156191 -0.3937568 -0.5592643 -0.6805238 -0.5363931
#> log(c3/c10)  4.5381912  1.4925217 -1.4772627 -1.1054224 -0.8329926 -1.1568063
#> log(c4/c10)  2.1888986  0.2963163 -1.5491108 -1.3180489 -1.1487607 -1.3499789
#> log(c5/c10) -4.3977748 -3.1492357 -1.9318050 -2.0842369 -2.1959164 -2.0631726
#> log(c6/c10) -7.8666033 -5.0254405 -2.2550675 -2.6019400 -2.8560769 -2.5540063
#> log(c7/c10)  3.1440542  1.6527182  0.1985400  0.3806145  0.5140119  0.3554540
#> log(c8/c10)  2.2184882  1.3497179  0.5025937  0.6086603  0.6863702  0.5940031
#> log(c9/c10) -1.2459067 -0.2578045  0.7056784  0.5850428  0.4966588  0.6017132
#>                    s30
#> log(c1/c10)  1.7452351
#> log(c2/c10) -0.9135002
#> log(c3/c10) -0.3095722
#> log(c4/c10) -0.8235067
#> log(c5/c10) -2.4104870
#> log(c6/c10) -3.3443514
#> log(c7/c10)  0.7703088
#> log(c8/c10)  0.8356747
#> log(c9/c10)  0.3268463
#>