Predict response from new data

Predict response from new data

Usage

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                      s1          s2         s3         s4          s5
#> log(c1/c10) -0.70404170  -1.0723053 -0.5401210 -0.9485413 -0.63356566
#> log(c2/c10) -0.18980937  -7.7979140  3.1966933 -5.2410247  1.26618322
#> log(c3/c10) -2.63165792 -10.9043126  1.0506478 -8.1240850 -1.04848781
#> log(c4/c10) -0.18526606  -8.3487948  3.4484658 -5.6052416  1.37702019
#> log(c5/c10)  2.19245163   6.1463273  0.4325113  4.8175308  1.43578307
#> log(c6/c10) -0.01765007  -0.5256616  0.2084749 -0.3549319  0.07957007
#> log(c7/c10)  3.96296538   0.9533097  5.3026167  1.9647779  4.53893491
#> log(c8/c10) -1.06290426   4.3362744 -3.4661747  2.5217486 -2.09616608
#> log(c9/c10) -4.14987279  -4.2936457 -4.0858769 -4.2453273 -4.12235841
#>                      s6         s7         s8           s9        s10
#> log(c1/c10) -0.72281157 -0.8824675 -0.7842723 -0.777165454 -0.5425485
#> log(c2/c10) -0.57758358 -3.8759788 -1.8473252 -1.700502257  3.1465432
#> log(c3/c10) -3.05330330 -6.6398056 -4.4339538 -4.274306273  0.9961171
#> log(c4/c10) -0.60134944 -4.1405416 -1.9637876 -1.806245966  3.3946544
#> log(c5/c10)  2.39397502  4.1081266  3.0538503  2.977547539  0.4585739
#> log(c6/c10) -0.04354269 -0.2637845 -0.1283264 -0.118522677  0.2051262
#> log(c7/c10)  3.80956753  2.5047703  3.3072762  3.365357208  5.2827780
#> log(c8/c10) -0.78771584  1.5530281  0.1133708  0.009176261 -3.4305850
#> log(c9/c10) -4.15720069 -4.2195316 -4.1811954 -4.178420864 -4.0868246
#>                     s11         s12        s13          s14        s15
#> log(c1/c10) -0.70474316 -0.76041767 -0.4192360 -0.776121738 -0.8240135
#> log(c2/c10) -0.20430126 -1.35450317  5.6941047 -1.678939717 -2.6683549
#> log(c3/c10) -2.64741564 -3.89808498  3.7662021 -4.250860296 -5.3266986
#> log(c4/c10) -0.20081592 -1.43498747  6.1281988 -1.783109269 -2.8447560
#> log(c5/c10)  2.19998296  2.79773438 -0.8653749  2.966341646  3.4805334
#> log(c6/c10) -0.01861773 -0.09541948  0.3752331 -0.117082895 -0.1831485
#> log(c7/c10)  3.95723260  3.50222942  6.2905563  3.373887036  2.9824887
#> log(c8/c10) -1.05261992 -0.23636596 -5.2384913 -0.006125842  0.6960240
#> log(c9/c10) -4.15014665 -4.17188240 -4.0386825 -4.178013390 -4.1967107
#>                    s16        s17        s18        s19        s20
#> log(c1/c10) -0.9556897 -0.8554379 -0.7868995 -0.3634543 -0.8001980
#> log(c2/c10) -5.3887062 -3.3175632 -1.9016024  6.8465218 -2.1763403
#> log(c3/c10) -8.2846662 -6.0326137 -4.4929720  5.0192802 -4.7917077
#> log(c4/c10) -5.7637045 -3.5413592 -2.0220273  7.3647473 -2.3168223
#> log(c5/c10)  4.8942798  3.8179221  3.0820578 -1.4642776  3.2248371
#> log(c6/c10) -0.3647929 -0.2264977 -0.1319506  0.4521827 -0.1502955
#> log(c7/c10)  1.9063573  2.7256714  3.2858049  6.7464358  3.1771226
#> log(c8/c10)  2.6265524  1.1567421  0.1518893 -6.0563173  0.3468602
#> log(c9/c10) -4.2481181 -4.2089790 -4.1822211 -4.0169049 -4.1874129
#>                    s21        s22        s23        s24         s25
#> log(c1/c10) -0.4876735 -0.8731613 -0.6065824 -0.4267227 -0.65880368
#> log(c2/c10)  4.2802265 -3.6837192  1.8236416  5.5394347  0.74478081
#> log(c3/c10)  2.2288251 -6.4307526 -0.4423368  3.5980221 -1.61543349
#> log(c4/c10)  4.6111015 -3.9342462  1.9751754  5.9622373  0.81755318
#> log(c5/c10) -0.1305929  4.0082107  1.1460761 -0.7849941  1.70675204
#> log(c6/c10)  0.2808250 -0.2509469  0.1167929  0.3649054  0.04475477
#> log(c7/c10)  5.7312467  2.5808254  4.7594573  6.2293711  4.33267568
#> log(c8/c10) -4.2351164  1.4165889 -2.4917728 -5.1287280 -1.72614688
#> log(c9/c10) -4.0654010 -4.2158984 -4.1118239 -4.0416053 -4.13221152
#>                    s26        s27        s28        s29        s30
#> log(c1/c10) -0.7858590 -0.9922782 -0.7340740 -0.6745065 -0.9711141
#> log(c2/c10) -1.8801062 -6.1446032 -0.8102577  0.4203703 -5.7073648
#> log(c3/c10) -4.4695982 -9.1065889 -3.3063010 -1.9681804 -8.6311589
#> log(c4/c10) -1.9989618 -6.5747852 -0.8510098  0.4694594 -6.1056266
#> log(c5/c10)  3.0708864  5.2871139  2.5148941  1.8753458  5.0598843
#> log(c6/c10) -0.1305153 -0.4152660 -0.0590789  0.0230931 -0.3860706
#> log(c7/c10)  3.2943085  1.6073354  3.7175250  4.2043436  1.7803005
#> log(c8/c10)  0.1366343  3.1629834 -0.6225960 -1.4959253  2.8526922
#> log(c9/c10) -4.1818149 -4.2624025 -4.1615976 -4.1383420 -4.2541399
#> 
#> , , 2
#> 
#>                      s1          s2          s3         s4         s5
#> log(c1/c10) -0.78034714  -0.6578935 -0.83485344 -0.6990470 -0.8037816
#> log(c2/c10)  0.09605189  -7.0955013  3.29714012 -4.6786043  1.4723274
#> log(c3/c10) -3.23050215 -10.4350195 -0.02364334 -8.0137656 -1.8517456
#> log(c4/c10)  0.49931591  -7.0058217  3.83998616 -4.4835370  1.9356033
#> log(c5/c10)  1.35801089   5.7621210 -0.60233686  4.2820123  0.5151792
#> log(c6/c10) -0.30332304   0.9731533 -0.87150535  0.5441622 -0.5476073
#> log(c7/c10)  3.99882759   3.0602134  4.41662145  3.3756576  4.1784538
#> log(c8/c10) -2.04712903   3.5101742 -4.52078356  1.6425068 -3.1106518
#> log(c9/c10) -4.68344683  -2.8566426 -5.49658988 -3.4705848 -5.0330495
#>                     s6         s7          s8          s9         s10
#> log(c1/c10) -0.7741059 -0.7210177 -0.75366916 -0.75603230 -0.83404626
#> log(c2/c10) -0.2704913 -3.3882960 -1.47071326 -1.33192903  3.24973572
#> log(c3/c10) -3.5977061 -6.7211313 -4.80009170 -4.66105728 -0.07113319
#> log(c4/c10)  0.1167898 -3.1369654 -1.13576733 -0.99093148  3.79051472
#> log(c5/c10)  1.5824821  3.4918270  2.31749851  2.23250699 -0.57330639
#> log(c6/c10) -0.2382629  0.3151369 -0.02522754 -0.04986127 -0.86309123
#> log(c7/c10)  3.9509878  3.5440638  3.79433936  3.81245295  4.41043441
#> log(c8/c10) -1.7638812  0.6454156 -0.83640460 -0.94365071 -4.48415161
#> log(c9/c10) -4.5903372 -3.7983499 -4.28545587 -4.32070995 -5.48454818
#>                     s11        s12        s13        s14        s15
#> log(c1/c10) -0.78011389 -0.7616012 -0.8750497 -0.7563793 -0.7404546
#> log(c2/c10)  0.08235345 -1.0048737  5.6578157 -1.3115471 -2.2467908
#> log(c3/c10) -3.24422529 -4.3334124  2.3412878 -4.6406386 -5.5775682
#> log(c4/c10)  0.48502016 -0.6496151  6.3035979 -0.9696608 -1.9456854
#> log(c5/c10)  1.36639982  2.0322182 -2.0480155  2.2200251  2.7927687
#> log(c6/c10) -0.30089161 -0.1079125 -1.2905173 -0.0534790  0.1125236
#> log(c7/c10)  3.99703972  3.8551390  4.7247278  3.8151131  3.6930487
#> log(c8/c10) -2.03654350 -1.1963841 -6.3450056 -0.9594010 -0.2366874
#> log(c9/c10) -4.67996714 -4.4037888 -6.0962506 -4.3258874 -4.0883160
#>                    s16        s17         s18       s19         s20
#> log(c1/c10) -0.6966701 -0.7300054 -0.75279556 -0.893598 -0.74837361
#> log(c2/c10) -4.8182001 -2.8604542 -1.52201871  6.747137 -1.78171450
#> log(c3/c10) -8.1536130 -6.1923379 -4.85148964  3.432573 -5.11165358
#> log(c4/c10) -4.6292198 -2.5861073 -1.18930994  7.440418 -1.46032964
#> log(c5/c10)  4.3675008  3.1685764  2.34891799 -2.715116  2.50795579
#> log(c6/c10)  0.5689400  0.2214468 -0.01612099 -1.483868  0.02997413
#> log(c7/c10)  3.3574381  3.6129557  3.78764317  4.866902  3.75374867
#> log(c8/c10)  1.7503800  0.2375236 -0.79675809 -7.186783 -0.59607707
#> log(c9/c10) -3.4351245 -3.9324327 -4.27242321 -6.372961 -4.20645508
#>                    s21        s22        s23        s24        s25
#> log(c1/c10) -0.8522931 -0.7241121 -0.8127540 -0.8725602 -0.7953895
#> log(c2/c10)  4.3213488 -3.2065628  1.9992643  5.5116140  0.9794723
#> log(c3/c10)  1.0024117 -6.5390705 -1.3238588  2.1948226 -2.3454892
#> log(c4/c10)  4.9088550 -2.9473079  2.4855171  6.1510212  1.4212575
#> log(c5/c10) -1.2295627  3.3805335  0.1924828 -1.9584816  0.8170039
#> log(c6/c10) -1.0532989  0.2828799 -0.6411368 -1.2645670 -0.4601272
#> log(c7/c10)  4.5502973  3.5677829  4.2472276  4.7056462  4.1141283
#> log(c8/c10) -5.3122451  0.5049805 -3.5178445 -6.2320277 -2.7297959
#> log(c9/c10) -5.7567602 -3.8445139 -5.1669024 -6.0591124 -4.9078540
#>                     s26        s27        s28        s29        s30
#> log(c1/c10) -0.75314154 -0.6845038 -0.7703609 -0.7901681 -0.6915412
#> log(c2/c10) -1.50169947 -5.5327109 -0.4904263  0.6728236 -5.1194118
#> log(c3/c10) -4.83113377 -8.8694119 -3.8180376 -2.6526907 -8.4553677
#> log(c4/c10) -1.16810468 -5.3748866 -0.1127354  1.1012375 -4.9435658
#> log(c5/c10)  2.33647448  4.8050675  1.7171704  1.0047956  4.5519630
#> log(c6/c10) -0.01972759  0.6957633 -0.1992251 -0.4056981  0.6224041
#> log(c7/c10)  3.79029516  3.2641829  3.9222827  4.0741056  3.3181251
#> log(c8/c10) -0.81245987  2.3025213 -1.5939255 -2.4928319  1.9831427
#> log(c9/c10) -4.27758472 -3.2536239 -4.5344692 -4.8299589 -3.3586105
#>