Predict response from new data

Predict response from new data

# S3 method for pibblefit
predict(
  object,
  newdata = NULL,
  response = "LambdaX",
  size = NULL,
  use_names = TRUE,
  summary = FALSE,
  iter = NULL,
  from_scratch = FALSE,
  ...
)

Arguments

object

An object of class pibblefit

newdata

An optional matrix for which to evaluate predictions. If NULL (default), the original data of the model is used.

response

Options = "LambdaX":Mean of regression, "Eta", "Y": counts

size

the number of counts per sample if response="Y" (as vector or matrix), default if newdata=NULL and response="Y" is to use colsums of m$Y. Otherwise uses median colsums of m$Y as default. If passed as a matrix should have dimensions ncol(newdata) x iter.

use_names

if TRUE apply names to output

summary

if TRUE, posterior summary of predictions are returned rather than samples

iter

number of iterations to return if NULL uses object$iter

from_scratch

should predictions of Y come from fitted Eta or from predictions of Eta from posterior of Lambda? (default: false)

...

other arguments passed to summarise_posterior

Value

(if summary==FALSE) array D x N x iter; (if summary==TRUE) tibble with calculated posterior summaries

Details

currently only implemented for pibblefit objects in coord_system "default" "alr", or "ilr".

Examples

sim <- pibble_sim()
fit <- pibble(sim$Y, sim$X)
predict(fit)[,,1:2] # just show 2 samples
#> , , 1
#> 
#>                     s1        s2         s3         s4         s5         s6
#> log(c1/c10)  1.5851530 -2.566587  0.1085145  0.5505849  0.8893789  0.9997148
#> log(c2/c10) -1.1023540 -4.348626 -2.2569471 -1.9112894 -1.6463842 -1.5601118
#> log(c3/c10)  2.8182258  2.212074  2.6026372  2.6671793  2.7166431  2.7327521
#> log(c4/c10) -1.0272309  1.484045 -0.1340521 -0.4014486 -0.6063759 -0.6731151
#> log(c5/c10) -4.0803216  2.121927 -1.8743840 -2.5347892 -3.0409107 -3.2057405
#> log(c6/c10)  4.7896053 -8.187742  0.1739862  1.5557927  2.6147819  2.9596656
#> log(c7/c10)  0.2493988  3.615942  1.4467685  1.0883045  0.8135848  0.7241161
#> log(c8/c10)  2.4429598 10.469327  5.2976770  4.4430426  3.7880677  3.5747604
#> log(c9/c10)  3.9171739  9.630361  5.9491681  5.3408374  4.8746248  4.7227921
#>                     s7         s8         s9        s10        s11        s12
#> log(c1/c10)  2.4845897  1.0681934  1.5209413  1.3312062  0.9870117  1.3774965
#> log(c2/c10) -0.3990787 -1.5065680 -1.1525615 -1.3009166 -1.5700444 -1.2647220
#> log(c3/c10)  2.9495432  2.7427499  2.8088509  2.7811497  2.7308974  2.7879080
#> log(c4/c10) -1.5712758 -0.7145359 -0.9883909 -0.8736253 -0.6654314 -0.9016250
#> log(c5/c10) -5.4239823 -3.3080400 -3.9843962 -3.7009526 -3.1867635 -3.7701052
#> log(c6/c10)  7.6010294  3.1737131  4.5888948  3.9958282  2.9199587  4.1405206
#> log(c7/c10) -0.4799322  0.6685886  0.3014665  0.4553180  0.7344168  0.4177824
#> log(c8/c10)  0.7041202  3.4423740  2.5670973  2.9339035  3.5993187  2.8444126
#> log(c9/c10)  2.6794639  4.6285592  4.0055353  4.2666287  4.7402728  4.2029289
#>                    s13        s14        s15        s16         s17        s18
#> log(c1/c10)  0.5849769  1.7111174  0.6637714  1.3436076  1.98409613  2.2265055
#> log(c2/c10) -1.8843980 -1.0038616 -1.8227882 -1.2912198 -0.79041779 -0.6008764
#> log(c3/c10)  2.6722005  2.8366166  2.6837045  2.7829603  2.87647134  2.9118630
#> log(c4/c10) -0.4222514 -1.1034233 -0.4699120 -0.8811266 -1.26854079 -1.4151677
#> log(c5/c10) -2.5861671 -4.2684987 -2.7038775 -3.7194790 -4.67629929 -5.0384325
#> log(c6/c10)  1.6632940  5.1833398  1.9095866  4.0345923  6.03660619  6.7943199
#> log(c7/c10)  1.0604169  0.1472574  0.9965244  0.4452620 -0.07409436 -0.2706581
#> log(c8/c10)  4.3765541  2.1994387  4.2242244  2.9099283  1.67170139  1.2030626
#> log(c9/c10)  5.2935107  3.7438350  5.1850821  4.2495631  3.36819039  3.0346122
#>                    s19        s20        s21        s22        s23        s24
#> log(c1/c10)  2.1274082  1.2263281  2.6632322  2.4317766  1.3449893 -1.0662664
#> log(c2/c10) -0.6783612 -1.3829214 -0.2593969 -0.4403736 -1.2901395 -3.1755158
#> log(c3/c10)  2.8973948  2.7658375  2.9756249  2.9418325  2.7831620  2.4311197
#> log(c4/c10) -1.3552264 -0.8101874 -1.6793319 -1.5393306 -0.8819623  0.5765411
#> log(c5/c10) -4.8903920 -3.5442762 -5.6908549 -5.3450853 -3.7215431 -0.1193889
#> log(c6/c10)  6.4845655  3.6680045  8.1594233  7.4359483  4.0389110 -3.4980980
#> log(c7/c10) -0.1903026  0.5403610 -0.6247891 -0.4371074  0.4441416  2.3993694
#> log(c8/c10)  1.3946428  3.1366594  0.3587588  0.8062214  2.9072572  7.5688268
#> log(c9/c10)  3.1709794  4.4109508  2.4336349  2.7521397  4.2476618  7.5657777
#>                      s25        s26        s27        s28         s29
#> log(c1/c10) -0.127361982  1.7556760  2.1463766  1.2607666  0.08398491
#> log(c2/c10) -2.441380389 -0.9690209 -0.6635296 -1.3559937 -2.27612690
#> log(c3/c10)  2.568199361  2.8431221  2.9001642  2.7708655  2.59905592
#> log(c4/c10)  0.008623191 -1.1303756 -1.3666999 -0.8310183 -0.11921484
#> log(c5/c10) -1.522010177 -4.3350644 -4.9187287 -3.5957236 -1.83773948
#> log(c6/c10) -0.563307179  5.3226195  6.5438562  3.7756511  0.09731269
#> log(c7/c10)  1.638034935  0.1111259 -0.2056836  0.5124357  1.46665895
#> log(c8/c10)  5.753686101  2.1132955  1.3579720  3.0700810  5.34509889
#> log(c9/c10)  6.273756439  3.6825181  3.1448771  4.3635602  5.98292313
#>                    s30
#> log(c1/c10)  0.3711916
#> log(c2/c10) -2.0515581
#> log(c3/c10)  2.6409880
#> log(c4/c10) -0.2929384
#> log(c5/c10) -2.2667951
#> log(c6/c10)  0.9950522
#> log(c7/c10)  1.2337701
#> log(c8/c10)  4.7898554
#> log(c9/c10)  5.5876995
#> 
#> , , 2
#> 
#>                      s1          s2         s3         s4         s5         s6
#> log(c1/c10)  1.57381394  -0.1180356  0.9720782  1.1522236  1.2902834  1.3352457
#> log(c2/c10) -0.96412494  -2.1852177 -1.3984278 -1.2684079 -1.1687632 -1.1363116
#> log(c3/c10)  2.79931640   0.5485122  1.9987787  2.2384407  2.4221128  2.4819298
#> log(c4/c10) -1.07605504   1.5173769 -0.1536559 -0.4298002 -0.6414318 -0.7103544
#> log(c5/c10) -3.64917136   2.9403618 -1.3054893 -2.0071319 -2.5448570 -2.7199792
#> log(c6/c10)  3.73912865 -10.3263040 -1.2634871  0.2341770  1.3819572  1.7557576
#> log(c7/c10)  0.02482697   5.0771662  1.8217793  1.2838145  0.8715290  0.7372589
#> log(c8/c10)  2.69812851   6.4244863  4.0234725  3.6266960  3.3226144  3.2235834
#> log(c9/c10)  3.37015205   7.9058798  4.9833625  4.5004057  4.1302771  4.0097364
#>                     s7         s8         s9        s10        s11        s12
#> log(c1/c10)  1.9403378  1.3631510  1.5476474  1.4703296  1.3300691  1.4891931
#> log(c2/c10) -0.6995863 -1.1161710 -0.9830106 -1.0388147 -1.1400478 -1.0252000
#> log(c3/c10)  3.2869327  2.5190544  2.7645050  2.6616429  2.4750430  2.6867385
#> log(c4/c10) -1.6378985 -0.7531303 -1.0359445 -0.9174242 -0.7024193 -0.9463400
#> log(c5/c10) -5.0767338 -2.8286665 -3.5472562 -3.2461136 -2.6998172 -3.3195843
#> log(c6/c10)  6.7862767  1.9877519  3.5215896  2.8787973  1.7127215  3.0356214
#> log(c7/c10) -1.0697164  0.6539260  0.1029676  0.3338602  0.7527176  0.2775286
#> log(c8/c10)  1.8908470  3.1621211  2.7557612  2.9260563  3.2349849  2.8845088
#> log(c9/c10)  2.3875279  3.9349245  3.4403026  3.6475860  4.0236144  3.5970144
#>                    s13        s14        s15        s16        s17        s18
#> log(c1/c10)  1.1662384  1.6251449  1.1983475  1.4753833  1.7363848  1.8351675
#> log(c2/c10) -1.2582926 -0.9270768 -1.2351179 -1.0351673 -0.8467894 -0.7754930
#> log(c3/c10)  2.2570858  2.8676061  2.2998030  2.6683662  3.0155975  3.1470161
#> log(c4/c10) -0.4512836 -1.1547401 -0.5005035 -0.9251710 -1.3252594 -1.4766832
#> log(c5/c10) -2.0617180 -3.8490987 -2.1867785 -3.2657969 -4.2823635 -4.6671092
#> log(c6/c10)  0.3506916  4.1658758  0.6176347  2.9208115  5.0906843  5.9119285
#> log(c7/c10)  1.2419622 -0.1284617  1.1460755  0.3187686 -0.4606553 -0.7556483
#> log(c8/c10)  3.5958278  2.5850703  3.5251065  2.9149254  2.3400606  2.1224882
#> log(c9/c10)  4.4628328  3.2325374  4.3767508  3.6340376  2.9343113  2.6694819
#>                    s19        s20        s21        s22        s23        s24
#> log(c1/c10)  1.7947850  1.4275915  2.0131353  1.9188162  1.4759463  0.4933506
#> log(c2/c10) -0.8046391 -1.0696610 -0.6470447 -0.7151195 -1.0347609 -1.7439496
#> log(c3/c10)  3.0932920  2.6047848  3.3837811  3.2583008  2.6691152  1.3618887
#> log(c4/c10) -1.4147810 -0.8519110 -1.7494896 -1.6049082 -0.9260340  0.5801845
#> log(c5/c10) -4.5098246 -3.0796539 -5.3602705 -4.9929102 -3.2679898  0.5590922
#> log(c6/c10)  5.5762028  2.5234872  7.3914891  6.6073543  2.9254924 -5.2434573
#> log(c7/c10) -0.6350547  0.4614885 -1.2871102 -1.0054470  0.3170872  3.2513947
#> log(c8/c10)  2.2114321  3.0201886  1.7305080  1.9382489  2.9136853  5.0778867
#> log(c9/c10)  2.7777445  3.7621641  2.1923630  2.4452256  3.6325281  6.2667970
#>                      s25        s26        s27        s28        s29        s30
#> log(c1/c10)  0.875957676  1.6433027  1.8025147  1.4416253  0.9620823  1.0791201
#> log(c2/c10) -1.467802829 -0.9139714 -0.7990602 -1.0595321 -1.4056423 -1.3211703
#> log(c3/c10)  1.870901812  2.8917629  3.1035755  2.6234551  1.9854804  2.1411852
#> log(c4/c10) -0.006312932 -1.1825742 -1.4266298 -0.8734234 -0.1383332 -0.3177402
#> log(c5/c10) -0.931112339 -3.9198210 -4.5399308 -3.1343138 -1.2665566 -1.7224036
#> log(c6/c10) -2.062598889  4.3168334  5.6404647  2.6401594 -1.3465893 -0.3735788
#> log(c7/c10)  2.108822279 -0.1826860 -0.6581378  0.4195795  1.8516298  1.5021220
#> log(c8/c10)  4.235181296  2.5450771  2.1944072  2.9892786  4.0454887  3.7877089
#> log(c9/c10)  5.241054785  3.1838576  2.7570218  3.7245404  5.0101608  4.6963908
#>