Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]       [,2]        [,3]      [,4]       [,5]       [,6]
#>  [1,]  0.17972467  1.9802786 -1.36203693 2.3910726 -0.5855506  1.0551165
#>  [2,] -1.15706044  1.6927125 -0.27725224 1.8363074 -0.7706088  2.1438777
#>  [3,]  0.03407989 -0.8061162  0.12094053 1.2705244 -0.1613689  2.1109857
#>  [4,] -0.13786343  1.5455236 -2.46801814 2.4360575  0.1250337  2.3001989
#>  [5,]  0.17392603  2.0416970 -0.12527124 2.7519016 -1.1464897 -0.2803633
#>  [6,] -1.19129312  0.9170129  0.23878118 2.8795472 -0.6748613  1.7854100
#>  [7,] -0.97044552  1.7197830  0.09035569 2.9105705 -0.9100598  0.8951849
#>  [8,] -0.41652160  1.3324674  0.16082763 2.6868630 -3.6659757  0.6003186
#>  [9,] -1.38894582  1.7470582 -0.00531996 0.9954528 -1.1013759  0.2680274
#>              [,7]        [,8]       [,9]       [,10]
#>  [1,] -1.48547792  0.45045483  0.3015782  1.00008657
#>  [2,]  0.94512546  0.53748813  0.8003629  0.61981132
#>  [3,] -0.11777393  0.38515607 -0.7661034  0.72963559
#>  [4,]  0.04414912  0.04441348 -3.1309734  0.54625475
#>  [5,] -0.92830810 -0.80096151  0.2620506 -0.83227154
#>  [6,] -0.77288381  0.65580581  0.3821280  0.64777610
#>  [7,]  0.25661419 -2.38862334 -0.1597224  0.98832850
#>  [8,]  0.78055002  0.75761298  0.8321701 -0.12687649
#>  [9,]  0.56923429  0.06221731  0.1323470  0.09365772