Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]        [,2]       [,3]        [,4]        [,5]       [,6]
#>  [1,]  0.05988830  0.22200244 -1.4439121  0.80539493 -2.70084841  0.5072587
#>  [2,]  0.01672398 -1.35839796 -1.1614268  0.33551194 -0.90739775  0.5424467
#>  [3,] -1.09778885 -1.84265776 -1.0000079 -0.33851902 -0.26832977  0.2897846
#>  [4,] -2.29520750 -0.13245094  0.1212353  0.39332271 -1.13352399  0.6182419
#>  [5,] -3.21000060 -2.28041530 -1.2620538 -0.48656269 -2.41421377 -0.3063738
#>  [6,] -0.42918128 -0.07894671 -0.8784495  0.41237694 -1.42169104 -1.2888813
#>  [7,] -0.83625202 -1.52028213 -0.4071517  0.54743551  0.02062842  0.6633794
#>  [8,] -0.25345392  0.28567793  0.1945765 -0.19979189 -0.50650425  0.8244303
#>  [9,] -1.92039001 -3.07950895 -0.6219828  0.01372531 -0.02184874  0.7874963
#>              [,7]       [,8]      [,9]     [,10]
#>  [1,] -1.54705683  0.8120087 0.9848125 1.4716509
#>  [2,] -0.27638104 -3.6206569 0.3746930 0.1301129
#>  [3,]  0.01925536 -0.7510413 2.3270459 0.7554408
#>  [4,] -0.12966404  0.6003681 1.4067822 0.9723411
#>  [5,] -1.01197444  0.9902941 2.2193658 0.8528919
#>  [6,] -0.90019375 -2.2282258 1.8514772 0.7125314
#>  [7,] -0.47988395 -0.8652033 1.7423041 0.7370355
#>  [8,] -0.40337816  0.1342081 1.8039174 1.1817202
#>  [9,]  0.04039148  0.6778454 2.1271482 0.6892405