Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>            [,1]       [,2]      [,3]       [,4]        [,5]      [,6]
#>  [1,] 1.2153253  0.9340758 1.8325972  0.1027748  0.80573358 2.8123942
#>  [2,] 0.3483113  0.7480699 0.8598669 -0.1347005  0.81927864 2.4558345
#>  [3,] 0.6421553  1.1002411 0.1654081  0.6292127  0.60995249 1.6564402
#>  [4,] 1.0339161  1.3575223 2.8314375 -1.0825468  0.27634028 1.3474448
#>  [5,] 1.1183622 -1.0905063 2.0544604 -0.4607769 -1.30893720 0.4892664
#>  [6,] 0.9269112 -0.6553561 2.5768584 -1.4741856 -0.02396701 2.7677428
#>  [7,] 0.6221744  0.7171533 3.3728362  0.5457138 -0.73620589 2.6225213
#>  [8,] 0.6941650  1.1205609 2.5030193 -0.5281420 -0.76802125 2.3236549
#>  [9,] 0.3950974  1.0343786 3.0023001  0.5017858 -1.49124868 1.9547879
#>              [,7]        [,8]        [,9]      [,10]
#>  [1,] -0.90746377 -0.45790848 -0.94330072 -0.9392825
#>  [2,] -1.23387849 -1.14534384 -4.30362487 -1.0134271
#>  [3,] -0.68243648 -0.94990431  0.01934845 -0.1373806
#>  [4,]  0.04685008 -0.52468634 -2.84118371 -0.2267075
#>  [5,] -0.11964573 -1.62578915 -1.24602479 -1.7811165
#>  [6,] -0.39002261 -0.28312217 -0.06364970 -1.8179565
#>  [7,] -0.37804015 -1.79769213 -0.26827356  0.1417628
#>  [8,] -0.49908266 -0.13095219 -0.01893911  0.1163476
#>  [9,] -0.74737832 -0.02233365 -0.32478936 -0.1857853