Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>             [,1]       [,2]       [,3]        [,4]       [,5]        [,6]
#>  [1,]  0.6131891 -0.7203225  0.4522318 -0.00845716 -2.1959515  0.84817795
#>  [2,]  0.5911955 -0.1940089  0.3033746 -0.17168658  0.2891779  0.83391466
#>  [3,]  1.3119522 -0.3731142  0.2873582 -2.06943410  0.4316306  0.01587504
#>  [4,]  0.5719570 -1.3873689 -0.7983095  0.01121772 -0.9366671  0.21296798
#>  [5,]  0.5557382 -0.3243571 -0.1949834 -0.10619234  0.4204216 -0.20384401
#>  [6,]  1.2923652 -0.8654809 -0.3687572 -0.56331688  0.1561763  0.03784684
#>  [7,]  1.3007714 -0.2387411 -0.6958059  0.06147090  0.3987320 -1.30699628
#>  [8,] -0.2671719  0.2400303 -0.5741078 -2.86661879  0.3613795 -2.79740134
#>  [9,] -0.5971632 -1.3280356 -2.0119404 -0.13390507 -0.8284417  0.68164643
#>             [,7]       [,8]        [,9]      [,10]
#>  [1,]  1.4432600  0.9912663  0.91668010 -1.4277206
#>  [2,] -0.4418566  1.9672021  1.07824820  0.1226104
#>  [3,]  1.7521005  0.6226385  0.93744213 -3.0705974
#>  [4,]  1.8328474  1.0763731  0.51440687  0.4941637
#>  [5,]  0.8197578  1.5829468 -0.83911389 -0.8468248
#>  [6,]  1.6542235 -1.4089965 -0.12049153  0.6759656
#>  [7,]  0.2763297  1.8197982  0.97878515  0.3561646
#>  [8,] -0.0532524  1.4374276  1.03151261 -1.9456742
#>  [9,]  1.4096835  1.7763660 -0.09442699  0.4808184