Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>             [,1]        [,2]       [,3]        [,4]        [,5]       [,6]
#>  [1,]  0.6584148 -0.21066816 -0.3842282 -2.12051076 -0.09626796 -0.6963568
#>  [2,]  0.5533425 -1.03299169 -0.3983781 -0.32479086 -3.02387405 -0.7748447
#>  [3,]  0.7050845 -0.29153797 -4.0754182 -2.19927058 -1.60692422  0.5841643
#>  [4,]  0.4352458 -1.29372797 -0.1997916 -1.60353006 -1.09472657  1.0255368
#>  [5,] -1.3409491 -0.02742845  0.0924987 -0.10548236 -3.34576498  0.7005520
#>  [6,] -1.8911774 -1.37488768 -0.9770387 -1.30997579 -0.60595424  0.5719092
#>  [7,] -0.1101902 -0.84504970 -0.7379570 -0.23448947 -0.79583121  0.3351429
#>  [8,] -0.6948371 -3.75211136 -2.2264137 -0.16773875 -2.01115716  0.5425372
#>  [9,] -0.4942060 -0.46790225  0.0101942 -0.04677696 -2.91741555  0.3061803
#>             [,7]        [,8]       [,9]       [,10]
#>  [1,]  0.3211066  0.55812816 -0.3798741 -1.39327114
#>  [2,] -0.9163810  0.16926525 -0.2741771  0.05887542
#>  [3,]  0.3963488 -1.18586958  0.4016049 -1.52327312
#>  [4,] -0.3917108 -0.39316039  0.2249946 -0.56910777
#>  [5,]  0.3367028 -2.20724681  0.3571894 -0.30878318
#>  [6,] -0.3693546 -0.10903308 -0.6930909 -0.20764306
#>  [7,] -1.9462349 -0.06551661  0.2420207 -0.34396426
#>  [8,]  0.2675678 -0.56392242 -0.5747669 -0.54715691
#>  [9,] -0.6874161 -1.32056195  0.5106642 -0.69075404