Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]        [,2]      [,3]       [,4]        [,5]        [,6]
#>  [1,]  0.23335018  0.52313414 0.9289976  0.2591024 -2.01033021  0.36550120
#>  [2,] -1.26390928  0.44845581 0.1267756  1.1749439  0.06366321  0.24342540
#>  [3,] -0.35249827 -0.78047815 1.8215898  0.4483661  0.81638220  0.04101842
#>  [4,]  0.42293861  0.17502914 1.6258030 -0.7859999  0.73859223  0.27999364
#>  [5,] -0.63147246  0.07823502 0.3807474  1.1876572  0.46302673 -0.45210105
#>  [6,] -0.23665704 -1.66139024 1.4106729 -0.4500333  1.04072459 -0.73501844
#>  [7,] -0.06314497  0.54376563 1.7128631  0.4044740 -1.85861702  0.31746679
#>  [8,] -1.48048188 -0.10328418 0.6038343  0.6994952  1.00550971  0.31520781
#>  [9,]  0.29243179 -0.21771396 1.9036338  0.6247671 -1.50111884  0.20376928
#>             [,7]       [,8]       [,9]       [,10]
#>  [1,]  0.2383502  0.5678267 -0.7627573 -0.29175856
#>  [2,] -3.0406897  1.0224983 -0.1341829 -0.41113968
#>  [3,]  0.6363111 -1.4671745 -0.4635436 -0.12853328
#>  [4,] -1.1202700  0.3721943 -1.3831264 -0.09188346
#>  [5,]  0.7651000 -1.2287189 -1.4413219 -1.35413032
#>  [6,] -0.5519723  1.1010085 -0.2975555 -0.81176063
#>  [7,]  0.9364162  1.1838267 -1.4963886  0.03638191
#>  [8,] -0.4895502 -2.0330977 -0.4169160 -1.89374704
#>  [9,] -2.6393401  0.5258394 -0.8236054 -0.87415395