Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>             [,1]       [,2]      [,3]       [,4]        [,5]        [,6]
#>  [1,]  0.9419647 -0.7659246 2.2307141 -0.2418225 -0.76821197 -0.15576099
#>  [2,] -0.0435244 -0.3176658 2.4849588 -1.4945825 -0.07239541 -0.04352735
#>  [3,]  0.7878844  0.1278067 1.8589697  0.2376483 -0.15630688 -0.18866871
#>  [4,]  1.0440072 -0.9060128 1.8910793  0.2519324  0.01963468 -0.29761587
#>  [5,]  1.1254362  0.1079424 1.1325941 -0.5782783 -1.55570532 -1.48715814
#>  [6,]  0.8497555 -0.8444903 2.4995170  0.0606125 -2.00167151 -0.44283666
#>  [7,]  0.8305718 -0.1378097 2.1559472 -2.3223913 -0.52700039 -2.90425664
#>  [8,] -0.1664457 -0.7895832 0.4812728 -1.5882365 -0.46048614 -1.01451130
#>  [9,] -1.0033781 -0.4872759 2.3096140 -1.1948137 -0.48800118 -0.41289869
#>              [,7]       [,8]       [,9]      [,10]
#>  [1,] -0.28041711  0.3110943  2.6609019 -0.3301271
#>  [2,]  0.01275770  0.1787909 -0.2327726 -0.7925546
#>  [3,] -0.06329584 -0.1154282  0.3983172 -0.5504075
#>  [4,] -2.07208582 -0.8691299  2.3599682  0.3391662
#>  [5,] -1.63984276  0.7703148  2.1687608 -0.9850789
#>  [6,]  0.18946398 -0.8005394  3.0770325 -3.0085082
#>  [7,]  0.24557081  0.6303853  3.4423683  1.0950647
#>  [8,] -0.16385688 -0.4216512  3.3837093 -0.1656980
#>  [9,]  0.42597118  0.6158176  2.9289959  0.7185146