Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>            [,1]        [,2]       [,3]        [,4]       [,5]      [,6]
#>  [1,] 3.0902137 -0.03777503  0.6469273 -0.20945578 -2.3428291 1.4065048
#>  [2,] 2.8576554 -0.60198450  0.5076632 -0.03177581 -2.5778620 0.5781131
#>  [3,] 3.8515198  0.33796091  0.3263868 -1.06138469 -0.7370009 1.0113351
#>  [4,] 3.9131891 -1.55994335 -1.2503263 -0.54247226 -0.4372750 1.5187596
#>  [5,] 0.6336341  0.07599477  0.3062141 -0.17732931  0.3630054 1.4929786
#>  [6,] 2.7637961 -0.28597495 -1.3444678 -0.67861381 -1.2851143 0.6967902
#>  [7,] 3.8977533 -0.90990968 -1.5480211 -0.07555305 -2.0696451 1.0642952
#>  [8,] 3.5506538 -1.54134887  0.7393279 -0.66115520 -0.9570645 1.3746556
#>  [9,] 0.1420021 -1.24163356  0.5885159 -1.16412669 -1.0954214 0.5011682
#>             [,7]        [,8]        [,9]       [,10]
#>  [1,]  1.2505314  0.59671862 -0.06841876  0.81129446
#>  [2,]  0.5811263  0.52406628 -0.14727878 -1.05035081
#>  [3,] -0.6356646 -0.63338797 -1.04562631  0.99267243
#>  [4,]  0.9513191 -1.25072241  0.36803844 -0.03227908
#>  [5,]  1.7243380 -0.97482840  0.16899641  1.11664835
#>  [6,]  1.0119515  0.65553767  0.04417951  0.96632274
#>  [7,]  1.1849297  0.02266733  0.37951915  0.91927557
#>  [8,]  1.6162853  0.12257753  0.58684263  1.19654637
#>  [9,]  1.8303033  0.72768435 -0.49117020  0.85910583