Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]       [,2]       [,3]       [,4]        [,5]        [,6]
#>  [1,]  0.48184501 -1.1512379 -1.6743095  3.2597422 -0.06178952  0.71477691
#>  [2,]  0.60449279 -0.2781032  0.3025288  3.3628533 -3.84005222 -0.31867851
#>  [3,]  0.72657015 -2.6950694 -0.7035808  3.1070975  0.04210371  1.46406529
#>  [4,]  0.07959015 -1.8544047 -0.4954855 -0.2891117  0.20489029  0.39840176
#>  [5,] -1.19601910 -0.6214062  0.3449052  3.2635169 -0.09586154  1.01395644
#>  [6,]  0.59321948 -0.9087480  0.4721827  1.4966606  0.18276383 -0.46738601
#>  [7,]  0.24572054 -0.7125731 -0.1706042  2.9981875 -0.25804078  0.63320409
#>  [8,]  0.74064611  0.2088646 -2.4306735  2.7767105 -0.04443093  0.05249657
#>  [9,]  0.63885758 -0.5841221 -0.9857492  2.0709182  0.10199313  1.31793619
#>              [,7]        [,8]      [,9]       [,10]
#>  [1,]  1.23396603 -0.43549571 1.7279824  0.03344978
#>  [2,]  1.20712859  0.47656601 1.5589525 -1.22492271
#>  [3,]  1.14605630 -2.01708879 0.4758807 -1.46685269
#>  [4,]  0.97941377  0.26859282 1.7830523 -0.53959374
#>  [5,]  1.26197779 -0.01740323 2.3057458 -2.24199313
#>  [6,]  0.09580698  0.44205980 2.4578203 -1.57125826
#>  [7,] -0.34799129 -0.38623430 2.3759005 -1.42861686
#>  [8,]  0.86909255 -0.11803584 2.0654541 -0.55130733
#>  [9,]  0.10291214  0.05799732 0.4146906 -0.86054188