Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>             [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#>  [1,] -1.0668639 -1.87420699 -0.27536961  0.7097967  1.22400776 -0.90463373
#>  [2,] -0.5255439 -0.17694117  0.57540001  0.1516420  1.13955934 -0.39169123
#>  [3,]  0.1902102  0.02312380  0.31409387 -0.5991870 -0.07382403 -0.06529549
#>  [4,]  0.5204824 -0.08933938  0.65437865 -0.9886719 -0.41069526 -0.48538841
#>  [5,] -1.3183160  0.22049433 -0.10542609  0.2315321  0.52998909 -1.04030174
#>  [6,]  0.3272156  0.09110182  0.42075074  0.8859210  1.33156153 -1.45097028
#>  [7,] -2.2378574 -3.31111888  0.08189964  0.7238870  1.15039766  0.01329535
#>  [8,] -0.1335634  0.27440311  1.07472006  0.8038833 -1.74625737 -0.98368650
#>  [9,] -1.7888948 -0.37936397  1.12679642 -0.3911684  0.42180656 -0.31836487
#>              [,7]       [,8]       [,9]     [,10]
#>  [1,] -0.44761666  1.4949527  0.1814388 2.0127516
#>  [2,]  0.07162560  0.3107088  0.2540059 1.1665363
#>  [3,] -0.21253277 -1.0646295  0.3050951 2.3110200
#>  [4,] -1.69421481  1.8751256 -0.6309054 2.1094513
#>  [5,]  0.04326972  1.6703871 -2.4414323 1.5175939
#>  [6,] -2.21552104  0.8720759 -0.4635019 2.1260628
#>  [7,] -0.09448114  1.1691564  0.1406521 1.9708368
#>  [8,] -1.29215622  1.6302800 -3.9308819 0.8505534
#>  [9,]  0.17855523  0.5703536 -0.2811132 1.3992395