Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]        [,2]       [,3]        [,4]       [,5]     [,6]
#>  [1,] -1.36433165  0.07135603  1.3608003  0.87543733  0.6287675 2.084703
#>  [2,]  0.08530188 -0.74008613  0.8433065  1.53376823  0.1790994 1.620020
#>  [3,] -2.27303589  0.01420400  1.4136940  0.50632166 -2.2923248 2.944876
#>  [4,] -0.23913660 -0.76769669  1.3866338  0.73079481 -0.8002513 3.037416
#>  [5,] -0.76560549 -0.38795775 -0.6299860  1.84672323  0.7523287 2.984400
#>  [6,] -3.53645395 -0.13916539  1.6021265  0.03646888 -0.2120000 2.278363
#>  [7,]  0.32975662 -0.60492830  1.5858000 -1.54454933  0.4958364 1.995850
#>  [8,]  0.18138328 -0.92497361  0.4026580  2.04512426 -0.9183943 2.625441
#>  [9,]  0.32171046  0.08715023 -0.3786447  1.47911436 -0.1097035 1.797878
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,] -2.11568894 -0.59768503 -1.06938487 -0.77566187
#>  [2,]  0.29900924 -0.41215724 -0.83320278 -0.42391837
#>  [3,] -0.21141428  0.24313296 -0.73902412  0.24712247
#>  [4,]  0.25043014  0.07326004 -2.23772275  0.39875639
#>  [5,]  0.32205411 -0.03213435  0.40333847 -1.09245530
#>  [6,] -0.08037665 -0.14023747 -0.09392278 -0.26959568
#>  [7,]  0.33499486 -0.24987849 -0.27067099  0.46021254
#>  [8,]  0.28165527 -0.47751744  0.63095802  0.06182568
#>  [9,] -0.47951901  0.19978546 -1.59436066 -1.47232498