Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.3170046 -1.55080772 -0.71011131 -0.82314976  0.86213572 -0.76379225
#>  [2,]  1.2328442 -0.21532558 -1.58161832 -0.28221958  0.20217913 -0.84292120
#>  [3,]  1.0768467 -0.03832228 -1.02924931 -3.60300180  0.63078163  0.18006082
#>  [4,]  1.7103364 -0.78276935 -1.20647771 -0.17999509  0.72083601 -2.08970674
#>  [5,]  1.4170848 -0.66260539 -1.71210830 -1.13385853  0.23744435  0.31326187
#>  [6,] -1.3553257 -1.60738802 -0.32064207 -0.76813318 -0.57163186 -0.46409110
#>  [7,]  0.8029483 -1.53374068  0.14572489 -0.04028753 -0.99603418 -0.01172994
#>  [8,]  1.2923850 -0.56671283 -0.34917151 -0.08478833  0.69820348  0.35838956
#>  [9,]  0.7875726 -2.07000500 -0.07004371 -2.67352980  0.01901132 -1.68815282
#>               [,7]        [,8]       [,9]       [,10]
#>  [1,] -0.024196166  0.11700768  0.5089546 -0.07957895
#>  [2,] -0.289032777 -0.37978322  0.6201055 -3.13609425
#>  [3,] -0.763486729 -0.71285681 -1.5250806  0.10208235
#>  [4,] -0.389386556 -0.31915975  0.5292085 -1.48022729
#>  [5,] -4.241269612 -0.03182593 -0.4643899 -2.48265736
#>  [6,] -0.201519416 -0.67677544  0.7063239 -0.48869925
#>  [7,] -0.821957471 -0.93026320  0.1105759 -0.84266529
#>  [8,] -0.002203711 -0.23992668  0.4671984 -0.09315359
#>  [9,] -1.528008431 -2.56083649  0.6807728 -0.36385436