Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]       [,2]       [,3]        [,4]      [,5]       [,6]
#>  [1,]  0.23660466  1.4598363  0.9658481  0.18864204 1.2016553  0.4090269
#>  [2,] -1.19376204  0.7459717 -0.8392955  0.27854944 1.3081819  1.5798430
#>  [3,]  0.19207150 -0.5465259  0.2568837  0.18131340 2.0076284  1.2946190
#>  [4,] -2.18542495  1.6698889  0.5886388 -0.68991605 0.8630323  1.5565141
#>  [5,] -0.26741449  0.9773509  0.7888081  0.62469770 1.1535058 -0.2702583
#>  [6,]  0.32496192  0.7853063  0.2360171  0.64754510 2.1580763  0.9309895
#>  [7,] -0.08265820  1.9155669  0.2074024  0.06431048 1.4820427 -1.4216753
#>  [8,]  0.01016646  1.7677720  0.8654985  0.03223031 2.1358697  1.0339394
#>  [9,]  0.08949171  1.2311659 -2.3995358  0.65230881 0.3264818  1.1043988
#>             [,7]       [,8]       [,9]       [,10]
#>  [1,]  0.4111793  0.2060477  0.9374666 -0.07689473
#>  [2,]  0.3751373  0.2438592  0.7169684  0.39788338
#>  [3,] -0.9219538 -1.4136812  0.8927650 -1.60776520
#>  [4,]  0.2549624 -3.4222411  0.8726415  0.47129714
#>  [5,]  0.7444977 -2.6268533  1.1067329 -1.18005676
#>  [6,] -1.3447426 -1.2379741  0.1702146 -0.47998559
#>  [7,] -2.8299551  0.3144693  1.0313656  0.18794529
#>  [8,]  0.6958352 -0.3683971 -0.6394268 -0.65754528
#>  [9,]  0.5662617  0.1261886  0.7910319 -0.84975115