Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] -0.13743727 -0.3312002 -1.5952317  2.5896872  1.0364971  0.7134912
#>  [2,] -0.03605466 -1.0946291  0.4302611  1.7693657  1.2425171  0.8972791
#>  [3,] -0.98053047 -1.0006552 -0.3529571  2.8586754  0.1886809  0.8603231
#>  [4,] -0.04712701  0.3486153 -1.2748981 -0.8325446  0.6011610  1.8485723
#>  [5,] -0.20535453  0.5282856 -0.6949068  2.4347884  0.7002748 -0.3407350
#>  [6,]  0.43013291 -0.2163443 -0.8447681  1.4769015  0.2124275  1.9787427
#>  [7,]  0.14265373 -0.0971531 -1.4255502  1.8172756 -0.5752273  1.2281001
#>  [8,] -2.58595155 -1.0625784  0.0375343  2.5399985 -0.9885152  1.6870342
#>  [9,] -0.46374826 -0.9394908  0.4786111  2.8058892  0.6653587  2.0580245
#>             [,7]        [,8]        [,9]      [,10]
#>  [1,]  0.7968142  0.44298835 -0.15629481  1.6804717
#>  [2,]  0.4671822 -1.06688846 -2.42821498  1.9322217
#>  [3,]  0.8580515  0.26008043 -0.13261153  2.5100225
#>  [4,]  1.0455422 -0.36516180  0.09458404  2.5321603
#>  [5,]  0.8988951  0.43221338 -0.67354694  2.3401178
#>  [6,] -3.1160542 -0.33808636 -2.04372820  0.9544726
#>  [7,]  0.9815253 -0.02679923  0.13288455 -0.7066310
#>  [8,]  1.1179594  0.19562534 -0.20936590  1.7681255
#>  [9,]  0.6984627 -0.60785405  0.17259204  2.3099825