Provide random initialization for pibble model

Randomly initializes based on ALR transform of counts plus random pseudocounts uniformily distributed between 0 and 1.

Usage

random_pibble_init(Y)

Arguments

Y

matrix (D x N) of counts

Value

(D-1) x N matrix

Details

Notation: N is number of samples and D is number of multinomial categories

Examples

Y <- matrix(sample(1:100, 100), 10, 10)
random_pibble_init(Y)
#>              [,1]        [,2]        [,3]      [,4]         [,5]      [,6]
#>  [1,]  0.25691678 -0.02542428 -1.71077550 0.5618306  0.048199384 1.5362480
#>  [2,]  0.04894514  0.38300048  0.71802318 1.0470637  0.465271644 1.6852486
#>  [3,] -0.41315757 -0.41230658  0.68130379 1.7962744 -0.736030783 0.2340033
#>  [4,] -1.69404220 -1.73348819  0.41878855 1.0833641  0.364209970 1.6273682
#>  [5,] -0.88520088  0.54355855 -0.70996077 2.1150139  0.067786718 0.8303504
#>  [6,] -0.14272512  0.54033132  0.36402135 1.3477515 -0.007851812 1.5560963
#>  [7,]  0.23082723  0.09026151 -0.09574289 1.1643665  0.565025610 1.6725503
#>  [8,] -1.48815729 -2.94704710 -1.80081519 2.0748132 -0.154641831 1.2196970
#>  [9,] -0.15613723 -0.35301335  0.55896341 2.2709071  0.623345290 1.2509881
#>             [,7]       [,8]       [,9]      [,10]
#>  [1,]  1.4969441 -1.2685237 -3.9462950  0.8770409
#>  [2,]  0.8154524  0.6967698 -0.2393166  1.0340328
#>  [3,]  1.3149137 -1.1849037 -0.9445366  0.9045163
#>  [4,]  1.4670036  0.5115539 -1.7799723  0.3241133
#>  [5,] -0.4840297  1.4934215 -3.0968369 -0.3972719
#>  [6,]  0.5678864  0.9174696 -0.5296983  0.0643022
#>  [7,]  1.5125982  0.3827255 -1.5916887  0.6279326
#>  [8,] -2.2549465  0.2295093 -0.4690426 -0.2958178
#>  [9,] -0.3789016  1.5317857 -0.3711958  0.4671585